% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{action}
\alias{action}
\alias{option}
\alias{plinkr}
\title{Main functions}
\usage{
action(.string, action, ...)

option(.string, option, ...)

plinkr(.string = "", ...)
}
\arguments{
\item{.string}{A character string}

\item{action}{A character string indicating plink flag.}

\item{option}{A character string indicating an option from a flag}
}
\description{
A set of functions to manage plink strings.
}
