% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{assoc}
\alias{assoc}
\title{Let PLINK detect an association with one binary traits.}
\usage{
assoc(assoc_params, plink_options = create_plink_options(), verbose = FALSE)
}
\arguments{
\item{assoc_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc}),
as can be created by \link{create_assoc_params}}

\item{verbose}{set to TRUE for more output}
}
\value{
a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{trait_name}: name of the quantitive trait,
taken from the phenotype table column name
\item \code{CHR}: Chromosome number
\item \code{SNP}: SNP identifier
\item \code{BP}: Physical position (base-pair)
\item \code{NMISS}: Number of non-missing genotypes
\item \code{BETA}: Regression coefficient
\item \code{SE}: Standard error
\item \code{R2}: Regression r-squared
\item \code{T}: Wald test (based on t-distribution)
\item \code{P}: Wald test asymptotic p-value
}

The table with have as much rows as the number of SNPs multiplied
by the number of traits.

Note that parameters in uppercase are named as such by PLINK.
}
\description{
Let PLINK detect an association with one binary traits.
}
\note{
This function is named after the \code{--assoc} flag used by PLINK
}
\examples{
if (is_plink_installed() && 1 == 2) {
  assoc(create_demo_assoc_params())
}
}
\author{
Richèl J.C. Bilderbeek
}
