% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_from_bfile.R
\name{assoc_from_bfile}
\alias{assoc_from_bfile}
\title{Let \code{PLINK}/\code{PLINK2}
do a binary-trait/case-control association}
\usage{
assoc_from_bfile(bfile, out)
}
\arguments{
\item{bfile}{the base filename of the binary files (i.e.
a \code{.bed}, \code{.bim} and \code{.fam} file).
This parameter is named after the \code{PLINK}
\code{--bfile} flag}

\item{out}{the base filename of the output files.
This parameter is named after the \code{PLINK}
\code{--out} flag}
}
\value{
a \link{list} with the following columns:
\itemize{
\item \code{assoc_table}: a \link[tibble]{tibble} with associations
found by \code{PLINK}.
See \link{read_plink_assoc_file} for an explanation of the
column names.
\item \code{log}: the log file as text as created by \code{PLINK}
}
}
\description{
Let \code{PLINK}/\code{PLINK2}
do a binary-trait/case-control association
from PLINK1 binary filenames
}
\note{
This function is named after the \code{--assoc} flag
used by \code{PLINK}
}
\author{
Richèl J.C. Bilderbeek
}
