% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_qt_covar.R
\name{assoc_qt_covar}
\alias{assoc_qt_covar}
\title{Let PLINK detect an association with one ore more quantitative traits.}
\usage{
assoc_qt_covar(
  assoc_qt_covar_data,
  assoc_qt_covar_params,
  plink_options = create_plink_options(),
  verbose = FALSE
)
}
\arguments{
\item{assoc_qt_covar_params}{parameters to do an association analysis
for a quantitative trait with covariates
(i.e. using \link{assoc_qt_covar}),
as can be created by \link{create_assoc_qt_covar_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{trait_name}: name of the quantitive trait,
taken from the phenotype table column name
\item \code{CHR}: Chromosome number
\item \code{SNP}: SNP identifier
\item \code{BP}: Physical position (base-pair)
\item \code{NMISS}: Number of non-missing genotypes
\item \code{BETA}: Regression coefficient
\item \code{SE}: Standard error
\item \code{R2}: Regression r-squared
\item \code{T}: Wald test (based on t-distribution)
\item \code{P}: Wald test asymptotic p-value
}

The table with have as much rows as the number of SNPs multiplied
by the number of traits.

Note that parameters in uppercase are named as such by PLINK.
}
\description{
Let PLINK detect an association with one ore more quantitative traits.
}
\note{
This function is named after the \code{--assoc} flag used by PLINK
}
\author{
Richèl J.C. Bilderbeek
}
