% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_qt_on_plink2_bin_data.R
\name{assoc_qt_on_plink2_bin_data}
\alias{assoc_qt_on_plink2_bin_data}
\title{Let PLINK2 detect an association with one or more quantitative traits
with data in PLINK2 binary format.}
\usage{
assoc_qt_on_plink2_bin_data(
  assoc_qt_params,
  plink_options = create_plink_options(),
  verbose = FALSE
)
}
\arguments{
\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
A \link[tibble]{tibble} with the detected associations.
The table with have as much rows as the number of SNPs multiplied
by the number of traits.
}
\description{
Let PLINK2 detect an association with one or more quantitative traits.
with data in PLINK2 binary format.
}
\note{
This function is named after the \code{--assoc} flag used by PLINK.
}
\author{
Richèl J.C. Bilderbeek
}
