% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_qt_on_plink_bin_files.R
\name{assoc_qt_on_plink_bin_files}
\alias{assoc_qt_on_plink_bin_files}
\title{Let \code{PLINK} detect an association with one or more quantitative traits
with files in \code{PLINK} binary format.}
\usage{
assoc_qt_on_plink_bin_files(
  assoc_qt_data,
  assoc_qt_params,
  plink_options = create_plink_v1_9_options(),
  verbose = FALSE
)
}
\arguments{
\item{assoc_qt_data}{data use to used by \link{assoc_qt}
(as created by \link{create_assoc_qt_data}),
which equals 'regular' data (see \link{check_data}) and a
phenotype table/file}

\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \link{list} with the following elements:
\itemize{
\item \code{qassoc_filenames}: names of the created association files
\item \code{log_filename}: name of the PLINK log file
}
}
\description{
Let \code{PLINK} detect an association with one or more quantitative traits
with files in \code{PLINK} binary format.
}
\note{
This function is named after the \code{--assoc} flag used by PLINK.
}
\seealso{
use \link{assoc_qt_on_plink_bin_data} to work on data directly
}
\author{
Richèl J.C. Bilderbeek
}
