% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_qt_on_plink_text_data.R
\name{assoc_qt_on_plink_text_data}
\alias{assoc_qt_on_plink_text_data}
\title{Let PLINK detect an association with one or more quantitative traits
with data in PLINK text format.}
\usage{
assoc_qt_on_plink_text_data(
  assoc_qt_data,
  assoc_qt_params,
  plink_options = create_plink_options(),
  verbose = FALSE
)
}
\arguments{
\item{assoc_qt_data}{data use to used by \link{assoc_qt}
(as created by \link{create_assoc_qt_data}),
which equals 'regular' data (see \link{check_data}) and a
phenotype table/file}

\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
A \link[tibble]{tibble} with the detected associations.
The table with have as much rows as the number of SNPs multiplied
by the number of traits.

If the \code{assoc_qt_params} used PLINK1 text or PLINK1 binary data,
the \link[tibble]{tibble} has the following columns:
\itemize{
\item \code{trait_name}: name of the quantitive trait,
taken from the phenotype table column name
\item \code{CHR}: Chromosome number
\item \code{SNP}: SNP identifier
\item \code{BP}: Physical position (base-pair)
\item \code{NMISS}: Number of non-missing genotypes
\item \code{BETA}: Regression coefficient
\item \code{SE}: Standard error
\item \code{R2}: Regression r-squared
\item \code{T}: Wald test (based on t-distribution)
\item \code{P}: Wald test asymptotic p-value
}

Note that parameters in uppercase are named as such by PLINK.
}
\description{
Let PLINK detect an association with one or more quantitative traits.
with data in PLINK text format.
}
\note{
This function is named after the \code{--assoc} flag used by PLINK.
}
\examples{
if (is_plink_installed()) {
  assoc_qt(create_demo_assoc_qt_data())
}
}
\author{
Richèl J.C. Bilderbeek
}
