% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autom.R
\name{builder}
\alias{builder}
\alias{one}
\alias{epi1}
\alias{hwe2}
\alias{mh2}
\alias{p2}
\alias{r2}
\alias{t2}
\alias{adjust}
\alias{all_pheno}
\alias{allele1234}
\alias{alleleacgt}
\alias{allow_no_sex}
\alias{alt_group}
\alias{alt_snp}
\alias{annotate}
\alias{aperm}
\alias{assoc}
\alias{bd}
\alias{bed}
\alias{beta}
\alias{bfile}
\alias{bim}
\alias{bmerge}
\alias{case_only}
\alias{cc}
\alias{cell}
\alias{cfile}
\alias{chap}
\alias{check_sex}
\alias{chr}
\alias{ci}
\alias{clump}
\alias{clump_allow_overlap}
\alias{clump_annotate}
\alias{clump_best}
\alias{clump_field}
\alias{clump_index_first}
\alias{clump_kb}
\alias{clump_p1}
\alias{clump_p2}
\alias{clump_r2}
\alias{clump_range}
\alias{clump_range_border}
\alias{clump_replicate}
\alias{clump_verbose}
\alias{cluster}
\alias{cluster_missing}
\alias{cnv_blue}
\alias{cnv_border}
\alias{cnv_brown}
\alias{cnv_count}
\alias{cnv_del}
\alias{cnv_disrupt}
\alias{cnv_drop_no_segment}
\alias{cnv_dup}
\alias{cnv_exclude}
\alias{cnv_freq_excldue_above}
\alias{cnv_freq_excldue_below}
\alias{cnv_freq_excldue_exact}
\alias{cnv_freq_incldue_exact}
\alias{cnv_freq_method2}
\alias{cnv_green}
\alias{cnv_indiv_perm}
\alias{cnv_intersect}
\alias{cnv_kb}
\alias{cnv_list}
\alias{cnv_make_map}
\alias{cnv_max_kb}
\alias{cnv_max_score}
\alias{cnv_overlap}
\alias{cnv_red}
\alias{cnv_region_overlap}
\alias{cnv_report_regions}
\alias{cnv_score}
\alias{cnv_seglist}
\alias{cnv_subset}
\alias{cnv_test_2sided}
\alias{cnv_test_region}
\alias{cnv_test_window}
\alias{cnv_track}
\alias{cnv_union_overlap}
\alias{cnv_unique}
\alias{cnv_verbose_report_regions}
\alias{cnv_write}
\alias{cnv_write_freq}
\alias{compound_genotypes}
\alias{condition}
\alias{condition_list}
\alias{control}
\alias{counts}
\alias{covar}
\alias{covar_name}
\alias{covar_number}
\alias{cow}
\alias{dfam}
\alias{distance_matrix}
\alias{dog}
\alias{dominant}
\alias{dummy}
\alias{dummy_coding}
\alias{each_versus_others}
\alias{each_vs_others}
\alias{epistasis}
\alias{exclude}
\alias{exclude_before_extract}
\alias{extract}
\alias{fam}
\alias{family}
\alias{fast_epistasis}
\alias{file}
\alias{filter}
\alias{filter_cases}
\alias{filter_controls}
\alias{filter_females}
\alias{filter_founders}
\alias{filter_males}
\alias{filter_nonfounders}
\alias{fisher}
\alias{flip}
\alias{flip_scan}
\alias{flip_subset}
\alias{freq}
\alias{from}
\alias{from_bp}
\alias{from_kb}
\alias{from_mb}
\alias{gap}
\alias{gene}
\alias{gene_list}
\alias{gene_list_border}
\alias{gene_report}
\alias{gene_report_empty}
\alias{gene_subset}
\alias{genedrop}
\alias{genepi}
\alias{geno}
\alias{genome}
\alias{genome}
\alias{genotypic}
\alias{gfile}
\alias{gxe}
\alias{hap}
\alias{hap_assoc}
\alias{hap_freq}
\alias{hap_impute}
\alias{hap_max_phase}
\alias{hap_min_phase_prob}
\alias{hap_miss}
\alias{hap_phase}
\alias{hap_phase_wide}
\alias{hap_pp}
\alias{hap_snps}
\alias{hap_tdt}
\alias{hap_window}
\alias{hardy}
\alias{hardy2}
\alias{help}
\alias{het}
\alias{homog}
\alias{homozyg_group}
\alias{homozyg_het}
\alias{homozyg_kb}
\alias{homozyg_match}
\alias{homozyg_snp}
\alias{homozyg_verbose}
\alias{horse}
\alias{hwe}
\alias{hwe_all}
\alias{ibm}
\alias{impossible}
\alias{impute_sex}
\alias{indep}
\alias{indep_pairwise}
\alias{independent_effect}
\alias{interaction}
\alias{keep}
\alias{keep_allele_order}
\alias{keep_before_remove}
\alias{lambda}
\alias{ld_window}
\alias{lfile}
\alias{liability}
\alias{linear}
\alias{list}
\alias{logistic}
\alias{lookup}
\alias{lookup_gene}
\alias{lookup_list}
\alias{loop_assoc}
\alias{maf}
\alias{make_bed}
\alias{make_founders}
\alias{make_pheno}
\alias{make_pheno}
\alias{map}
\alias{map3}
\alias{match}
\alias{match_type}
\alias{matrix}
\alias{max}
\alias{max_maf}
\alias{mc}
\alias{mcc}
\alias{mcovar}
\alias{me}
\alias{mendel}
\alias{merge}
\alias{merge_list}
\alias{merge_mode}
\alias{meta_analysis}
\alias{mfilter}
\alias{mh}
\alias{mhf}
\alias{min}
\alias{mind}
\alias{missing}
\alias{missing_genotype}
\alias{missing_phenotype}
\alias{model}
\alias{model_fisher}
\alias{model_dom}
\alias{model_gen}
\alias{model_rec}
\alias{model_trend}
\alias{mouse}
\alias{mperm}
\alias{mpheno}
\alias{must_have_sex}
\alias{mwithin}
\alias{neighbour}
\alias{no_fid}
\alias{no_parents}
\alias{no_pheno}
\alias{no_sex}
\alias{nonfounders}
\alias{nop}
\alias{nudge}
\alias{null_group}
\alias{null_snp}
\alias{oblig_cluster}
\alias{oblig_missing}
\alias{out}
\alias{output_missing_genotype}
\alias{output_missing_phenotype}
\alias{parameters}
\alias{parentdt1}
\alias{parentdt2}
\alias{ped}
\alias{pedigree}
\alias{perm}
\alias{pheno}
\alias{pheno_name}
\alias{plist}
\alias{poo}
\alias{ppc}
\alias{ppc_gap}
\alias{proxy_assoc}
\alias{proxy_b_kb}
\alias{proxy_b_maxsnp}
\alias{proxy_b_r2}
\alias{proxy_b_threshold}
\alias{proxy_b_window}
\alias{proxy_dosage}
\alias{proxy_drop}
\alias{proxy_flanking}
\alias{proxy_geno}
\alias{proxy_genotypic_concordance}
\alias{proxy_glm}
\alias{proxy_impute}
\alias{proxy_impute_threshold}
\alias{proxy_kb}
\alias{proxy_list}
\alias{proxy_maf}
\alias{proxy_maxsnp}
\alias{proxy_mhf}
\alias{proxy_r2}
\alias{proxy_r2_no_filter}
\alias{proxy_replace}
\alias{proxy_show_proxies}
\alias{proxy_sub_maxsnp}
\alias{proxy_sub_r2}
\alias{proxy_tdt}
\alias{proxy_verbose}
\alias{proxy_window}
\alias{prune}
\alias{qmatch}
\alias{qq_plot}
\alias{qt}
\alias{qual_geno_max_threshold}
\alias{qual_geno_scores}
\alias{qual_geno_threshold}
\alias{qual_max_threshold}
\alias{qual_scores}
\alias{qual_threshold}
\alias{r}
\alias{rank}
\alias{read_genome}
\alias{recessive}
\alias{recode}
\alias{recode_bimbam}
\alias{recode_fastphase}
\alias{recode_lgen}
\alias{recode_rlist}
\alias{recode_structure}
\alias{recode12}
\alias{recodea}
\alias{recodead}
\alias{recodehv}
\alias{reference_allele}
\alias{rel_check}
\alias{remove}
\alias{script}
\alias{set}
\alias{set_by_all}
\alias{set_hh_missing}
\alias{set_max}
\alias{set_me_missing}
\alias{set_p}
\alias{set_r2}
\alias{set_table}
\alias{set_test}
\alias{sex}
\alias{sheep}
\alias{silent}
\alias{simulate}
\alias{simulate_haps}
\alias{simulate_label}
\alias{simulate_ncases}
\alias{simulate_ncontrols}
\alias{simulate_prevalence}
\alias{simulate_qt}
\alias{simulate_tags}
\alias{snp}
\alias{snps}
\alias{specific_haplotype}
\alias{subset}
\alias{swap_parents}
\alias{swap_sibs}
\alias{swap_unrel}
\alias{tab}
\alias{tdt}
\alias{test_all}
\alias{test_mishap}
\alias{test_missing}
\alias{test_snp}
\alias{tests}
\alias{tfam}
\alias{tfile}
\alias{to}
\alias{to_bp}
\alias{to_kb}
\alias{to_mb}
\alias{tped}
\alias{tucc}
\alias{twolocus}
\alias{update_alleles}
\alias{update_chr}
\alias{update_cm}
\alias{update_ids}
\alias{update_map}
\alias{update_name}
\alias{update_parents}
\alias{update_sex}
\alias{whap}
\alias{window}
\alias{with_phenotype}
\alias{within}
\alias{within}
\alias{write_covar}
\alias{write_snplist}
\alias{zero_cluster}
\title{PLINK flags}
\usage{
one(.string, ...)

epi1(.string, ...)

hwe2(.string, ...)

mh2(.string, ...)

p2(.string, ...)

r2(.string, ...)

t2(.string, ...)

adjust(.string, ...)

all_pheno(.string, ...)

allele1234(.string, ...)

alleleacgt(.string, ...)

allow_no_sex(.string, ...)

alt_group(.string, ...)

alt_snp(.string, ...)

annotate(.string, ...)

aperm(.string, ...)

assoc(.string, ...)

bd(.string, ...)

bed(.string, ...)

beta(.string, ...)

bfile(.string, ...)

bim(.string, ...)

bmerge(.string, ...)

case_only(.string, ...)

cc(.string, ...)

cell(.string, ...)

cfile(.string, ...)

chap(.string, ...)

check_sex(.string, ...)

chr(.string, ...)

ci(.string, ...)

clump(.string, ...)

clump_allow_overlap(.string, ...)

clump_annotate(.string, ...)

clump_best(.string, ...)

clump_field(.string, ...)

clump_index_first(.string, ...)

clump_kb(.string, ...)

clump_p1(.string, ...)

clump_p2(.string, ...)

clump_r2(.string, ...)

clump_range(.string, ...)

clump_range_border(.string, ...)

clump_replicate(.string, ...)

clump_verbose(.string, ...)

cluster(.string, ...)

cluster_missing(.string, ...)

cnv_blue(.string, ...)

cnv_border(.string, ...)

cnv_brown(.string, ...)

cnv_count(.string, ...)

cnv_del(.string, ...)

cnv_disrupt(.string, ...)

cnv_drop_no_segment(.string, ...)

cnv_dup(.string, ...)

cnv_exclude(.string, ...)

cnv_freq_excldue_above(.string, ...)

cnv_freq_excldue_below(.string, ...)

cnv_freq_excldue_exact(.string, ...)

cnv_freq_incldue_exact(.string, ...)

cnv_freq_method2(.string, ...)

cnv_green(.string, ...)

cnv_indiv_perm(.string, ...)

cnv_intersect(.string, ...)

cnv_kb(.string, ...)

cnv_list(.string, ...)

cnv_make_map(.string, ...)

cnv_max_kb(.string, ...)

cnv_max_score(.string, ...)

cnv_overlap(.string, ...)

cnv_red(.string, ...)

cnv_region_overlap(.string, ...)

cnv_report_regions(.string, ...)

cnv_score(.string, ...)

cnv_seglist(.string, ...)

cnv_subset(.string, ...)

cnv_test_2sided(.string, ...)

cnv_test_region(.string, ...)

cnv_test_window(.string, ...)

cnv_track(.string, ...)

cnv_union_overlap(.string, ...)

cnv_unique(.string, ...)

cnv_verbose_report_regions(.string, ...)

cnv_write(.string, ...)

cnv_write_freq(.string, ...)

compound_genotypes(.string, ...)

condition(.string, ...)

condition_list(.string, ...)

control(.string, ...)

counts(.string, ...)

covar(.string, ...)

covar_name(.string, ...)

covar_number(.string, ...)

cow(.string, ...)

dfam(.string, ...)

distance_matrix(.string, ...)

dog(.string, ...)

dominant(.string, ...)

dummy(.string, ...)

dummy_coding(.string, ...)

each_versus_others(.string, ...)

each_vs_others(.string, ...)

epistasis(.string, ...)

exclude(.string, ...)

exclude_before_extract(.string, ...)

extract(.string, ...)

fam(.string, ...)

family(.string, ...)

fast_epistasis(.string, ...)

file(.string, ...)

filter(.string, ...)

filter_cases(.string, ...)

filter_controls(.string, ...)

filter_females(.string, ...)

filter_founders(.string, ...)

filter_males(.string, ...)

filter_nonfounders(.string, ...)

fisher(.string, ...)

flip(.string, ...)

flip_scan(.string, ...)

flip_subset(.string, ...)

freq(.string, ...)

from(.string, ...)

from_bp(.string, ...)

from_kb(.string, ...)

from_mb(.string, ...)

gap(.string, ...)

gene(.string, ...)

gene_list(.string, ...)

gene_list_border(.string, ...)

gene_report(.string, ...)

gene_report_empty(.string, ...)

gene_subset(.string, ...)

genedrop(.string, ...)

genepi(.string, ...)

geno(.string, ...)

genome(.string, ...)

genome(.string, ...)

genotypic(.string, ...)

gfile(.string, ...)

gxe(.string, ...)

hap(.string, ...)

hap_assoc(.string, ...)

hap_freq(.string, ...)

hap_impute(.string, ...)

hap_max_phase(.string, ...)

hap_min_phase_prob(.string, ...)

hap_miss(.string, ...)

hap_phase(.string, ...)

hap_phase_wide(.string, ...)

hap_pp(.string, ...)

hap_snps(.string, ...)

hap_tdt(.string, ...)

hap_window(.string, ...)

hardy(.string, ...)

hardy2(.string, ...)

help(.string, ...)

het(.string, ...)

homog(.string, ...)

homozyg_group(.string, ...)

homozyg_het(.string, ...)

homozyg_kb(.string, ...)

homozyg_match(.string, ...)

homozyg_snp(.string, ...)

homozyg_verbose(.string, ...)

horse(.string, ...)

hwe(.string, ...)

hwe_all(.string, ...)

ibm(.string, ...)

impossible(.string, ...)

impute_sex(.string, ...)

indep(.string, ...)

indep_pairwise(.string, ...)

independent_effect(.string, ...)

interaction(.string, ...)

keep(.string, ...)

keep_allele_order(.string, ...)

keep_before_remove(.string, ...)

lambda(.string, ...)

ld_window(.string, ...)

lfile(.string, ...)

liability(.string, ...)

linear(.string, ...)

list(.string, ...)

logistic(.string, ...)

lookup(.string, ...)

lookup_gene(.string, ...)

lookup_list(.string, ...)

loop_assoc(.string, ...)

maf(.string, ...)

make_bed(.string, ...)

make_founders(.string, ...)

make_pheno(.string, ...)

make_pheno(.string, ...)

map(.string, ...)

map3(.string, ...)

match(.string, ...)

match_type(.string, ...)

matrix(.string, ...)

max(.string, ...)

max_maf(.string, ...)

mc(.string, ...)

mcc(.string, ...)

mcovar(.string, ...)

me(.string, ...)

mendel(.string, ...)

merge(.string, ...)

merge_list(.string, ...)

merge_mode(.string, ...)

meta_analysis(.string, ...)

mfilter(.string, ...)

mh(.string, ...)

mhf(.string, ...)

min(.string, ...)

mind(.string, ...)

missing(.string, ...)

missing_genotype(.string, ...)

missing_phenotype(.string, ...)

model(.string, ...)

model_fisher(.string, ...)

model_dom(.string, ...)

model_gen(.string, ...)

model_rec(.string, ...)

model_trend(.string, ...)

mouse(.string, ...)

mperm(.string, ...)

mpheno(.string, ...)

must_have_sex(.string, ...)

mwithin(.string, ...)

neighbour(.string, ...)

no_fid(.string, ...)

no_parents(.string, ...)

no_pheno(.string, ...)

no_sex(.string, ...)

nonfounders(.string, ...)

nop(.string, ...)

nudge(.string, ...)

null_group(.string, ...)

null_snp(.string, ...)

oblig_cluster(.string, ...)

oblig_missing(.string, ...)

out(.string, ...)

output_missing_genotype(.string, ...)

output_missing_phenotype(.string, ...)

parameters(.string, ...)

parentdt1(.string, ...)

parentdt2(.string, ...)

ped(.string, ...)

pedigree(.string, ...)

perm(.string, ...)

pheno(.string, ...)

pheno_name(.string, ...)

plist(.string, ...)

poo(.string, ...)

ppc(.string, ...)

ppc_gap(.string, ...)

proxy_assoc(.string, ...)

proxy_b_kb(.string, ...)

proxy_b_maxsnp(.string, ...)

proxy_b_r2(.string, ...)

proxy_b_threshold(.string, ...)

proxy_b_window(.string, ...)

proxy_dosage(.string, ...)

proxy_drop(.string, ...)

proxy_flanking(.string, ...)

proxy_geno(.string, ...)

proxy_genotypic_concordance(.string, ...)

proxy_glm(.string, ...)

proxy_impute(.string, ...)

proxy_impute_threshold(.string, ...)

proxy_kb(.string, ...)

proxy_list(.string, ...)

proxy_maf(.string, ...)

proxy_maxsnp(.string, ...)

proxy_mhf(.string, ...)

proxy_r2(.string, ...)

proxy_r2_no_filter(.string, ...)

proxy_replace(.string, ...)

proxy_show_proxies(.string, ...)

proxy_sub_maxsnp(.string, ...)

proxy_sub_r2(.string, ...)

proxy_tdt(.string, ...)

proxy_verbose(.string, ...)

proxy_window(.string, ...)

prune(.string, ...)

qmatch(.string, ...)

qq_plot(.string, ...)

qt(.string, ...)

qual_geno_max_threshold(.string, ...)

qual_geno_scores(.string, ...)

qual_geno_threshold(.string, ...)

qual_max_threshold(.string, ...)

qual_scores(.string, ...)

qual_threshold(.string, ...)

r(.string, ...)

rank(.string, ...)

read_genome(.string, ...)

recessive(.string, ...)

recode(.string, ...)

recode_bimbam(.string, ...)

recode_fastphase(.string, ...)

recode_lgen(.string, ...)

recode_rlist(.string, ...)

recode_structure(.string, ...)

recode12(.string, ...)

recodea(.string, ...)

recodead(.string, ...)

recodehv(.string, ...)

reference_allele(.string, ...)

rel_check(.string, ...)

remove(.string, ...)

script(.string, ...)

set(.string, ...)

set_by_all(.string, ...)

set_hh_missing(.string, ...)

set_max(.string, ...)

set_me_missing(.string, ...)

set_p(.string, ...)

set_r2(.string, ...)

set_table(.string, ...)

set_test(.string, ...)

sex(.string, ...)

sheep(.string, ...)

silent(.string, ...)

simulate(.string, ...)

simulate_haps(.string, ...)

simulate_label(.string, ...)

simulate_ncases(.string, ...)

simulate_ncontrols(.string, ...)

simulate_prevalence(.string, ...)

simulate_qt(.string, ...)

simulate_tags(.string, ...)

snp(.string, ...)

snps(.string, ...)

specific_haplotype(.string, ...)

subset(.string, ...)

swap_parents(.string, ...)

swap_sibs(.string, ...)

swap_unrel(.string, ...)

tab(.string, ...)

tdt(.string, ...)

test_all(.string, ...)

test_mishap(.string, ...)

test_missing(.string, ...)

test_snp(.string, ...)

tests(.string, ...)

tfam(.string, ...)

tfile(.string, ...)

to(.string, ...)

to_bp(.string, ...)

to_kb(.string, ...)

to_mb(.string, ...)

tped(.string, ...)

tucc(.string, ...)

twolocus(.string, ...)

update_alleles(.string, ...)

update_chr(.string, ...)

update_cm(.string, ...)

update_ids(.string, ...)

update_map(.string, ...)

update_name(.string, ...)

update_parents(.string, ...)

update_sex(.string, ...)

whap(.string, ...)

window(.string, ...)

with_phenotype(.string, ...)

within(.string, ...)

within(.string, ...)

write_covar(.string, ...)

write_snplist(.string, ...)

zero_cluster(.string, ...)
}
\arguments{
\item{.string}{A character string.}

\item{...}{Other arguments.}
}
\description{
A list of all PLINK flags defined in \code{plinkr}
}
