% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_additive_phenotype_values.R
\name{calc_additive_phenotype_values}
\alias{calc_additive_phenotype_values}
\title{Calculate the phenotypic values for a trait that is perfectly additive.}
\usage{
calc_additive_phenotype_values(snvs)
}
\arguments{
\item{snvs}{a \link[tibble]{tibble} that contains the
two nucleotide calls for multiple individuals.
Each column is a haplotypes, hence, for a diploid organism,
there are two columns.
Each individual is represented by a row.
Nucleotides are in uppercase.
Per SNV table, there can be only two different nucleotides,
as this is how a SNP works; a SNP from A to C is a different one (and
has a different SNP ID) than a SNP that encodes a mutation
for A to G.}
}
\description{
Calculate the phenotypic values for a trait that is perfectly additive.
A phenotypic value starts at 10.0.
}
\details{
\itemize{
\item each \code{A} denotes an increase of 0.0
\item each \code{T} denotes an increase of 0.5
}

The SNVs ('Single Nucleotide Variants') must reflect this genetic
architecture, i.e. must consist out of adenines and thymines
}
\note{
the reason a phenotypic value starts at 10.0, instead of 0.0,
is due to PLINK: if phenotypic values are all only zeroes and ones
(and twos), PLINK will interpret these as case-control codes.
There is no way to force PLINK to do a quantitative trait analysis on
zeroes and ones.

As a workaround, phenotypic traits start from ten.
}
\author{
Richèl J.C. Bilderbeek
}
