% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dominant_phenotype_values.R
\name{calc_dominant_phenotype_values}
\alias{calc_dominant_phenotype_values}
\title{Calculate the phenotypic values for a trait that is dominant.}
\usage{
calc_dominant_phenotype_values(
  snvs,
  phenotype_value_recessive = 10,
  phenotype_value_dominant = 11
)
}
\arguments{
\item{snvs}{a \link[tibble]{tibble} that contains the
two nucleotide calls for multiple individuals.
Each column is a haplotype, hence, for a diploid organism,
there are two columns.
Each individual is represented by a row.
Nucleotides are in uppercase.
Per SNV table, there can be only two different nucleotides,
as this is how a SNP works; a SNP from A to C is a different one (and
has a different SNP ID) than a SNP that encodes a mutation
for A to G.
Use \link{create_snvs} to create a \code{snvs}.}

\item{phenotype_value_recessive}{phenotypic value
for the recessive variant, i.e. that genotype that is homozygous for the
rare allele}

\item{phenotype_value_dominant}{phenotypic value
for the dominant variant, i.e. that genotype that has at least one
version of the common common allele}
}
\description{
Calculate the phenotypic values for a trait that is dominant.
It is assumed that adenine is the dominant genotype,
and the other nucleotides are recessive.
Per SNP, if there is at least one adenine,
the phenotypic value is \code{phenotype_value_dominant}.
else the phenotypic value is \code{phenotype_value_recessive}.
}
\details{
If there are more SNPs, the phenotypic values are averaged.
}
\note{
the reason a these phenotypic value are 10.0 and 11.0,
instead of -say- 0.0 and 1.0,
is due to PLINK: if phenotypic values are all only zeroes and ones
(and twos), PLINK will interpret these as case-control codes.
There is no way to force PLINK to do a quantitative trait analysis on
zeroes and ones.

As a workaround, phenotypic values start from ten.
}
\author{
Richèl J.C. Bilderbeek
}
