% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_epistatic_phenotype_values.R
\name{calc_epistatic_phenotype_values}
\alias{calc_epistatic_phenotype_values}
\title{Calculate the phenotypic values for a trait that is epistatic.}
\usage{
calc_epistatic_phenotype_values(snvs)
}
\arguments{
\item{snvs}{a \link[tibble]{tibble} that contains the
two nucleotide calls for multiple individuals.
Each column is a haplotypes, hence, for a diploid organism,
there are two columns.
Each individual is represented by a row.
Nucleotides are in uppercase.
Per SNV table, there can be only two different nucleotides,
as this is how a SNP works; a SNP from A to C is a different one (and
has a different SNP ID) than a SNP that encodes a mutation
for A to G.}
}
\description{
Calculate the phenotypic values for a trait that is epistatic.
}
\details{
A phenotypic value starts at 20.0.
Only genotypes without adenines results in a phenotype of 21.0
}
\note{
the reason a phenotypic value starts at 20.0, instead of 0.0,
is due to PLINK: if phenotypic values are all only zeroes and ones
(and twos), PLINK will interpret these as case-control codes.
There is no way to force PLINK to do a quantitative trait analysis on
zeroes and ones.

As a workaround, epistatic traits start from twenty.
}
\author{
Richèl J.C. Bilderbeek
}
