% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_runif_phenotype_values.R
\name{calc_runif_phenotype_values}
\alias{calc_runif_phenotype_values}
\title{Calculate the phenotypic values for a trait that is random and uniform.}
\usage{
calc_runif_phenotype_values(snvs)
}
\arguments{
\item{snvs}{a \link[tibble]{tibble} that contains the
two nucleotide calls for multiple individuals.
Each column is a haplotypes, hence, for a diploid organism,
there are two columns.
Each individual is represented by a row.
Nucleotides are in uppercase.
Per SNV table, there can be only two different nucleotides,
as this is how a SNP works; a SNP from A to C is a different one (and
has a different SNP ID) than a SNP that encodes a mutation
for A to G.}
}
\description{
Calculate the phenotypic values for a trait that is random and uniform,
as produced by \link[stats]{runif}.
}
\examples{
t <- tibble::tibble(
  snv_1a = c("A", "A", "C", "C"),
  snv_1b = c("A", "C", "A", "C")
)
calc_runif_phenotype_values(t)
}
\author{
Richèl J.C. Bilderbeek
}
