% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_plink_version_and_data_can_work_together.R
\name{can_plink_version_and_data_can_work_together}
\alias{can_plink_version_and_data_can_work_together}
\title{Determine if the \code{data} and the version of \code{PLINK}/\code{PLINK2} can work together}
\usage{
can_plink_version_and_data_can_work_together(
  data,
  plink_options,
  verbose = FALSE
)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the \code{data} and the version
of \code{PLINK}/\code{PLINK2} can work together
}
\description{
Determine if the \code{data} and the version of \code{PLINK}/\code{PLINK2} can work together
}
\author{
Richèl J.C. Bilderbeek
}
