% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assoc_params.R
\name{check_assoc_params}
\alias{check_assoc_params}
\title{Check if the \code{assoc_params} are valid, will stop otherwise}
\usage{
check_assoc_params(assoc_params)
}
\value{
Nothing.
}
\description{
Check if the \code{assoc_params} are valid, will stop otherwise
}
\note{
This function is named after the \code{--assoc} PLINK flag.
}
\author{
Richèl J.C. Bilderbeek
}
