% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assoc_qt_covar_params.R
\name{check_assoc_qt_covar_params}
\alias{check_assoc_qt_covar_params}
\title{Check if the \code{assoc_qt_covar_params} are valid, will stop otherwise}
\usage{
check_assoc_qt_covar_params(assoc_qt_covar_params)
}
\arguments{
\item{assoc_qt_covar_params}{parameters to do an association analysis
for a quantitative trait with covariates
(i.e. using \link{assoc_qt_covar}),
as can be created by \link{create_assoc_qt_covar_params}}
}
\value{
Nothing.
}
\description{
Check if the \code{assoc_qt_covar_params} are valid, will stop otherwise
}
\note{
This function is named after the \code{--assoc} PLINK flag.
}
\examples{
assoc_qt_covar_params <- create_test_assoc_qt_covar_params()
check_assoc_qt_covar_params(assoc_qt_covar_params = assoc_qt_covar_params)
}
\author{
Richèl J.C. Bilderbeek
}
