% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assoc_qt_data.R
\name{check_assoc_qt_data}
\alias{check_assoc_qt_data}
\title{Check if the \code{assoc_qt_data} is valid}
\usage{
check_assoc_qt_data(assoc_qt_data)
}
\arguments{
\item{assoc_qt_data}{data use to used by \link{assoc_qt},
which equals 'regular' data (see \link{check_data}) and a
phenotype table/file}
}
\description{
Check if the \code{assoc_qt_data} is valid.
}
\details{
\code{assoc_qt_data} is a list combined of:
\itemize{
\item 'regular' data, as checked by \link{check_data}
\item phenotype data, as checked by \link{check_phenotype_data}
}

Will \link{stop} if the \code{assoc_qt_data} is invalid
}
\author{
Richèl J.C. Bilderbeek
}
