% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assoc_qt_result.R
\name{check_assoc_qt_result}
\alias{check_assoc_qt_result}
\title{Check if a \code{assoc_qt_result} is valid.}
\usage{
check_assoc_qt_result(assoc_qt_result)
}
\arguments{
\item{assoc_qt_result}{the result of \link{assoc_qt},
which is a \link{list} with elements:
\itemize{
\item \code{qassoc_table}: the quantitative analysis results table,
as can be checked by \link{check_qassoc_table}
\item \code{log}: the text from the log file created by \code{PLINK}/\code{PLINK2}
when doing \link{assoc_qt}
}}
}
\value{
Nothing. Will \link{stop} if the \code{assoc_qt_result} is invalid
}
\description{
Check if a \code{assoc_qt_result} is valid.
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
