% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assoc_table.R
\name{check_assoc_table}
\alias{check_assoc_table}
\title{Check if a \code{.assoc} table is valid.}
\usage{
check_assoc_table(assoc_table)
}
\value{
Nothing. Will \link{stop} if the \code{.assoc} is invalid
}
\description{
Check if a \code{.assoc} table is valid.
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
