% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_base_output_filename.R
\name{check_base_output_filename}
\alias{check_base_output_filename}
\title{Check if the \code{base_output_filename} is valid.
Check if the \code{base_output_filename},
which is a filename without an extension}
\usage{
check_base_output_filename(base_output_filename)
}
\arguments{
\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}
}
\value{
Nothing. Will \link{stop} if the output
is a filename
}
\description{
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
