% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bed_table.R
\name{check_bed_table}
\alias{check_bed_table}
\title{Check if a \code{.bed} table is valid.}
\usage{
check_bed_table(bed_table)
}
\arguments{
\item{bed_table}{a table that maps the SNPs to the individuals,
of which the column names are the names of the individuals,
the row names are the names of the SNPs,
and the values are the SNP variant.
Use \link{get_test_bed_table} to get a \code{.bed} table as used in testing.
Use \link{read_plink_bed_file} to read a \code{PLINK} \code{.bed} file.
Use \link{check_bed_table} to test if a \code{.bed} table is valid.}
}
\value{
Nothing. Will \link{stop} if the \code{bed_table} is invalid
}
\description{
Check if a \code{.bed} table is valid.
Will \link{stop} if not
}
\examples{
check_bed_table(get_test_bed_table())
}
\author{
Richèl J.C. Bilderbeek
}
