% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bim_table.R
\name{check_bim_table}
\alias{check_bim_table}
\title{Check if a \code{.bim} table is valid.}
\usage{
check_bim_table(bim_table)
}
\arguments{
\item{bim_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{chr}: the chromosome number
\item \code{id}: the SNP ID
\item \code{posg}: the position
\item \code{pos}: the position
\item \code{ref}: something
\item \code{alt}: something
}

Use \link{get_test_bim_table} to get a \code{.bim} table as used in testing.
Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.
Use \link{check_bim_table} to test if a \code{.bim} table is valid.}
}
\value{
Nothing. Will \link{stop} if the \code{bim_table} is invalid
}
\description{
Check if a \code{.bim} table is valid.
Will \link{stop} if not
}
\examples{
check_bim_table(get_test_bim_table())
check_bim_table(read_plink_bim_file(get_plinkr_filename("toy_data.bim")))
}
\author{
Richèl J.C. Bilderbeek
}
