% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_calc_phenotype_function.R
\name{check_calc_phenotype_function}
\alias{check_calc_phenotype_function}
\title{Check if the \code{calc_phenotype_function} is valid.
Check if the \code{calc_phenotype_function}
is valid.}
\usage{
check_calc_phenotype_function(calc_phenotype_function)
}
\value{
Nothing. Will \link{stop} if the input
\code{calc_phenotype_function} is invalid
}
\description{
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
