% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chromosome_number.R
\name{check_chromosome_number}
\alias{check_chromosome_number}
\title{Check if the chromosome number is valid.
Check if the chromosome number is valid.
Will \link{stop} if not}
\usage{
check_chromosome_number(chromosome_number)
}
\arguments{
\item{chromosome_number}{the chromosome number,
as can be checked by \link{check_chromosome_number}}
}
\value{
Nothing. Will \link{stop} if the chromosome number is invalid
}
\description{
Check if the chromosome number is valid.
Check if the chromosome number is valid.
Will \link{stop} if not
}
\examples{
check_chromosome_number(1)
check_chromosome_number(2)
}
\author{
Richèl J.C. Bilderbeek
}
