% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chromosome_selector.R
\name{check_chromosome_selector}
\alias{check_chromosome_selector}
\title{Check if a chromosome selector is valid.}
\usage{
check_chromosome_selector(chromosome_selector)
}
\arguments{
\item{chromosome_selector}{a SNP selector (see \link{create_snps_selector})
that allows one to select a single chromsome,
as created by \link{create_chromosome_selector}
and checked by \link{check_chromosome_selector}}
}
\value{
Nothing.
}
\description{
Check if a chromosome selector is valid.
Will \link{stop} if not.
}
\examples{
check_chromosome_selector(create_test_chromosome_selector())
}
\seealso{
There are multiple SNP selector checking functions,
see \link{create_snps_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
