% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty_plinkr_folder.R
\name{check_empty_plinkr_folder}
\alias{check_empty_plinkr_folder}
\title{Check there are no files in the default \code{PLINK} folder}
\usage{
check_empty_plinkr_folder(plinkr_folder = get_plinkr_folder())
}
\value{
Nothing.
}
\description{
Check there are no files in the default \code{PLINK} folder.
The goal is to make sure no temporary files are left undeleted.
Will \link{stop} if there are files in the \link{plinkr} folder
}
\author{
Richèl J.C. Bilderbeek
}
