% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_equal_number_of_snvs.R
\name{check_equal_number_of_snvs}
\alias{check_equal_number_of_snvs}
\title{Internal function}
\usage{
check_equal_number_of_snvs(data)
}
\arguments{
\item{data}{PLINK or PLINK2 text or binary data:
\itemize{
\item PLINK text data:
a \code{.ped} table (see \link{read_plink_ped_file})
and a \code{.map} (see \link{read_plink_map_file}),
as created by \link{create_plink_text_data}.
\item PLINK binary data:
a \code{.bed} table (see \link{read_plink_bed_file})
a \code{.bim} table (see \link{read_plink_bim_file})
and a \code{.fam} (see \link{read_plink_fam_file}),
as created by \link{create_plink_bin_data}.
\item PLINK2 binary data: use \link{create_plink2_bin_data}
}}
}
\description{
Check if the data has the same number of
SNVs in its tables.
Will \link{stop} if not.
}
\examples{
check_equal_number_of_snvs(create_test_plink_text_data())
check_equal_number_of_snvs(create_test_plink_bin_data())
check_equal_number_of_snvs(create_test_plink2_bin_data())
}
\author{
Richèl J.C. Bilderbeek
}
