% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fam_filename.R
\name{check_fam_filename}
\alias{check_fam_filename}
\title{Checks that the \code{fam_filename} is valid}
\usage{
check_fam_filename(fam_filename)
}
\arguments{
\item{fam_filename}{name of a \code{PLINK} \code{.fam} file
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.}
}
\value{
nothing
}
\description{
Checks that the \code{fam_filename} is valid.
Calls \link{stop} if not
}
\examples{
check_fam_filename("a.fam")
check_fam_filename("/folder/a.fam")
check_fam_filename("~/a.fam")
}
\author{
Richèl J.C. Bilderbeek
}
