% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fid.R
\name{check_fid}
\alias{check_fid}
\title{Check if a family ID is valid}
\usage{
check_fid(fid)
}
\arguments{
\item{fid}{the family ID, which is called \code{FID} in \code{PLINK}.
Use \link{check_fid} to check if a \code{fid} is valid.}
}
\value{
nothing
}
\description{
Check if a family ID is valid.
Will \link{stop} if not
}
\examples{
check_fid("1")
check_fid("Doe")
}
\author{
Richèl J.C. Bilderbeek
}
