% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_iid.R
\name{check_iid}
\alias{check_iid}
\title{Check if a within-family ID is valid}
\usage{
check_iid(iid)
}
\value{
nothing
}
\description{
Check if a within-family ID is valid.
Will \link{stop} if not
}
\examples{
check_iid("1")
check_iid("Doe")
}
\author{
Richèl J.C. Bilderbeek
}
