% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_iids.R
\name{check_iids}
\alias{check_iids}
\title{Check if one or more within-family IDs are valid}
\usage{
check_iids(iids)
}
\arguments{
\item{iids}{one or more within-family IDs,
as can be checked by \link{check_iids}}
}
\value{
nothing
}
\description{
Check if one or more within-family IDs are valid
Will \link{stop} if not.
}
\examples{
check_iids("1")
check_iids("Doe")
}
\author{
Richèl J.C. Bilderbeek
}
