% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ld_window_r2.R
\name{check_ld_window_r2}
\alias{check_ld_window_r2}
\title{Check if the \code{ld_window_r2} is valid.
Check if the \code{ld_window_r2} is valid,
which is one number in range zero (including zero)
to one (including one)}
\usage{
check_ld_window_r2(ld_window_r2)
}
\arguments{
\item{ld_window_r2}{the minimal squared correlation coefficient (also
known as \code{r^2}) between markers. This parameter is named after the
\code{--ld_window_r2} flag, as documented at
\url{https://www.cog-genomics.org/plink/1.9/ld}.}
}
\value{
Nothing. Will \link{stop} if the input
is not one number in range zero
to one
}
\description{
Will \link{stop} if not
}
\examples{
check_ld_window_r2(0.0)
check_ld_window_r2(0.2)
check_ld_window_r2(1.0)
}
\author{
Richèl J.C. Bilderbeek
}
