% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_maf.R
\name{check_maf}
\alias{check_maf}
\title{Check if the \code{maf} is valid.
Check if the \code{maf}, an abbreviation of 'Minor Allele
Frequency' is valid, which is when it is a number between
0.0 and 0.5 (i.e. excluding 0.0 and 0.5).}
\usage{
check_maf(maf)
}
\arguments{
\item{maf}{minor allele frequency}
}
\value{
Nothing. Will \link{stop} if the input
is not number between
0.0 and 0.5
}
\description{
Will \link{stop} if not
}
