% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_individuals.R
\name{check_n_individuals}
\alias{check_n_individuals}
\title{Check if the number of individual is valid.}
\usage{
check_n_individuals(n_individuals)
}
\arguments{
\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}
}
\description{
Check if the number of individual is valid.
Will \link{stop} if not
}
