% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_samples.R
\name{check_n_samples}
\alias{check_n_samples}
\title{Check if the number of samples is valid.
Check if the number of samples is valid.
Will \link{stop} if not}
\usage{
check_n_samples(n_samples)
}
\arguments{
\item{n_samples}{the number of samples/individuals,
as can be checked by \link{check_n_samples}}
}
\value{
Nothing. Will \link{stop} if the number of samples is invalid
}
\description{
Check if the number of samples is valid.
Check if the number of samples is valid.
Will \link{stop} if not
}
\examples{
check_n_samples(0)
check_n_samples(1)
check_n_samples(2)
}
\author{
Richèl J.C. Bilderbeek
}
