% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_snps.R
\name{check_n_snps}
\alias{check_n_snps}
\title{Check if the number of SNPs is valid.
Check if the number of SNPs is valid.
Will \link{stop} if not}
\usage{
check_n_snps(n_snps)
}
\arguments{
\item{n_snps}{the number of SNPs}
}
\value{
Nothing. Will \link{stop} if the number of SNPs is invalid
}
\description{
Check if the number of SNPs is valid.
Check if the number of SNPs is valid.
Will \link{stop} if not
}
\examples{
check_n_snps(1)
check_n_snps(2)
}
\author{
Richèl J.C. Bilderbeek
}
