% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_os.R
\name{check_os}
\alias{check_os}
\title{Checks if the operating system is supported}
\usage{
check_os(os)
}
\value{
nothing. Will stop if the OS is unsupported
}
\description{
Checks if the operating system is supported
}
\examples{
check_os("mac")
check_os("unix")
check_os("win")
}
\author{
Richèl J.C. Bilderbeek
}
