% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pca_result_filenames.R
\name{check_pca_result_filenames}
\alias{check_pca_result_filenames}
\title{Check if a \code{pca_result_filenames} is valid.}
\usage{
check_pca_result_filenames(pca_result_filenames)
}
\value{
Nothing. Will \link{stop} if the \code{pca_result_filenames} is invalid
}
\description{
Check if a \code{pca_result_filenames} is valid.
Will \link{stop} if not
}
\examples{
check_pca_result_filenames(create_test_pca_result_filenames())
}
\author{
Richèl J.C. Bilderbeek
}
