% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ped_table.R
\name{check_ped_table}
\alias{check_ped_table}
\title{Check if a \code{.ped} is valid. Will \link{stop} if not}
\usage{
check_ped_table(ped_table)
}
\arguments{
\item{ped_table}{a table obtained by reading a PLINK \code{.ped} file
Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.}
}
\value{
Nothing.
Will \link{stop} if the pedigree table is invalid.
}
\description{
Check if a \code{.ped} is valid. Will \link{stop} if not
}
