% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pgen_table.R
\name{check_pgen_table}
\alias{check_pgen_table}
\title{Check if a \code{.pgen} table is valid.}
\usage{
check_pgen_table(pgen_table)
}
\arguments{
\item{pgen_table}{a \link{matrix} that maps the individuals
to their SNPs. TODO}
}
\value{
Nothing. Will \link{stop} if the \code{.pgen} is invalid
}
\description{
Check if a \code{.pgen} table is valid.
Will \link{stop} if not
}
\examples{
check_pgen_table(get_test_pgen_table())
}
\author{
Richèl J.C. Bilderbeek
}
