% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phe_file_ok_for_qt.R
\name{check_phe_file_ok_for_qt}
\alias{check_phe_file_ok_for_qt}
\title{Check if a phenotype files holds a table that is valid
for use in a quantitative trait analysis}
\usage{
check_phe_file_ok_for_qt(phe_filename)
}
\arguments{
\item{phe_filename}{name of a phenotype (\code{.phe}) file,
as checked by \link{check_phe_filename}.
Use \link{read_plink_phe_file} to read a raw phenotype file.}
}
\value{
Nothing. Will \link{stop} if the phenotype table in the file
is invalid for use in a quantitative trait analysis
}
\description{
Check if a phenotype files holds a table that is valid
for use in a quantitative trait analysis.
Will \link{stop} if not
}
\details{
For a phenotype table to be valid for use in a quantitative trait analysis,
in each column, at least 1 value must differ from 0, 1, and 2.
This is because \code{PLINK} will assume a case-control analysis
when all values in a column are 0, 1 and 2.
}
\examples{
phe_filename <- get_plinkr_filename("demo_assoc_qt.phe")
check_phe_file_ok_for_qt(phe_filename)
}
\seealso{
Use \link{check_phe_table_ok_for_qt} to check if an in-memory
phenotype table is valid
for use in a quantitative trait analysis
}
\author{
Richèl J.C. Bilderbeek
}
