% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phe_table.R
\name{check_phe_table}
\alias{check_phe_table}
\title{Check if a phenotype table is valid.}
\usage{
check_phe_table(phe_table)
}
\arguments{
\item{phe_table}{a table of phenotypes.

A \code{phe_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_phe_file} to read a phenotype file.
Use \link{check_phe_table} to test if a phenotype table is valid.}
}
\value{
Nothing. Will \link{stop} if the \code{phe_table} is invalid
}
\description{
Check if a phenotype table is valid.
Will \link{stop} if not
}
\examples{
check_phe_table(get_test_phe_table())
}
\seealso{
use \link{check_phe_table_ok_for_qt} to check if the \code{phe_table}
is valid for a quantitative trait analysis.
}
\author{
Richèl J.C. Bilderbeek
}
