% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phe_table_ok_for_qt.R
\name{check_phe_table_ok_for_qt}
\alias{check_phe_table_ok_for_qt}
\title{Check if a phenotype table is valid for use in a quantitative trait analysis}
\usage{
check_phe_table_ok_for_qt(phe_table)
}
\arguments{
\item{phe_table}{a table of phenotypes.

A \code{phe_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_phe_file} to read a phenotype file.
Use \link{check_phe_table} to test if a phenotype table is valid.}
}
\value{
Nothing. Will \link{stop} if the \code{phe_table} is invalid for
use in a quantitative trait analysis
}
\description{
Check if a phenotype table is valid for use in a quantitative trait analysis.
Will \link{stop} if not
}
\details{
For a phenotype table to be valid for use in a quantitative trait analysis,
in each column, at least 1 value must differ from 0, 1, and 2.
This is because \code{PLINK} will assume a case-control analysis
when all values in a column are 0, 1 and 2.
}
\examples{
check_phe_table_ok_for_qt(get_test_phe_table())
}
\seealso{
use \link{check_phe_table} to check if the \code{phe_table}
is valid in general
}
\author{
Richèl J.C. Bilderbeek
}
