% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phenotype_data_filename.R
\name{check_phenotype_data_filename}
\alias{check_phenotype_data_filename}
\title{Check if the \code{phenotype_data_filename} is valid.}
\usage{
check_phenotype_data_filename(phenotype_data_filename)
}
\arguments{
\item{phenotype_data_filename}{phenotype data
(as checked by \link{check_phenotype_data}))
in the form of a \link{list} with a filename to a phenotype table}
}
\value{
nothing
}
\description{
Check if the \code{phenotype_data_filename} is valid.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
