% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phenotypes.R
\name{check_phenotypes}
\alias{check_phenotypes}
\title{Check if the phenotypes are indeed phenotypes with a clear genetic setup.}
\usage{
check_phenotypes(phenotypes)
}
\arguments{
\item{phenotypes}{one ore more phenotypes,
named after their genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive,
see \link{calc_additive_phenotype_values} for the exact calculation
\item \code{epistatic} the phenotype is epistatic,
see \link{calc_epistatic_phenotype_values} for the exact calculation
}}
}
\value{
Nothing
Will \link{stop} if the phenotypes are not
in the supported list of phenotypes with a clear genetic setup,
as can be obtained with \link{get_phenotypes}
}
\description{
Check if the phenotypes are indeed phenotypes with a clear genetic setup.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
