% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink2_bin_data.R
\name{check_plink2_bin_data}
\alias{check_plink2_bin_data}
\title{Check if the \code{PLINK2} bin data is valid.}
\usage{
check_plink2_bin_data(plink2_bin_data)
}
\arguments{
\item{plink2_bin_data}{the genetic and phenotypic data for \code{PLINK2} to work
on, in \code{PLINK2} binary format, as created by \link{create_plink2_bin_data}.}
}
\value{
nothing
}
\description{
Check if the \code{PLINK2} bin data is valid.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
