% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink2_bin_filenames.R
\name{check_plink2_bin_filenames}
\alias{check_plink2_bin_filenames}
\title{Checks that the \code{plink2_bin_filenames} is valid}
\usage{
check_plink2_bin_filenames(plink2_bin_filenames)
}
\value{
nothing
}
\description{
Checks that the \code{plink2_bin_filenames} is valid.
Calls \link{stop} if not
}
\examples{
check_plink2_bin_filenames(
  create_plink2_bin_filenames(
    pgen_filename = "file.pgen",
    psam_filename = "file.psam",
    pvar_filename = "file.pvar"
  )
)
}
\seealso{
\code{plinkr} has can check the following check sets of filenames,
depending on the type of files:
\itemize{
\item \code{PLINK} text files: use \link{check_plink_text_filenames}
\item \code{PLINK} binary files: use \link{check_plink_bin_filenames}
\item \code{PLINK2} binary files: use \link{check_plink2_bin_filenames}
}
}
\author{
Richèl J.C. Bilderbeek
}
