% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink_bin_data.R
\name{check_plink_bin_data}
\alias{check_plink_bin_data}
\title{Check if the \code{PLINK} bin data is valid.}
\usage{
check_plink_bin_data(plink_bin_data)
}
\arguments{
\item{plink_bin_data}{the genetic and phenotypic data for \code{PLINK} to work
on, in \code{PLINK} binary format, as created by \link{create_plink_bin_data}.}
}
\value{
nothing
}
\description{
Check if the \code{PLINK} bin data is valid.
Will \link{stop} if not.
}
\examples{
plink_bin_data <- create_test_plink_bin_data()
check_plink_bin_data(plink_bin_data)
}
\author{
Richèl J.C. Bilderbeek
}
