% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink_filenames.R
\name{check_plink_filenames}
\alias{check_plink_filenames}
\title{Check if a \code{plink_filenames} is valid.}
\usage{
check_plink_filenames(plink_filenames)
}
\value{
Nothing. Will \link{stop} if the \code{plink_filenames} is invalid
}
\description{
Check if a \code{plink_filenames} is valid.
Will \link{stop} if not
}
\examples{
# For PLINK, a .phe table contains the sample IDs
phe_table <- get_test_phe_table()
}
\author{
Richèl J.C. Bilderbeek
}
