% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink_is_installed.R
\name{check_plink_is_installed}
\alias{check_plink_is_installed}
\title{Check if PLINK is installed}
\usage{
check_plink_is_installed(plink_options = create_plink_options())
}
\arguments{
\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
Nothing. Will \link{stop} if PLINK is not installed.
}
\description{
Check if PLINK is installed. Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
