% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink_options.R
\name{check_plink_options}
\alias{check_plink_options}
\title{Check if the options to run PLINK are valid.}
\usage{
check_plink_options(plink_options)
}
\arguments{
\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the PLINK options,
as can be checked by \link{check_plink_options}
}
\description{
Check if the options to run PLINK are valid.
Will \link{stop} if not.
}
\details{
Note that this function does not check if PLINK is actually present in
that folder, similar to \link{normalizePath}
with \code{mustWork} set to \link{FALSE}
}
\examples{
check_plink_options(create_plink_options())
check_plink_options(create_plink_v1_7_options())
check_plink_options(create_plink_v1_9_options())
check_plink_options(create_custom_plink_options("custom_plink_path"))
}
\author{
Richèl J.C. Bilderbeek
}
