% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink_optionses.R
\name{check_plink_optionses}
\alias{check_plink_optionses}
\title{Check if the multiple sets of options to run PLINK are valid.}
\usage{
check_plink_optionses(plink_optionses)
}
\arguments{
\item{plink_optionses}{a list of one or more options to run PLINK,
as created by \link{create_plink_optionses}.
The reduplicated plural was used to express this is a list
of \code{plink_options}, instead of one set of \code{plink_options}}
}
\value{
the PLINK options,
as can be checked by \link{check_plink_options}
}
\description{
Check if the multiple sets of options to run PLINK are valid.
Will \link{stop} if not.
}
\details{
Note that this function does not check if PLINK is actually present in
that folder, similar to \link{normalizePath}
with \code{mustWork} set to \link{FALSE}
}
\examples{
check_plink_optionses(create_plink_optionses())
}
\author{
Richèl J.C. Bilderbeek
}
