% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink_text_data.R
\name{check_plink_text_data}
\alias{check_plink_text_data}
\title{Check if the \code{PLINK} text data is valid.}
\usage{
check_plink_text_data(plink_text_data)
}
\arguments{
\item{plink_text_data}{the genetic and phenotypic data for \code{PLINK} to work
on, in \code{PLINK} text format, as created by \link{create_plink_text_data}.}
}
\value{
nothing
}
\description{
Check if the \code{PLINK} text data is valid.
Will \link{stop} if not.
}
\examples{
plink_text_data <- create_plink_text_data(
  ped_table = get_test_ped_table(),
  map_table = get_test_map_table()
)
check_plink_text_data(plink_text_data)
}
\seealso{
\code{plinkr} has the following functions to check if in-memory \code{data}
is valid:
\itemize{
\item \code{PLINK} text data: use \link{check_plink_text_data}
\item \code{PLINK} binary data: use \link{check_plink_bin_data}
\item \code{PLINK2} binary data: use \link{check_plink2_bin_data}
\item any of these: use \link{check_data}
}
}
\author{
Richèl J.C. Bilderbeek
}
