% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink_text_filenames.R
\name{check_plink_text_filenames}
\alias{check_plink_text_filenames}
\title{Checks that the \code{plink_text_filenames} is valid}
\usage{
check_plink_text_filenames(plink_text_filenames)
}
\value{
nothing
}
\description{
Checks that the \code{plink_text_filenames} is valid.
Calls \link{stop} if not
}
\examples{
check_plink_text_filenames(
  create_plink_text_filenames(
    map_filename = "file.map",
    ped_filename = "file.ped"
  )
)
}
\seealso{
\code{plinkr} has the following functions to check sets of filenames,
depending on the data:
\itemize{
\item \code{PLINK} text filenames: use \link{check_plink_text_filenames}
\item \code{PLINK} binary filenames: use \link{check_plink_bin_filenames}
\item \code{PLINK2} binary filenames: use \link{check_plink2_bin_filenames}
}
}
\author{
Richèl J.C. Bilderbeek
}
