% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_psam_table.R
\name{check_psam_table}
\alias{check_psam_table}
\title{Check if a \code{.psam} table is valid.}
\usage{
check_psam_table(psam_table)
}
\arguments{
\item{psam_table}{a \link[tibble]{tibble} with as much rows as
individuals. It has the following columns:
\itemize{
\item \code{FID}: the family ID
\item \code{IID}: the within-family ID
\item \code{SEX}: the sex
\item \code{PHENO1}: a phenotypic value
}

The uppercase column names are those as used by PLINK2.}
}
\value{
Nothing. Will \link{stop} if the \code{psam_table} is invalid
}
\description{
Check if a \code{.psam} table is valid.
Will \link{stop} if not
}
\examples{
check_psam_table(get_test_psam_table())
}
\author{
Richèl J.C. Bilderbeek
}
