% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pvar_table.R
\name{check_pvar_table}
\alias{check_pvar_table}
\title{Check if a \code{.pvar} table is valid.}
\usage{
check_pvar_table(pvar_table)
}
\arguments{
\item{pvar_table}{a \link[tibble]{tibble} with as much rows as
SNPs. It has the following columns:
\itemize{
\item \code{CHROM}: the chromosome
\item \code{POS}: the position on the chromosome
\item \code{ID}: the SNP ID
\item \code{REF}: the reference variant
\item \code{ALT}: the alternate variant
}

The uppercase column names are those as used by PLINK2.}
}
\value{
Nothing. Will \link{stop} if the \code{pvar_table} is invalid
}
\description{
Check if a \code{.pvar} table is valid.
Will \link{stop} if not
}
\examples{
check_pvar_table(get_test_pvar_table())
}
\author{
Richèl J.C. Bilderbeek
}
