% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qassoc_table.R
\name{check_qassoc_table}
\alias{check_qassoc_table}
\title{Check if a \code{qassoc_table} is valid.}
\usage{
check_qassoc_table(qassoc_table)
}
\arguments{
\item{qassoc_table}{the table that hold the result of an association
with a quantitative trait, as created by \code{PLINK}/\code{PLINK2}
and stored as a \code{.qassoc} file (for \code{PLINK}) or
\verb{[basename].<phenotype>.glm.linear} (for \code{PLINK2}).}
}
\value{
Nothing. Will \link{stop} if the \code{qassoc_table} is invalid
}
\description{
Check if a \code{qassoc_table} is valid.
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
