% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_random_snp_selector.R
\name{check_random_snp_selector}
\alias{check_random_snp_selector}
\title{Check if a random-SNP selector is valid.}
\usage{
check_random_snp_selector(random_snp_selector)
}
\arguments{
\item{random_snp_selector}{a SNP selector (see \link{create_snp_selector})
that allows one to select one or more random SNPs,
as created by \link{create_random_snp_selector}
and checked by \link{check_random_snp_selector}}
}
\value{
Nothing.
}
\description{
Check if a random-SNP selector is valid.
Will \link{stop} if not.
}
\examples{
check_random_snp_selector(create_test_random_snp_selector())
}
\seealso{
There are multiple SNP selector checking functions,
see \link{create_snp_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
