% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample_ids.R
\name{check_sample_ids}
\alias{check_sample_ids}
\title{Check if a \code{sample_ids} is valid.}
\usage{
check_sample_ids(sample_ids)
}
\arguments{
\item{sample_ids}{sample IDs, which is a \link[tibble]{tibble}
with two columns: the first column holds the family ID
(called \code{fid} or \code{FID} by PLINK/PLINK2),
the second column holds the within-family ID
(called \code{iid} or \code{IID} by PLINK/PLINK2),
as can be checked by \link{check_sample_ids}}
}
\value{
Nothing. Will \link{stop} if the \code{sample_ids} is invalid
}
\description{
Check if a \code{sample_ids} is valid.
Will \link{stop} if not
}
\examples{
# For PLINK, a .phe table contains the sample IDs
phe_table <- get_test_phe_table()
sample_ids <- phe_table[, 1:2]
check_sample_ids(sample_ids)

# For PLINK2, a .psam table contains the sample IDs
psam_table <- get_test_psam_table()
sample_ids <- psam_table[, 1:2]
check_sample_ids(sample_ids)
}
\author{
Richèl J.C. Bilderbeek
}
