% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample_ids_filename.R
\name{check_sample_ids_filename}
\alias{check_sample_ids_filename}
\title{Checks that the \code{sample_ids_filename} is valid}
\usage{
check_sample_ids_filename(sample_ids_filename)
}
\arguments{
\item{sample_ids_filename}{name of a file to store \code{sample_ids}
(see \link{check_sample_ids}) to}
}
\value{
nothing
}
\description{
Checks that the \code{sample_ids_filename} is valid.
Calls \link{stop} if not
}
\examples{
check_sample_ids_filename("sample_ids.txt")
check_sample_ids_filename("/folder/sample_ids.txt")
check_sample_ids_filename("~/sample_ids.txt")
}
\author{
Richèl J.C. Bilderbeek
}
