% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sample_selector.R
\name{check_sample_selector}
\alias{check_sample_selector}
\title{Check if sample selector is valid.}
\usage{
check_sample_selector(sample_selector)
}
\arguments{
\item{sample_selector}{a sample/individual selector,
a way to select one or more samples/individuals.
See \link{create_samples_selector} for all sample selectors.}
}
\value{
Nothing.
}
\description{
Check if sample selector is valid.
Will \link{stop} if not.
}
\examples{
check_sample_selector(create_test_single_sample_selector())
}
\seealso{
There are multiple sample selectors,
see \link{create_samples_selector} for an overview.

There are multiple sample selector checking functions:
\itemize{
\item A single sample selector: \link{check_single_sample_selector}
}
}
\author{
Richèl J.C. Bilderbeek
}
