% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sim_qt_params.R
\name{check_sim_qt_params}
\alias{check_sim_qt_params}
\title{Check if the input is a valid set of parameters to
simulate a quantitative trait, will stop otherwise}
\usage{
check_sim_qt_params(sim_qt_params)
}
\arguments{
\item{sim_qt_params}{the parameters for a quantitative
traits simulation, as can be created by
\link{create_sim_qt_params}}
}
\description{
Check if the input is a valid set of parameters to
simulate a quantitative trait, will stop otherwise
}
