% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_simulate_qt_params.R
\name{check_simulate_qt_params}
\alias{check_simulate_qt_params}
\title{Check if the input is a valid set of parameters to
simulate a quantitative trait, will stop otherwise}
\usage{
check_simulate_qt_params(simulate_qt_params)
}
\arguments{
\item{simulate_qt_params}{the parameters for a quantitative
traits simulation, as can be created by
\link{create_simulate_qt_params}}
}
\description{
Check if the input is a valid set of parameters to
simulate a quantitative trait, will stop otherwise
}
\note{
This function is named after the \code{--simulate-qt} PLINK flag.
}
\author{
Richèl J.C. Bilderbeek
}
