% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_sample_selector.R
\name{check_single_sample_selector}
\alias{check_single_sample_selector}
\title{Check if a single-sample selector is valid.}
\usage{
check_single_sample_selector(single_sample_selector)
}
\value{
Nothing.
}
\description{
Check if a single-sample selector is valid.
Will \link{stop} if not.
}
\examples{
check_single_sample_selector(create_test_single_sample_selector())
}
\seealso{
There are multiple sample selector checking functions,
see \link{create_sample_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
