% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_snp_selector.R
\name{check_snp_selector}
\alias{check_snp_selector}
\title{Check if SNP selector is valid.}
\usage{
check_snp_selector(snp_selector)
}
\value{
Nothing.
}
\description{
Check if SNP selector is valid.
Will \link{stop} if not.
}
\examples{
# A SNP window selector
check_snp_selector(create_test_snp_window_selector())

# Other types of SNP selectors ...
}
\author{
Richèl J.C. Bilderbeek
}
