% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_snp_window_selector.R
\name{check_snp_window_selector}
\alias{check_snp_window_selector}
\title{Check if SNP windows selector is valid.}
\usage{
check_snp_window_selector(snp_window_selector)
}
\arguments{
\item{snp_window_selector}{a SNP selector (see \link{create_snp_selector}),
that allows one to select a focal SNP and a range/window of SNPs around it,
as created by \link{create_snp_window_selector}
and checked by \link{check_snp_window_selector}}
}
\value{
Nothing.
}
\description{
Check if SNP windows selector is valid.
Will \link{stop} if not.
}
\examples{
check_snp_window_selector(create_test_snp_window_selector())
}
\seealso{
There are multiple SNP selector checking functions,
see \link{create_snp_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
