% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_snvs.R
\name{check_snvs}
\alias{check_snvs}
\title{Check the single nucleotide variations.}
\usage{
check_snvs(snvs)
}
\arguments{
\item{snvs}{a \link[tibble]{tibble} that contains the
two nucleotide calls for multiple individuals.
Each column is a haplotype, hence, for a diploid organism,
there are two columns.
Each individual is represented by a row.
Nucleotides are in uppercase.
Per SNV table, there can be only two different nucleotides,
as this is how a SNP works; a SNP from A to C is a different one (and
has a different SNP ID) than a SNP that encodes a mutation
for A to G.
Use \link{create_snvs} to create a \code{snvs}.}
}
\description{
Check if the single nucleotide variations are valid,
will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
