% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trait.R
\name{check_trait}
\alias{check_trait}
\title{Check if a trait is valid.}
\usage{
check_trait(trait)
}
\arguments{
\item{trait}{one trait with a clear genetic architecture and a known
minor allele frequency, as created by \link{create_trait}.
Use \link{is_one_trait} to detect if something is one trait}
}
\value{
the checked trait, now of class type \code{trait}
}
\description{
Check if a trait (with a clear genetic architecture and a known
minor allele frequency) is valid.
Will \link{stop} if not
}
\examples{
check_trait(create_additive_trait())
check_trait(create_random_trait())
}
\author{
Richèl J.C. Bilderbeek
}
