% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_window_kb.R
\name{check_window_kb}
\alias{check_window_kb}
\title{Check if the window size (in kilobases) is valid.
Check if the window size (in kilobases) is valid.
Will \link{stop} if not}
\usage{
check_window_kb(window_kb)
}
\value{
Nothing. Will \link{stop} if the window
size (in kilobases) is invalid
}
\description{
Check if the window size (in kilobases) is valid.
Check if the window size (in kilobases) is valid.
Will \link{stop} if not
}
\examples{
check_window_kb(0.001) # 1 base pair
check_window_kb(1) # 1000 base pairs
}
\author{
Richèl J.C. Bilderbeek
}
