% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_plink_folder.R
\name{clear_plink_folder}
\alias{clear_plink_folder}
\title{Deletes all files in the default \code{PLINK} folder}
\usage{
clear_plink_folder(plink_folder = get_plink_folder())
}
\arguments{
\item{plink_folder}{folder where PLINK is installed}
}
\value{
Nothing.
}
\description{
Deletes all files in the default \code{PLINK} folder,
as obtained by \link{get_plink_folder}
}
\author{
Richèl J.C. Bilderbeek
}
