% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_plinkr_cache.R
\name{clear_plinkr_cache}
\alias{clear_plinkr_cache}
\title{Clear the \link{plinkr} cache}
\usage{
clear_plinkr_cache()
}
\value{
Nothing.
}
\description{
Clear the \link{plinkr} cache.
}
\author{
Richèl J.C. Bilderbeek
}
