% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_assoc_qt_speed.R
\name{compare_assoc_qt_speed}
\alias{compare_assoc_qt_speed}
\title{Compare speed when using \link{assoc_qt} for one \code{assoc_qt_params}
for (one or) multiple versions of \code{PLINK}/\code{PLINK2}}
\usage{
compare_assoc_qt_speed(
  assoc_qt_params,
  plink_optionses = create_plink_optionses(),
  verbose = FALSE
)
}
\arguments{
\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}

\item{plink_optionses}{a list of one or more options to run PLINK,
as created by \link{create_plink_optionses}.
The reduplicated plural was used to express this is a list
of \code{plink_options}, instead of one set of \code{plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
A \link[tibble]{tibble}
}
\description{
Compare speed when using \link{assoc_qt} for one \code{assoc_qt_params}
for (one or) multiple versions of \code{PLINK}/\code{PLINK2}
}
\author{
Richèl J.C. Bilderbeek
}
