% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bed_table_to_pgen_table.R
\name{convert_bed_table_to_pgen_table}
\alias{convert_bed_table_to_pgen_table}
\title{Convert a \code{PLINK} \code{.bed} table to a \code{PLINK2} \code{.pgen} table}
\usage{
convert_bed_table_to_pgen_table(bed_table, verbose = FALSE)
}
\arguments{
\item{bed_table}{a table that maps the SNPs to the individuals,
of which the column names are the names of the individuals,
the row names are the names of the SNPs,
and the values are the SNP variant.
Use \link{get_test_bed_table} to get a \code{.bed} table as used in testing.
Use \link{read_plink_bed_file} to read a \code{PLINK} \code{.bed} file.
Use \link{check_bed_table} to test if a \code{.bed} table is valid.
See also the \code{bed} file format reference at
\url{https://www.cog-genomics.org/plink2/formats#bed}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{PLINK2} \code{.pgen} table
}
\description{
Convert a \code{PLINK} \code{.bed} table to a \code{PLINK2} \code{.pgen} table
}
\author{
Richèl J.C. Bilderbeek
}
