% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bim_table_to_map_table.R
\name{convert_bim_table_to_map_table}
\alias{convert_bim_table_to_map_table}
\title{Convert a \code{PLINK} \code{.bim} table to a \code{PLINK} \code{.map} table}
\usage{
convert_bim_table_to_map_table(bim_table, verbose = FALSE)
}
\arguments{
\item{bim_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{chr}: the chromosome number
\item \code{id}: the SNP ID
\item \code{posg}: the position
\item \code{pos}: the position
\item \code{ref}: something
\item \code{alt}: something
}

Use \link{get_test_bim_table} to get a \code{.bim} table as used in testing.
Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.
Use \link{check_bim_table} to test if a \code{.bim} table is valid.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{PLINK} \code{.map} table
}
\description{
Convert a \code{PLINK} \code{.bim} table to a \code{PLINK} \code{.map} table
}
\author{
Richèl J.C. Bilderbeek
}
