% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pgen_table_to_bed_table.R
\name{convert_pgen_table_to_bed_table}
\alias{convert_pgen_table_to_bed_table}
\title{Convert a \code{PLINK2} \code{.pgen} table to a \code{PLINK} \code{.bed} table}
\usage{
convert_pgen_table_to_bed_table(pgen_table, verbose = FALSE)
}
\arguments{
\item{pgen_table}{a \link{matrix} that maps the individuals
to their SNPs. TODO}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{PLINK} \code{.bed} table
}
\description{
Convert a \code{PLINK2} \code{.pgen} table to a \code{PLINK} \code{.bed} table
}
\author{
Richèl J.C. Bilderbeek
}
