% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_plink2_bin_files_to_plink_text_files.R
\name{convert_plink2_bin_files_to_plink_text_files}
\alias{convert_plink2_bin_files_to_plink_text_files}
\title{Convert \code{PLINK2} binary data files to \code{PLINK} text data files}
\usage{
convert_plink2_bin_files_to_plink_text_files(
  base_input_filename,
  base_output_filename = file.path(dirname(base_input_filename), "output"),
  plink_options = create_plink_v2_0_options(),
  verbose = FALSE
)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a list with the following elements:
\itemize{
\item \code{ped_filename}: the full \code{.ped} filename
\item \code{map_filename}: the full \code{.map} filename
}
}
\description{
Convert \code{PLINK2} binary data files to \code{PLINK} text data files
}
\seealso{
these are the functions to convert between the \code{PLINK} and
\code{PLINK2} files:
\itemize{
\item To convert from PLINK1 text files
\itemize{
\item to PLINK1 binary files:
use \link{convert_plink_text_files_to_plink_bin_files}
\item to PLINK2 binary files,
use \link{convert_plink_text_files_to_plink2_bin_files}
}
\item To convert from PLINK1 binary files
\itemize{
\item to PLINK text files,
use \link{convert_plink_bin_files_to_plink_text_files}
\item to PLINK2 binary files,
use \link{convert_plink_bin_files_to_plink2_bin_files}
}
\item To convert from PLINK2 binary files
\itemize{
\item to PLINK text files,
use \link{convert_plink2_bin_files_to_plink_text_files}
\item to PLINK binary files,
use \link{convert_plink2_bin_files_to_plink_bin_files}
}
}
}
\author{
Richèl J.C. Bilderbeek
}
