% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_plink_bin_files_to_plink_vcf_files.R
\name{convert_plink_bin_files_to_plink_vcf_files}
\alias{convert_plink_bin_files_to_plink_vcf_files}
\title{Convert PLINK binary files to PLINK VCF files}
\usage{
convert_plink_bin_files_to_plink_vcf_files(
  base_input_filename,
  base_output_filename,
  plink_options = create_plink_options(),
  verbose = FALSE
)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
the full paths to the VCF files created
}
\description{
Convert PLINK binary files to PLINK VCF files
}
\author{
Richèl J.C. Bilderbeek
}
