% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/convert_plink_binary_files_to_plink_text_files.R
\name{convert_plink_binary_files_to_plink_text_files}
\alias{convert_plink_binary_files_to_plink_text_files}
\title{Convert PLINK binary files to PLINK plain-text files}
\usage{
convert_plink_binary_files_to_plink_text_files()
}
\description{
Convert PLINK binary files to PLINK plain-text files
}
