% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_plink_text_data_to_plink_bin_data.R
\name{convert_plink_text_data_to_plink_bin_data}
\alias{convert_plink_text_data_to_plink_bin_data}
\title{Convert the \code{PLINK} text data to \code{PLINK} binary data}
\usage{
convert_plink_text_data_to_plink_bin_data(
  plink_text_data,
  plink_options = create_plink_v1_9_options(),
  verbose = FALSE
)
}
\arguments{
\item{plink_text_data}{the genetic and phenotypic data for \code{PLINK} to work
on, in \code{PLINK} text format, as created by \link{create_plink_text_data}.}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
the same \link{list} that was used as input,
except that the following elements
\itemize{
\item \code{ped_table} (see \link{check_ped_table})
\item \code{map_table} (see \link{check_map_table})
}

are replaced by:
\itemize{
\item \code{bim_table} (see \link{check_bim_table})
\item \code{fam_table} (see \link{check_fam_table})
\item \code{bed_table} (see \link{check_bed_table})
}

The other list elements are left as-is.
}
\description{
Convert the \code{PLINK} text data to \code{PLINK} binary data.
}
\examples{
if (is_on_ci()) {
  if (is_plink_installed(plink_options = create_plink_v1_9_options())) {

    # Use testing data
    convert_plink_text_data_to_plink_bin_data(
      plink_text_data = create_test_plink_text_data()
    )

    # Use simulated data
    assoc_qt_data <- create_demo_assoc_qt_data()
    # Is PLINK text data
    is_plink_text_data(assoc_qt_data$data)
    assoc_qt_data$data <- convert_plink_text_data_to_plink_bin_data(
      assoc_qt_data$data
    )
    is_plink_bin_data(assoc_qt_data$data)
  }
}
}
\seealso{
these are the functions to convert between the \code{PLINK} and
\code{PLINK2} formats:
\itemize{
\item To convert from PLINK1 text data
\itemize{
\item to PLINK1 binary data:
use \link{convert_plink_text_data_to_plink_bin_data}
\item to PLINK2 binary data,
use \link{convert_plink_text_data_to_plink2_bin_data}
}
\item To convert from PLINK1 binary data
\itemize{
\item to PLINK text data,
use \link{convert_plink_bin_data_to_plink_text_data}
\item to PLINK2 binary data,
use \link{convert_plink_bin_data_to_plink2_bin_data}
}
\item To convert from PLINK2 binary data
\itemize{
\item to PLINK text data,
use \link{convert_plink2_bin_data_to_plink_text_data}
\item to PLINK binary data,
use \link{convert_plink2_bin_data_to_plink_bin_data}
}
}
}
\author{
Richèl J.C. Bilderbeek
}
