% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_plink_text_files_to_plink2_bin_files.R
\name{convert_plink_text_files_to_plink2_bin_files}
\alias{convert_plink_text_files_to_plink2_bin_files}
\title{Convert \code{PLINK} text data files to \code{PLINK2} binary data files}
\usage{
convert_plink_text_files_to_plink2_bin_files(
  base_input_filename,
  base_output_plink1_filename = file.path(dirname(base_input_filename),
    "output_plink1"),
  base_output_plink2_filename = file.path(dirname(base_input_filename),
    "output_plink2"),
  plink1_options = create_plink_v1_9_options(),
  plink2_options = create_plink_v2_0_options(),
  verbose = FALSE
)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_plink1_filename}{temporary folder to create the \code{PLINK}
binary files in}

\item{base_output_plink2_filename}{temporary folder to create the \code{PLINK2}
binary files in}

\item{plink1_options}{the \code{PLINK} version,
as created by \link{create_plink_options},
for any version of \code{PLINK}}

\item{plink2_options}{the \code{PLINK2} version,
as created by \link{create_plink_options},
for any version of \code{PLINK2}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a list with the following elements:
\itemize{
\item \code{pgen_filename}: the full \code{.pgen} filename
\item \code{psam_filename}: the full \code{.psam} filename
\item \code{pvar_filename}: the full \code{.pvar} filename
\item \code{log_filename}: the full \code{.log} filename
}
}
\description{
Convert \code{PLINK} text data files to \code{PLINK2} binary data files, by
}
\details{
\enumerate{
\item convert \code{PLINK} text data files to \code{PLINK} binary data files,
using \link{convert_plink_text_files_to_plink_bin_files},
with \code{base_output_plink1_filename} as a temporary folder
and \code{plink1_options} as the PLINK version
\item convert \code{PLINK} binary data files to \code{PLINK2} binary data files,
using \link{convert_plink_bin_files_to_plink2_bin_files}
with \code{base_output_plink2_filename} as a temporary folder
and \code{plink2_options} as the PLINK2 version
}
}
\seealso{
these are the functions to convert between the \code{PLINK} and
\code{PLINK2} files:
\itemize{
\item To convert from PLINK1 text files
\itemize{
\item to PLINK1 binary files:
use \link{convert_plink_text_files_to_plink_bin_files}
\item to PLINK2 binary files,
use \link{convert_plink_text_files_to_plink2_bin_files}
}
\item To convert from PLINK1 binary files
\itemize{
\item to PLINK text files,
use \link{convert_plink_bin_files_to_plink_text_files}
\item to PLINK2 binary files,
use \link{convert_plink_bin_files_to_plink2_bin_files}
}
\item To convert from PLINK2 binary files
\itemize{
\item to PLINK text files,
use \link{convert_plink2_bin_files_to_plink_text_files}
\item to PLINK binary files,
use \link{convert_plink2_bin_files_to_plink_bin_files}
}
}
}
\author{
Richèl J.C. Bilderbeek
}
