% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_psam_table_to_fam_table.R
\name{convert_psam_table_to_fam_table}
\alias{convert_psam_table_to_fam_table}
\title{Convert a \code{PLINK2} \code{.psam} table to a \code{PLINK} \code{.fam} table}
\usage{
convert_psam_table_to_fam_table(psam_table, verbose = FALSE)
}
\arguments{
\item{psam_table}{a \link[tibble]{tibble} with as much rows as
individuals. It has the following columns:
\itemize{
\item \code{FID}: the family ID
\item \code{IID}: the within-family ID
\item \code{SEX}: the sex
\item \code{PHENO1}: a phenotypic value
}

The uppercase column names are those as used by PLINK2.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{PLINK} \code{.fam} table
}
\description{
Convert a \code{PLINK2} \code{.psam} table to a \code{PLINK} \code{.fam} table
}
\author{
Richèl J.C. Bilderbeek
}
