% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pvar_table_to_bim_table.R
\name{convert_pvar_table_to_bim_table}
\alias{convert_pvar_table_to_bim_table}
\title{Convert a \code{PLINK2} \code{.pvar} table to a \code{PLINK} \code{.bim} table}
\usage{
convert_pvar_table_to_bim_table(pvar_table, verbose = FALSE)
}
\arguments{
\item{pvar_table}{a \link[tibble]{tibble} with as much rows as
SNPs. It has the following columns:
\itemize{
\item \code{CHROM}: the chromosome
\item \code{POS}: the position on the chromosome
\item \code{ID}: the SNP ID
\item \code{REF}: the reference variant
\item \code{ALT}: the alternate variant
}

The uppercase column names are those as used by PLINK2.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{PLINK} \code{.bim} table
}
\description{
Convert a \code{PLINK2} \code{.pvar} table to a \code{PLINK} \code{.bim} table
}
\author{
Richèl J.C. Bilderbeek
}
