% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_args.R
\name{create_assoc_args}
\alias{create_assoc_args}
\title{Create the command-line arguments to call PLINK or PLINK2}
\usage{
create_assoc_args(
  assoc_data,
  assoc_params,
  plink_options = create_plink_options()
)
}
\arguments{
\item{assoc_data}{data to use for an association analysis
for a quantitative trait (i.e. using \link{assoc}),
as can be created by \link{create_assoc_data}}

\item{assoc_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc}),
as can be created by \link{create_assoc_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to call PLINK or PLINK2
}
\author{
Richèl J.C. Bilderbeek
}
