% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_data.R
\name{create_assoc_data}
\alias{create_assoc_data}
\title{Create data for the \link{assoc} function}
\usage{
create_assoc_data(data)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
\item data to be used by \link{assoc},
as created by \link{create_assoc_data}
\item data to be used by \link{assoc_qt},
as created by \link{create_assoc_qt_data}
}}
}
\description{
Create data to detect associations with a binary traits
(using \link{assoc}).
Use \link{create_test_assoc_data}) to create a dataset
that is used in testing.
}
\note{
This function is named after the \code{--assoc} PLINK flag.
}
\author{
Richèl J.C. Bilderbeek
}
