% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_params.R
\name{create_assoc_params}
\alias{create_assoc_params}
\title{Create parameters for the \link{assoc} function}
\usage{
create_assoc_params(
  confidence_interval = 0.95,
  maf = get_lowest_maf(),
  base_input_filename = file.path(get_plinkr_tempfilename(), "assoc_input"),
  base_output_filename = file.path(get_plinkr_tempfilename(), "assoc_output")
)
}
\arguments{
\item{confidence_interval}{confidence interval, a value between (and
excluding both) 0.0 and 1.0. The confidence interval helps assess
the certainty of an estimation: you can be 99 percent sure
a value is within the range of the 0.99 confidence interval

The \code{confidence_interval} is only used by \code{PLINK2}.
For \code{PLINK}, its value is ignored}

\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the \code{PLINK} analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
\code{PLINK} \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}
}
\description{
Create parameters to detect associations with a binary traits
(using \link{assoc}).
Use \link{create_test_assoc_params}) to create a set of parameters
that is used in testing.
}
\note{
This function is named after the \code{--assoc} PLINK flag.
}
\author{
Richèl J.C. Bilderbeek
}
