% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_qt_args.R
\name{create_assoc_qt_args}
\alias{create_assoc_qt_args}
\title{Create the command-line arguments to call \code{PLINK} or \code{PLINK2}
to do a quantitative trait analysis,
depending on the type of data.}
\usage{
create_assoc_qt_args(
  assoc_qt_data,
  assoc_qt_params,
  plink_options = create_plink_options()
)
}
\arguments{
\item{assoc_qt_data}{data use to used by \link{assoc_qt}
(as created by \link{create_assoc_qt_data}),
which equals 'regular' data (see \link{check_data}) and a
phenotype table/file}

\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to call \code{PLINK} or \code{PLINK2}
to do a quantitative trait analysis,
depending on the type of data.
}
\examples{
create_assoc_qt_args(
  assoc_qt_data = create_test_assoc_qt_data(),
  assoc_qt_params = create_test_assoc_qt_params(),
  plink_options = create_plink_options()
)
}
\seealso{
These are the specialized functions:
\itemize{
\item \link{create_assoc_qt_args_on_plink_text_data}
\item \link{create_assoc_qt_args_on_plink_bin_data}
\item \link{create_assoc_qt_args_on_plink2_bin_data}
\item \link{create_assoc_qt_args_on_plink_text_filenames}
\item \link{create_assoc_qt_args_on_plink_bin_filenames}
\item \link{create_assoc_qt_args_on_plink2_bin_filenames}
}
}
\author{
Richèl J.C. Bilderbeek
}
