% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_qt_args_on_plink_bin_filenames.R
\name{create_assoc_qt_args_on_plink_bin_filenames}
\alias{create_assoc_qt_args_on_plink_bin_filenames}
\title{Internal function}
\usage{
create_assoc_qt_args_on_plink_bin_filenames(
  assoc_qt_data,
  assoc_qt_params,
  plink_options = create_plink_options()
)
}
\arguments{
\item{assoc_qt_data}{data use to used by \link{assoc_qt}
(as created by \link{create_assoc_qt_data}),
which equals 'regular' data (see \link{check_data}) and a
phenotype table/file}

\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to call \code{PLINK} or \code{PLINK2}
to do a quantitative trait analysis, when
\code{PLINK} binary data is used
}
\examples{
assoc_qt_data <- create_assoc_qt_data(
  data = create_plink_bin_filenames(
    bed_filename = get_plinkr_filename("select_snps.bed"),
    bim_filename = get_plinkr_filename("select_snps.bim"),
    fam_filename = get_plinkr_filename("select_snps.fam")
  ),
  phenotype_data = create_test_phenotype_data_filename(
    phe_filename = get_plinkr_filename("select_snps.phe")
  )
)
create_assoc_qt_args_on_plink_bin_filenames(
  assoc_qt_data = assoc_qt_data,
  assoc_qt_params = create_test_assoc_qt_params()
)
}
\seealso{
see \link{create_assoc_qt_args} for all the specialized
functions to create the \code{PLINK}/\code{PLINK2} command-line arguments
}
\author{
Richèl J.C. Bilderbeek
}
