% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_qt_covar_args.R
\name{create_assoc_qt_covar_args}
\alias{create_assoc_qt_covar_args}
\title{Create the command-line arguments to call \code{PLINK} or \code{PLINK2}
to do a quantitative trait analysis with covariates}
\usage{
create_assoc_qt_covar_args(
  assoc_qt_covar_data,
  assoc_qt_covar_params,
  plink_options = create_plink_options()
)
}
\arguments{
\item{assoc_qt_covar_data}{data to do an association analysis
for a quantitative trait with covariates
(i.e. using \link{assoc_qt_covar}),
as can be created by \link{create_assoc_qt_covar_data}}

\item{assoc_qt_covar_params}{parameters to do an association analysis
for a quantitative trait with covariates
(i.e. using \link{assoc_qt_covar}),
as can be created by \link{create_assoc_qt_covar_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to call \code{PLINK} or \code{PLINK2}
to do a quantitative trait analysis with covariates
}
\author{
Richèl J.C. Bilderbeek
}
