% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_qt_covar_data.R
\name{create_assoc_qt_covar_data}
\alias{create_assoc_qt_covar_data}
\title{Create data for the \link{assoc_qt_covar} function}
\usage{
create_assoc_qt_covar_data(data, phenotype_data, cov_table)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
}}

\item{phenotype_data}{phenotype data in the form of either
a filename to a phenotype file (\code{phenotype_data_filename})
or a list with a phenotype table (\code{phenotype_data_table})}

\item{cov_table}{a table of covariates.

A \code{cov_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of covariate values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_cov_file} to read a covariates file.
Use \link{check_cov_table} to test if a covariates table is valid.}
}
\description{
Create data to detect associations with quantitative traits
and covariates
(using \link{assoc_qt_covar}).
Use \link{create_test_assoc_qt_covar_data})
to create a dataset
that is used in testing.
}
\author{
Richèl J.C. Bilderbeek
}
