% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_qt_covar_params.R
\name{create_assoc_qt_covar_params}
\alias{create_assoc_qt_covar_params}
\title{Create parameters for the \link{assoc_qt_covar} function}
\usage{
create_assoc_qt_covar_params(
  maf = get_lowest_maf(),
  confidence_interval = 0.95,
  allow_no_sex = FALSE,
  base_input_filename = file.path(get_plinkr_tempfilename(), "assoc_qt_covar_input"),
  base_output_filename = file.path(get_plinkr_tempfilename(), "assoc_qt_covar_input")
)
}
\arguments{
\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the \code{PLINK} analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
\code{PLINK} \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}

\item{confidence_interval}{confidence interval, a value between (and
excluding both) 0.0 and 1.0. The confidence interval helps assess
the certainty of an estimation: you can be 99 percent sure
a value is within the range of the 0.99 confidence interval}

\item{allow_no_sex}{set this to \link{TRUE}
to retain phenotype values for missing-sex samples.
This parameter is named after the \code{PLINK} \code{--allow-no-sex} flag}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}
}
\description{
Create parameters to detect associations with quantitative traits
and covariates  (using \link{assoc_qt_covar}).
Use \link{create_test_assoc_qt_covar_params}) to create a set of parameters
that is used in testing.
}
\note{
This function is named after the \code{--assoc}
and \code{--covar} \code{PLINK} flags, as well as forcing \code{PLINK}
to use quantitative trait phenotypic values.
}
\examples{
check_assoc_qt_covar_params(create_assoc_qt_covar_params())
check_assoc_qt_covar_params(create_test_assoc_qt_covar_params())
}
\author{
Richèl J.C. Bilderbeek
}
