% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_qt_params.R
\name{create_assoc_qt_params}
\alias{create_assoc_qt_params}
\title{Create parameters for the \link{assoc_qt} function}
\usage{
create_assoc_qt_params(
  maf = get_lowest_maf(),
  allow_no_sex = FALSE,
  base_input_filename = file.path(get_plinkr_tempfilename(), "assoc_input"),
  base_output_filename = file.path(get_plinkr_tempfilename(), "assoc_output")
)
}
\arguments{
\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the \code{PLINK} analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
\code{PLINK} \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}

\item{allow_no_sex}{set this to \link{TRUE}
to retain phenotype values for missing-sex samples.
This parameter is named after the \code{PLINK} \code{--allow-no-sex} flag}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}
}
\description{
Create parameters to detect associations with quantitative traits
(using \link{assoc_qt}).
Use \link{create_test_assoc_qt_params}) to create a set of parameters
that is used in testing.
}
\note{
This function is named after the \code{--assoc} \code{PLINK} flag.
Where \code{--PLINK} uses phenotypic values to determine if an association
is about a binary/case-control trait or a quantitative trait,
\link{plinkr} uses functions named \link{assoc} and \link{assoc_qt}.
}
\author{
Richèl J.C. Bilderbeek
}
