% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bgen_files_for_saige.R
\name{create_bgen_files_for_saige}
\alias{create_bgen_files_for_saige}
\title{Create the \code{.bgen} files for \code{SAIGE}}
\usage{
create_bgen_files_for_saige()
}
\description{
Create the \code{.bgen} files for \code{SAIGE}
}
\note{
thanks to Marcin Kierczak  (\url{https://github.com/mkierczak/})
for sharing his wisdom
}
\author{
Richèl J.C. Bilderbeek
}
