% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chromosome_selector.R
\name{create_chromosome_selector}
\alias{create_chromosome_selector}
\title{Create a selector to select all SNPs on a chromosome}
\usage{
create_chromosome_selector(chromosome_number)
}
\description{
Create a selector to select all SNPs on a chromosome
to be used in, for example, \link{select_snps}
}
\examples{
create_chromosome_selector(chromosome_number = 1)
create_chromosome_selector(chromosome_number = 22)
}
\seealso{
Use \link{create_test_chromosome_selector} to
create a single-SNP selector to be used in testing.

There are multiple SNP selectors,
see \link{create_snp_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
