% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_convert_plink_bin_files_to_plink_text_files_args.R
\name{create_convert_plink_bin_files_to_plink_text_files_args}
\alias{create_convert_plink_bin_files_to_plink_text_files_args}
\title{Create the command-line arguments to convert PLINK binary
files to PLINK text files}
\usage{
create_convert_plink_bin_files_to_plink_text_files_args(
  base_input_filename,
  base_output_filename,
  plink_options = create_plink_v1_9_options()
)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to convert PLINK binary
files to PLINK text files
}
\note{
the first argument is the first argument, not the path to
the \code{PLINK} or \code{PLINK2} binary
}
\examples{
create_convert_plink_bin_files_to_plink_text_files_args(
  base_input_filename = "base_input_filename",
  base_output_filename = "base_output_filename"
)
}
\author{
Richèl J.C. Bilderbeek
}
