% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cov_table_from_data.R
\name{create_cov_table_from_data}
\alias{create_cov_table_from_data}
\title{Create a single-covariate table from pedigree (\code{.ped}) table}
\usage{
create_cov_table_from_data(data)
}
\arguments{
\item{data}{PLINK or PLINK2 text or binary data:
\itemize{
\item PLINK text data:
a \code{.ped} table (see \link{read_plink_ped_file})
and a \code{.map} (see \link{read_plink_map_file}),
as created by \link{create_plink_text_data}.
\item PLINK binary data:
a \code{.bed} table (see \link{read_plink_bed_file})
a \code{.bim} table (see \link{read_plink_bim_file})
and a \code{.fam} (see \link{read_plink_fam_file}),
as created by \link{create_plink_bin_data}.
\item PLINK2 binary data: use \link{create_plink2_bin_data}
}}
}
\value{
a single-covariate \link[tibble]{tibble}, with columns:
\itemize{
\item \code{FID} the family ID
\item \code{IID} the within-family ID
\item \code{C1} the covariate, which is zero by default
}

The column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#covno}
}
\description{
Create a single-covariate table from pedigree (\code{.ped}) table.
The individuals' family and within-family ID's are copied.
}
\author{
Richèl J.C. Bilderbeek
}
