% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_custom_plink_options.R
\name{create_custom_plink_options}
\alias{create_custom_plink_options}
\title{Create the options to run a custom PLINK version}
\usage{
create_custom_plink_options(plink_folder)
}
\arguments{
\item{plink_folder}{folder where PLINK is installed}
}
\value{
the PLINK options,
as can be checked by \link{check_plink_options}
}
\description{
Create the options to run a custom PLINK version
}
\examples{
create_custom_plink_options("path_to_my_custom_plink_folder")
}
\author{
Richèl J.C. Bilderbeek
}
