% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_custom_trait.R
\name{create_custom_trait}
\alias{create_custom_trait}
\title{Create a custom trait in which the phenotype is
determined in a custom way
and the minor allele frequency is known.}
\usage{
create_custom_trait(
  mafs = 0.25,
  n_snps = 1,
  calc_phenotype_function = calc_random_phenotype_values
)
}
\arguments{
\item{mafs}{one or more minor allele frequencies.
These allele frequencies must be ordered decreasingly,
i.e. the MAF is at the first position, where the
even rarer alleles are at the second and third positions.

Note that \code{PLINK} cannot handle triallelic nor
quadallelic SNPs: \code{PLINK} will give a warning that it
is setting the rarest alleles to missing.}

\item{n_snps}{the number of SNPs}
}
\description{
Create a custom trait in which the phenotype is
determined in a custom way
and the minor allele frequency is known.
}
\examples{
create_additive_trait()
create_additive_trait(maf = 0.01)
}
\author{
Richèl J.C. Bilderbeek
}
