% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_assoc_params.R
\name{create_demo_assoc_params}
\alias{create_demo_assoc_params}
\title{Create a set of \code{assoc} parameters
to demonstrate PLINKs \code{--assoc} functionality}
\usage{
create_demo_assoc_params(
  n_individuals = 3,
  trait = create_random_case_control_trait(),
  confidence_interval = 0.95
)
}
\arguments{
\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}

\item{trait}{one trait with a clear genetic architecture and a known
minor allele frequency, as created by \link{create_trait}.
Use \link{is_one_trait} to detect if something is one trait}

\item{confidence_interval}{confidence interval, a value between (and
excluding both) 0.0 and 1.0. The confidence interval helps assess
the certainty of an estimation: you can be 99 percent sure
a value is within the range of the 0.99 confidence interval}
}
\description{
This function creates:
\itemize{
\item Individuals that have all combinations of SNPs exactly once,
as created by the \link{create_demo_ped_table} function
\item All SNPs that are on different chromosomes
as created by the \link{create_demo_map_table} function
\item Traits that follow different genotype to phenotype mapping:
\itemize{
\item \code{control}: a trait that is random
\item \code{additive}: a trait that is perfectly additive,
see \link{calc_additive_phenotype_values} for the exact calculation
\item \code{epistatic}: a trait that is epistatic,
see \link{calc_epistatic_phenotype_values} for the exact calculation
}
}
}
\note{
This function is named after the \code{--assoc} PLINK flag.
}
\examples{
# Default
create_demo_assoc_params()

# Add more individuals
create_demo_assoc_params(n_individuals = 16)

# Use a random trait
create_demo_assoc_params(
  trait = create_random_trait()
)

# Use an additive trait
create_demo_assoc_params(
  trait = create_additive_trait()
)

# Use an epistatic trait
create_demo_assoc_params(
  trait = create_epistatic_trait()
)
}
\author{
Richèl J.C. Bilderbeek
}
