% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_assoc_qt_covar_params.R
\name{create_demo_assoc_qt_covar_params}
\alias{create_demo_assoc_qt_covar_params}
\title{Create a set of \code{assoc_qt_covar} parameters
to demonstrate PLINKs \code{--assoc} and \code{--covar} functionality}
\usage{
create_demo_assoc_qt_covar_params(
  n_individuals = 3,
  traits = create_demo_traits()
)
}
\arguments{
\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}

\item{traits}{one or more traits
with a clear genetic architecture and a known minor allele frequency,
as, for example, created by \link{create_demo_traits}.}
}
\description{
This function creates:
\itemize{
\item Individuals that have all combinations of SNPs exactly once,
as created by the \link{create_demo_ped_table} function
\item All SNPs that are on different chromosomes
as created by the \link{create_demo_map_table} function
\item Traits that follow different genotype to phenotype mapping:
\itemize{
\item \code{control}: a trait that is random
\item \code{additive}: a trait that is perfectly additive,
see \link{calc_additive_phenotype_values} for the exact calculation
\item \code{epistatic}: a trait that is epistatic,
see \link{calc_epistatic_phenotype_values} for the exact calculation
}
}
}
\note{
This function is named after the \code{--assoc}
and \code{--covar} PLINK flags.
}
\examples{
# Default
create_demo_assoc_qt_covar_params()
}
\author{
Richèl J.C. Bilderbeek
}
