% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_map_table.R
\name{create_demo_map_table}
\alias{create_demo_map_table}
\title{Create a demo genetic mapping \code{.map} table}
\usage{
create_demo_map_table(n_snps = 1)
}
\arguments{
\item{n_snps}{the number of SNPs}
}
\description{
Create a demo genetic mapping \code{.map} table.
It is simply a map of SNPs that are located on different chromosomes.
}
\examples{
create_demo_map_table()
}
\author{
Richèl J.C. Bilderbeek
}
