% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_ped_table.R
\name{create_demo_ped_table}
\alias{create_demo_ped_table}
\title{Create a demo pedigree \code{.ped} table}
\usage{
create_demo_ped_table(n_individuals = 4, traits = create_demo_traits())
}
\arguments{
\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}

\item{traits}{one or more traits
with a clear genetic architecture and a known minor allele frequency,
as, for example, created by \link{create_demo_traits}.}
}
\description{
Create a demo pedigree \code{.ped} table,
which should allow the individuals to have all combinations of
biallelic SNPs exactly once.
As we assume two nucleotides and a diploid organism,
these are 4 individuals.
}
\examples{
create_demo_ped_table()
}
\author{
Richèl J.C. Bilderbeek
}
