% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_ped_table_lhs.R
\name{create_demo_ped_table_lhs}
\alias{create_demo_ped_table_lhs}
\title{Internal function}
\usage{
create_demo_ped_table_lhs(n_individuals = 4)
}
\arguments{
\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}
}
\description{
Internal function to create the left-hand side
of a demo pedigree \code{.ped} table
}
\examples{
create_demo_ped_table_lhs()
create_demo_ped_table_lhs(n_individuals = 4)
}
\author{
Richèl J.C. Bilderbeek
}
