% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_ped_table_rhs.R
\name{create_demo_ped_table_rhs}
\alias{create_demo_ped_table_rhs}
\title{Internal function}
\usage{
create_demo_ped_table_rhs(traits = create_demo_traits(), n_individuals = 4)
}
\arguments{
\item{traits}{one or more traits
with a clear genetic architecture and a known minor allele frequency,
as, for example, created by \link{create_demo_traits}.}

\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}
}
\description{
Internal function to create the right-hand side
of a demo pedigree \code{.ped} table
for all traits
}
\examples{
create_demo_ped_table_rhs()
}
\author{
Richèl J.C. Bilderbeek
}
