% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_ped_table_rhs_trait.R
\name{create_demo_ped_table_rhs_trait}
\alias{create_demo_ped_table_rhs_trait}
\title{Internal function}
\usage{
create_demo_ped_table_rhs_trait(
  trait = create_random_trait(),
  n_individuals = 4
)
}
\arguments{
\item{trait}{one trait with a clear genetic architecture and a known
minor allele frequency, as created by \link{create_trait}.
Use \link{is_one_trait} to detect if something is one trait}

\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}
}
\description{
Internal function to create the right-hand side
of a demo pedigree \code{.ped} table
for one trait
}
\examples{
create_demo_ped_table_rhs()
}
\author{
Richèl J.C. Bilderbeek
}
