% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_traits.R
\name{create_demo_traits}
\alias{create_demo_traits}
\title{Create some traits with a clear genetic architecture and a known
minor allele frequency for demonstration purposes.}
\usage{
create_demo_traits()
}
\description{
Create some traits with a clear genetic architecture and a known
minor allele frequency for demonstration purposes.
}
\examples{
traits <- create_demo_traits()
check_traits(traits)
}
\author{
Richèl J.C. Bilderbeek
}
