% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pca_args.R
\name{create_pca_args}
\alias{create_pca_args}
\title{Create the CLI arguments to call \code{PLINK} with the \code{--pca} flag}
\usage{
create_pca_args(bfile, base_output_filename)
}
\arguments{
\item{bfile}{the base filename of the binary files (i.e.
a \code{.bed}, \code{.bim} and \code{.fam} file).
This parameter is named after the \code{PLINK}
\code{--bfile} flag}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}
}
\value{
the CLI arguments
}
\description{
Create the CLI arguments to call \code{PLINK} with the \code{--pca} flag
}
\author{
Richèl J.C. Bilderbeek
}
