% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phe_table_from_data.R
\name{create_phe_table_from_data}
\alias{create_phe_table_from_data}
\title{Create a single-phenotype table from the PLINK data.}
\usage{
create_phe_table_from_data(data)
}
\arguments{
\item{data}{PLINK or PLINK2 text or binary data:
\itemize{
\item PLINK text data:
a \code{.ped} table (see \link{read_plink_ped_file})
and a \code{.map} (see \link{read_plink_map_file}),
as created by \link{create_plink_text_data}.
\item PLINK binary data:
a \code{.bed} table (see \link{read_plink_bed_file})
a \code{.bim} table (see \link{read_plink_bim_file})
and a \code{.fam} (see \link{read_plink_fam_file}),
as created by \link{create_plink_bin_data}.
\item PLINK2 binary data: use \link{create_plink2_bin_data}
}}
}
\value{
a single-phenotype \link[tibble]{tibble}, with columns:
\itemize{
\item \code{FID} the family ID
\item \code{IID} the within-family ID
\item \code{P1} the phenotype
}

The column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}
}
\description{
Create a single-phenotype table from the PLINK data.
The individuals' family and within-family ID's are copied.
The \code{case_control_code} is copied as well, but
renamed to \code{trait} and divided by then.
In this way, PLINK traits the phenotype table as having quantitative
traits.
}
\examples{
create_phe_table_from_data(create_test_plink_text_data())
create_phe_table_from_data(create_test_plink_bin_data())
create_phe_table_from_data(create_test_plink2_bin_data())
}
\author{
Richèl J.C. Bilderbeek
}
