% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phe_table_from_fam_table.R
\name{create_phe_table_from_fam_table}
\alias{create_phe_table_from_fam_table}
\title{Create a single-phenotype table from pedigree (\code{.fam}) table}
\usage{
create_phe_table_from_fam_table(fam_table)
}
\arguments{
\item{fam_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{fam}: the family ID
\item \code{id}: the individual's ID
\item \code{pat}: ID of father
\item \code{mat}: ID of mother
\item \code{sex}: the gender
\item \code{pheno}: a phenotype
}

Use \link{get_test_fam_table} to get a \code{.fam} table as used in testing.
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.
Use \link{check_fam_table} to test if a \code{.fam} table is valid.}
}
\value{
a single-phenotype \link[tibble]{tibble}, with columns:
\itemize{
\item \code{FID} the family ID
\item \code{IID} the within-family ID
\item \code{P1} the phenotype
}

The column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}
}
\description{
Create a single-phenotype table from pedigree (\code{.fam}) table.
The individuals' family and within-family ID's are copied.
The \code{pheno} is copied as well, but
renamed to \code{P1} and divided by ten.
In this way, PLINK treats the phenotype table as having quantitative
traits.
}
\examples{
create_phe_table_from_fam_table(get_test_fam_table())
}
\author{
Richèl J.C. Bilderbeek
}
