% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phe_table_from_plink2_bin_data.R
\name{create_phe_table_from_plink2_bin_data}
\alias{create_phe_table_from_plink2_bin_data}
\title{Create a single-phenotype table from a table
within the PLINK2 binary data.}
\usage{
create_phe_table_from_plink2_bin_data(plink2_bin_data)
}
\arguments{
\item{plink2_bin_data}{the genetic and phenotypic data for \code{PLINK2} to work
on, in \code{PLINK2} binary format, as created by \link{create_plink2_bin_data}.}
}
\value{
a single-phenotype \link[tibble]{tibble}, with columns:
\itemize{
\item \code{FID} the family ID
\item \code{IID} the within-family ID
\item \code{P1} the phenotype
}

The column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}
}
\description{
Create a single-phenotype table from a table
within the PLINK2 binary data.
The individuals' family and within-family ID's are copied.
The \code{case_control_code} is copied as well, but
renamed to \code{trait} and divided by then.
In this way, PLINK traits the phenotype table as having quantitative
traits.
}
\examples{
create_phe_table_from_plink_bin_data(create_test_plink_bin_data())
}
\author{
Richèl J.C. Bilderbeek
}
