% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phe_table_from_plink_bin2_filenames.R
\name{create_phe_table_from_plink_bin2_filenames}
\alias{create_phe_table_from_plink_bin2_filenames}
\title{Create a single-phenotype table from a table
from the PLINK2 binary files.}
\usage{
create_phe_table_from_plink_bin2_filenames(plink2_bin_filenames)
}
\arguments{
\item{plink2_bin_filenames}{the binary data files' names for \code{PLINK2}
to work on,
as created by \link{create_plink2_bin_filenames}.}
}
\value{
a single-phenotype \link[tibble]{tibble}, with columns:
\itemize{
\item \code{FID} the family ID
\item \code{IID} the within-family ID
\item \code{P1} the phenotype
}

The column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}
}
\description{
Create a single-phenotype table from a table
from the PLINK2 binary files.
}
\details{
The individuals' family and within-family ID's are copied.
The \code{case_control_code} is copied as well, but
renamed to \code{trait} and divided by then.
In this way, PLINK treats the phenotype table as having quantitative
traits.
}
\author{
Richèl J.C. Bilderbeek
}
