% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phe_table_from_plink_text_filenames.R
\name{create_phe_table_from_plink_text_filenames}
\alias{create_phe_table_from_plink_text_filenames}
\title{Create a single-phenotype table from pedigree (\code{.ped}) table
within the PLINK text files.}
\usage{
create_phe_table_from_plink_text_filenames(plink_text_filenames)
}
\arguments{
\item{plink_text_filenames}{the text data files' names for \code{PLINK}
to work on,
as created by \link{create_plink_text_filenames}.}
}
\value{
a single-phenotype \link[tibble]{tibble}, with columns:
\itemize{
\item \code{FID} the family ID
\item \code{IID} the within-family ID
\item \code{P1} the phenotype
}

The column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}
}
\description{
Create a single-phenotype table from pedigree (\code{.ped}) table
within the PLINK text files.
The individuals' family and within-family ID's are copied.
The \code{case_control_code} is copied as well, but
renamed to \code{trait} and divided by then.
In this way, PLINK treats the phenotype table as having quantitative
traits.
}
\author{
Richèl J.C. Bilderbeek
}
