% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phe_table_from_psam_table.R
\name{create_phe_table_from_psam_table}
\alias{create_phe_table_from_psam_table}
\title{Create a single-phenotype table from pedigree (\code{.psam}) table}
\usage{
create_phe_table_from_psam_table(psam_table)
}
\arguments{
\item{psam_table}{a \link[tibble]{tibble} with as much rows as
individuals. It has the following columns:
\itemize{
\item \code{FID}: the family ID
\item \code{IID}: the within-family ID
\item \code{SEX}: the sex
\item \code{PHENO1}: a phenotypic value
}

The uppercase column names are those as used by PLINK2.}
}
\value{
a single-phenotype \link[tibble]{tibble}, with columns:
\itemize{
\item \code{FID} the family ID
\item \code{IID} the within-family ID
\item \code{P1} the phenotype
}

The column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}
}
\description{
Create a single-phenotype table from pedigree (\code{.psam}) table.
The individuals' family and within-family ID's are copied.
The \code{pheno} is copied as well, but
renamed to \code{P1} and divided by ten.
In this way, PLINK treats the phenotype table as having quantitative
traits.
}
\examples{
create_phe_table_from_psam_table(get_test_psam_table())
}
\author{
Richèl J.C. Bilderbeek
}
