% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phenotype_data_filename.R
\name{create_phenotype_data_filename}
\alias{create_phenotype_data_filename}
\title{Create a \code{phenotype_data_filename}}
\usage{
create_phenotype_data_filename(phe_filename)
}
\arguments{
\item{phe_filename}{name of a phenotype (\code{.phe}) file.
Use \link{read_plink_phe_file} to read a raw phenotype file.}
}
\description{
Create a \code{phenotype_data_filename}
}
\examples{
phe_filename <- get_plinkr_filename("pheno_with_standard_header.phe")
create_phenotype_data_filename(phe_filename)
}
\seealso{
use \link{create_test_phenotype_data_filename}
to create a \code{phenotype_data_filename} to be used in testing
}
\author{
Richèl J.C. Bilderbeek
}
