% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phenotype_data_table.R
\name{create_phenotype_data_table}
\alias{create_phenotype_data_table}
\title{Create a \code{phenotype_data_table}}
\usage{
create_phenotype_data_table(phe_table)
}
\arguments{
\item{phe_table}{a table of phenotypes.

A \code{phe_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_phe_file} to read a phenotype file.
Use \link{check_phe_table} to test if a phenotype table is valid.}
}
\description{
Create a \code{phenotype_data_table}.
}
\examples{
create_test_phenotype_data_filename()
}
\seealso{
use \link{create_test_phenotype_data_table} to create a
\code{phenotype_data_table} to be used in testing
}
\author{
Richèl J.C. Bilderbeek
}
