% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phenotype_table_from_ped_table.R
\name{create_phenotype_table_from_ped_table}
\alias{create_phenotype_table_from_ped_table}
\title{Create a single-phenotype table from pedigree (\code{.ped}) table}
\usage{
create_phenotype_table_from_ped_table(ped_table)
}
\arguments{
\item{ped_table}{a 'pedigree' table.

A \code{ped_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{family_id} The family ID (\code{FID})
\item \code{within_family_id} Within-family ID
(\code{IID}, cannot be zero)
\item \code{within_family_id_father} Within-family ID of father
(\code{0} if father isn't in dataset)
\item \code{within_family_id_mother} Within-family ID of mother
(\code{0} if mother isn't in dataset)
\item \code{sex_code} Sex code
(\code{1} = male, \code{2} = female, \code{0} = unknown)
\item \code{case_control_code} Case control code
(\code{1} = control, \code{2} = case,
\code{9}/\code{0}/non-numeric = missing data if case/control)
\item \code{snv_[x][y]} Nucleotide for the \code{x}th variant
for haplotype \code{y} (\code{y} is either \code{a} or \code{b})
in the \code{.map file} (\code{0} = no call)
}

Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.
Use \link{check_ped_table} to test if a pedigree table is valid.}
}
\value{
a single-phenotype \link[tibble]{tibble}, with columns:
\itemize{
\item \code{family_id}
\item \code{within_family_id}
\item \code{trait}
}
}
\description{
Create a single-phenotype table from pedigree (\code{.ped}) table.
The individuals' family and within-family ID's are copied.
The \link{case_control_code} is copied as well, but
renamed to \link{trait} and divided by then.
In this way, PLINK traits the phenotype table as having quantitative
traits.
}
