% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plink2_bin_data.R
\name{create_plink2_bin_data}
\alias{create_plink2_bin_data}
\title{Create a set of \code{PLINK2} binary data}
\usage{
create_plink2_bin_data()
}
\description{
Create a set of \code{PLINK2} binary data
}
\author{
Richèl J.C. Bilderbeek
}
