% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plink2_bin_filenames.R
\name{create_plink2_bin_filenames}
\alias{create_plink2_bin_filenames}
\title{Create a \code{plink2_bin_filenames}}
\usage{
create_plink2_bin_filenames(pgen_filename, psam_filename, pvar_filename)
}
\arguments{
\item{pgen_filename}{name of a \code{PLINK2} \code{.pgen} file
Use \link{read_plink2_pgen_file} to read a \code{PLINK2} \code{.pgen} file}

\item{psam_filename}{name of a \code{PLINK2} \code{.psam} file
Use \link{read_plink2_psam_file} to read a \code{PLINK2} \code{.psam} file}

\item{pvar_filename}{name of a \code{PLINK2} \code{.pvar} file
Use \link{read_plink2_pvar_file} to read a \code{PLINK2} \code{.pvar} file}
}
\value{
a list with the following elements:
\itemize{
\item \code{pgen_filename}: the \code{.pgen} filename
\item \code{psam_filename}: the \code{.psam} filename
\item \code{pvar_filename}: the \code{.pvar} filename
}
}
\description{
Create a \code{plink2_bin_filenames},
which is a list of \code{PLINK2} binary filenames
}
\examples{
create_plink2_bin_filenames(
  pgen_filename = "file.pgen",
  psam_filename = "file.psam",
  pvar_filename = "file.pvar"
)
}
\seealso{
\code{plinkr} has the following functions to create sets of filenames,
depending on the data:
\itemize{
\item \code{PLINK} text files: use \link{create_plink_text_filenames}
\item \code{PLINK} binary files: use \link{create_plink_bin_filenames}
\item \code{PLINK2} binary files: use \link{create_plink2_bin_filenames}
}
}
\author{
Richèl J.C. Bilderbeek
}
