% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plink_bin_filenames.R
\name{create_plink_bin_filenames}
\alias{create_plink_bin_filenames}
\title{Create a \code{plink_bin_filenames}}
\usage{
create_plink_bin_filenames(bed_filename, bim_filename, fam_filename)
}
\arguments{
\item{bed_filename}{name of a \code{PLINK} \code{.bed} file
Use \link{read_plink_bed_file} to read a \code{PLINK} \code{.bed} file.}

\item{bim_filename}{name of a \code{PLINK} \code{.bim} file
Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.}

\item{fam_filename}{name of a \code{PLINK} \code{.fam} file
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.}
}
\value{
a list with the following elements:
\itemize{
\item \code{bed_filename}: the \code{.bed} filename
\item \code{bim_filename}: the \code{.bim} filename
\item \code{fam_filename}: the \code{.fam} filename
}
}
\description{
Create a \code{plink_bin_filenames},
which is a list of \code{PLINK} binary filenames
}
\examples{
create_plink_bin_filenames(
  bed_filename = "file.bed",
  bim_filename = "file.bim",
  fam_filename = "file.fam"
)
}
\seealso{
\code{plinkr} has the following functions to create sets of filenames,
depending on the data:
\itemize{
\item \code{PLINK} text files: use \link{create_plink_text_filenames}
\item \code{PLINK} binary files: use \link{create_plink_bin_filenames}
\item \code{PLINK2} binary files: use \link{create_plink2_bin_filenames}
}
}
\author{
Richèl J.C. Bilderbeek
}
