% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plink_options.R
\name{create_plink_options}
\alias{create_plink_options}
\title{Create the options to run PLINK}
\usage{
create_plink_options(
  plink_version = get_default_plink_version(),
  plink_folder = get_plink_folder(),
  add_noweb = TRUE,
  os = get_os()
)
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the PLINK version.
Use \link{get_plink_versions} to get all the supported PLINK versions.}

\item{plink_folder}{folder where PLINK is installed}

\item{add_noweb}{add the \code{--noweb} option, which prevents
that PLINK will check if it is in its latest version.
By default, this value is set to true,
hence plinkr will skip this check.
Use \link{get_plink_version} to get the current PLINK version.}

\item{os}{name of the operating system,
as returned by \link[rappdirs]{app_dir}}
}
\value{
the PLINK options,
as can be checked by \link{check_plink_options}
}
\description{
Create the options to run PLINK.
Options are:
}
\details{
\itemize{
\item \link{create_plink_v1_7_options}: use PLINK v1.7
\item \link{create_plink_v1_9_options}: use PLINK v1.9
\item \link{create_custom_plink_options}: use a custom version of PLINK
}

Prefer to call these more specialized functions.
}
\examples{
# Use the default PLINK version
create_plink_options()

# Use PLINK v1.7
create_plink_options(plink_version = "1.7")
create_plink_v1_7_options()

# Use PLINK v1.9
create_plink_options(plink_version = "1.9")
create_plink_v1_9_options()

# Use PLINK v2.0
create_plink_options(plink_version = "2.0")
create_plink_v2_0_options()

# Use a custom PLINK version
create_plink_options(
  plink_version = "custom",
  plink_folder = "path_to_my_custom_plink_folder"
)
create_custom_plink_options("path_to_my_custom_plink_folder")
}
\author{
Richèl J.C. Bilderbeek
}
