% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plink_optionses.R
\name{create_plink_optionses}
\alias{create_plink_optionses}
\title{Create the PLINK options for all supported versions of PLINK}
\usage{
create_plink_optionses(plink_folder = get_plink_folder())
}
\arguments{
\item{plink_folder}{folder where PLINK is installed}
}
\value{
a list of multiple PLINK options,
as can be checked by \link{check_plink_optionses}
}
\description{
Create the PLINK options for all supported versions of PLINK.
}
\note{
'optionses' is the reduplicated plural (or: Gollumese)
of 'options', to indicate this is a list of multiple options
}
\examples{
create_plink_optionses()
}
\author{
Richèl J.C. Bilderbeek
}
