% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plink_text_filenames.R
\name{create_plink_text_filenames}
\alias{create_plink_text_filenames}
\title{Create a \code{plink_text_filenames}}
\usage{
create_plink_text_filenames(map_filename, ped_filename)
}
\arguments{
\item{map_filename}{name of a \code{PLINK} \code{.map} file
Use \link{read_plink_map_file} to read a \code{PLINK} \code{.map} file.}

\item{ped_filename}{name of a \code{PLINK} \code{.ped} file.
Use \link{read_plink_ped_file} to read a \code{PLINK} \code{.ped} file.}
}
\value{
a list with the following elements:
\itemize{
\item \code{map_filename}: the \code{.map} filename
\item \code{ped_filename}: the \code{.ped} filename
}
}
\description{
Create a \code{plink_text_filenames},
which is a list of \code{PLINK} text filenames
}
\examples{
create_plink_text_filenames(
  map_filename = "file.map",
  ped_filename = "file.ped"
)
}
\seealso{
\code{plinkr} has the following functions to create sets of filenames,
depending on the data:
\itemize{
\item \code{PLINK} text files: use \link{create_plink_text_filenames}
\item \code{PLINK} binary files: use \link{create_plink_bin_filenames}
\item \code{PLINK2} binary files: use \link{create_plink2_bin_filenames}
}
}
\author{
Richèl J.C. Bilderbeek
}
