% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_samples_selector.R
\name{create_random_samples_selector}
\alias{create_random_samples_selector}
\title{Create a random samples selector}
\usage{
create_random_samples_selector(n_samples)
}
\arguments{
\item{n_samples}{the number of samples/individuals,
as can be checked by \link{check_n_samples}}
}
\description{
Create a random samples selector
to be used in, for example, \link{select_samples}
}
\examples{
create_random_samples_selector(
  n_samples = 42
)
}
\seealso{
Use \link{create_test_random_samples_selector} to
create a random sample selector to be used in testing.

There are multiple sample selectors,
see \link{create_samples_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
