% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_snp_selector.R
\name{create_random_snp_selector}
\alias{create_random_snp_selector}
\title{Create a random SNP selector}
\usage{
create_random_snp_selector(n_snps)
}
\arguments{
\item{n_snps}{the number of SNPs}
}
\description{
Create a random SNP selector
to be used in, for example, \link{select_snps}
}
\examples{
create_random_snp_selector(
  n_snps = 42
)
}
\seealso{
Use \link{create_test_random_snp_selector} to
create a random SNP selector to be used in testing.

There are multiple SNP selectors,
see \link{create_snp_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
