% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_trait.R
\name{create_random_trait}
\alias{create_random_trait}
\title{Create a trait with no relation between genotype and phenotype
and a known minor allele frequency.}
\usage{
create_random_trait(maf = 0.25)
}
\arguments{
\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the PLINK analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
PLINK \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}
}
\description{
Create a trait with no relation between genotype and phenotype
and a known minor allele frequency.
}
\examples{
create_random_trait()
create_random_trait(maf = 0.01)
}
\author{
Richèl J.C. Bilderbeek
}
