% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_runif_trait.R
\name{create_runif_trait}
\alias{create_runif_trait}
\title{Create a trait with no relation between genotype and a phenotype
and a known minor allele frequency.}
\usage{
create_runif_trait(mafs = 0.25, n_snps = 1)
}
\arguments{
\item{mafs}{one or more minor allele frequencies.
These allele frequencies must be ordered decreasingly,
i.e. the MAF is at the first position, where the
even rarer alleles are at the second and third positions.

Note that \code{PLINK} cannot handle triallelic nor
quadallelic SNPs: \code{PLINK} will give a warning that it
is setting the rarest alleles to missing.}

\item{n_snps}{the number of SNPs}
}
\description{
Create a trait with no relation between genotype and phenotype
and a known minor allele frequency.
The phenotypes are random and uniform,
as produced by \link[stats]{runif}.
}
\examples{
create_runif_trait()
create_runif_trait(maf = 0.01)
}
\author{
Richèl J.C. Bilderbeek
}
