% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_select_random_samples_args.R
\name{create_select_random_samples_args}
\alias{create_select_random_samples_args}
\title{Create the command-line arguments to select random samples}
\usage{
create_select_random_samples_args(
  plink_bin_filenames,
  random_samples_selector,
  sample_ids_filename,
  base_output_filename,
  plink_options = create_plink_options()
)
}
\arguments{
\item{plink_bin_filenames}{the binary data files' names for \code{PLINK}
to work on,
as created by \link{create_plink_bin_filenames}.}

\item{random_samples_selector}{a sample selector
(see \link{create_samples_selector})
that allows one to select one or more random samples,
as created by \link{create_random_samples_selector}
and checked by \link{check_random_samples_selector}}

\item{sample_ids_filename}{name of a file to store \code{sample_ids}
(see \link{check_sample_ids}) to}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to select random samples
}
\author{
Richèl J.C. Bilderbeek
}
