% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_select_random_snp_args.R
\name{create_select_random_snp_args}
\alias{create_select_random_snp_args}
\title{Create the command-line arguments to select a range SNPs}
\usage{
create_select_random_snp_args(
  plink_bin_filenames,
  random_snps_selector,
  base_output_filename = plinkr::get_plinkr_tempfilename(),
  plink_options = create_plink_options()
)
}
\arguments{
\item{plink_bin_filenames}{the binary data files' names for \code{PLINK}
to work on,
as created by \link{create_plink_bin_filenames}.}

\item{random_snps_selector}{a SNP selector
(see \link{create_snp_selector})
that allows one to select one or more random SNPs,
as created by \link{create_random_snps_selector}
and checked by \link{check_random_snps_selector}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to select a range SNPs.
}
\author{
Richèl J.C. Bilderbeek
}
