% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_select_samples_args.R
\name{create_select_samples_args}
\alias{create_select_samples_args}
\title{Create the command-line arguments to select samples}
\usage{
create_select_samples_args(
  plink_bin_filenames,
  sample_selector,
  sample_ids_filename = plinkr::get_plinkr_tempfilename(pattern = "sample_ids_",
    fileext = ".txt"),
  base_output_filename = plinkr::get_plinkr_tempfilename(),
  plink_options = create_plink_options()
)
}
\arguments{
\item{plink_bin_filenames}{the binary data files' names for \code{PLINK}
to work on,
as created by \link{create_plink_bin_filenames}.}

\item{sample_selector}{a sample/individual selector,
a way to select one or more samples/individuals.
See \link{create_samples_selector} for all sample selectors.}

\item{sample_ids_filename}{name of a file to store \code{sample_ids}
(see \link{check_sample_ids}) to}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to select samples.
}
\examples{
create_select_samples_args(
  plink_bin_filenames = create_test_plink_bin_filenames(),
  sample_selector = create_test_single_sample_selector()
)
create_select_samples_args(
  plink_bin_filenames = create_plink_bin_filenames(
    bed_filename = get_plinkr_filename("select_snps.bed"),
    bim_filename = get_plinkr_filename("select_snps.bim"),
    fam_filename = get_plinkr_filename("select_snps.fam")
  ),
  sample_selector = create_test_random_samples_selector()
)
}
\author{
Richèl J.C. Bilderbeek
}
