% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_select_single_sample_args.R
\name{create_select_single_sample_args}
\alias{create_select_single_sample_args}
\title{Create the command-line arguments to select a single sample}
\usage{
create_select_single_sample_args(
  plink_bin_filenames,
  single_sample_selector,
  sample_ids_filename,
  base_output_filename,
  plink_options = create_plink_options()
)
}
\arguments{
\item{plink_bin_filenames}{the binary data files' names for \code{PLINK}
to work on,
as created by \link{create_plink_bin_filenames}.}

\item{sample_ids_filename}{name of a file to store \code{sample_ids}
(see \link{check_sample_ids}) to}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to select a single sample
}
\author{
Richèl J.C. Bilderbeek
}
