% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_select_single_snp_args.R
\name{create_select_single_snp_args}
\alias{create_select_single_snp_args}
\title{Create the command-line arguments to select a single SNP}
\usage{
create_select_single_snp_args(
  plink_bin_filenames,
  single_snp_selector,
  base_output_filename,
  plink_options = create_plink_options()
)
}
\arguments{
\item{plink_bin_filenames}{the binary data files' names for \code{PLINK}
to work on,
as created by \link{create_plink_bin_filenames}.}

\item{single_snp_selector}{a SNP selector (see \link{check_snp_selector})
that allows one to select a single SNP,
as created by \link{create_single_snp_selector}
and checked by \link{check_single_snp_selector}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to select a single SNP
}
\author{
Richèl J.C. Bilderbeek
}
