% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_select_snp_range_args.R
\name{create_select_snp_range_args}
\alias{create_select_snp_range_args}
\title{Create the command-line arguments to select a range SNPs}
\usage{
create_select_snp_range_args(
  plink_bin_filenames,
  snp_range_selector,
  base_output_filename,
  plink_options = create_plink_options()
)
}
\arguments{
\item{plink_bin_filenames}{the binary data files' names for \code{PLINK}
to work on,
as created by \link{create_plink_bin_filenames}.}

\item{snp_range_selector}{a SNP selector (see \link{create_snps_selector}),
that allows one to select a range between two SNPs,
as created by \link{create_snp_range_selector}
and checked by \link{check_snp_range_selector}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to select a range SNPs.
}
\author{
Richèl J.C. Bilderbeek
}
