% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_select_snps_args.R
\name{create_select_snps_args}
\alias{create_select_snps_args}
\title{Create the command-line arguments to select SNPs}
\usage{
create_select_snps_args(
  plink_bin_filenames,
  snp_selector,
  base_output_filename = plinkr::get_plinkr_tempfilename(),
  plink_options = create_plink_options()
)
}
\arguments{
\item{plink_bin_filenames}{the binary data files' names for \code{PLINK}
to work on,
as created by \link{create_plink_bin_filenames}.}

\item{snp_selector}{a SNP selector, a way to select one or more SNPs.
See \link{create_snp_selector} for all SNP selectors.}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to select SNPs.
}
\examples{
create_select_snps_args(
  plink_bin_filenames = create_test_plink_bin_filenames(),
  snp_selector = create_test_single_snp_selector()
)
create_select_snps_args(
  plink_bin_filenames = create_plink_bin_filenames(
    bed_filename = get_plinkr_filename("select_snps.bed"),
    bim_filename = get_plinkr_filename("select_snps.bim"),
    fam_filename = get_plinkr_filename("select_snps.fam")
  ),
  snp_selector = create_test_random_snps_selector()
)
create_select_snps_args(
  plink_bin_filenames = create_test_plink_bin_filenames(),
  snp_selector = create_test_snp_range_selector()
)
create_select_snps_args(
  plink_bin_filenames = create_test_plink_bin_filenames(),
  snp_selector = create_test_snp_window_selector()
)
}
\author{
Richèl J.C. Bilderbeek
}
