% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simulate_qt_params.R
\name{create_simulate_qt_params}
\alias{create_simulate_qt_params}
\title{Create a set of parameters for simulating
a quantitative trait}
\usage{
create_simulate_qt_params(
  n_snps,
  snp_label,
  allele_frequency_lower,
  allele_frequency_upper,
  additive_genetic_variance,
  ratio_dominance_to_additive
)
}
\arguments{
\item{n_snps}{Number of SNPs in this set}

\item{snp_label}{Label of this set of SNPs}

\item{allele_frequency_lower}{Lower allele frequency range}

\item{allele_frequency_upper}{Upper allele frequency range}

\item{additive_genetic_variance}{The additive genetic variance}

\item{ratio_dominance_to_additive}{The ratio of dominance to additive effects}
}
\description{
Create a set of parameters for simulating
for simulating a quantitative trait (see \link{simulate_qt}).
Use \link{create_test_simulate_qt_params}) for a set of parameters
using in testing.
}
\details{
This function is named after the \code{--simulate-qt} PLINK flag.
}
