% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_single_sample_selector.R
\name{create_single_sample_selector}
\alias{create_single_sample_selector}
\title{Create a sample selector to select a single sample}
\usage{
create_single_sample_selector(fid, iid)
}
\arguments{
\item{fid}{the family ID, which is called \code{FID} in \code{PLINK}.
Use \link{check_fid} to check if a \code{fid} is valid.}

\item{iid}{a within-family ID, as can be checked by \link{check_iid}}
}
\description{
Create a sample selector to select a single sample,
to be used in, for example, \link{select_samples}
}
\examples{
create_single_sample_selector(
  fid = "adams",
  iid = "3"
)
}
\seealso{
Use \link{create_test_single_sample_selector} to
create a single-sample selector to be used in testing.

There are multiple sample selectors,
see \link{create_sample_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
