% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_single_snp_selector.R
\name{create_single_snp_selector}
\alias{create_single_snp_selector}
\title{Create a SNP selector to select a single SNP}
\usage{
create_single_snp_selector(snp)
}
\arguments{
\item{snp}{a SNP, e.g. \code{rs12345678},
as can be checked by \link{check_snp}}
}
\description{
Create a SNP selector to select a single SNP,
to be used in, for example, \link{select_snps}
}
\examples{
create_single_snp_selector(
  snp = "rs123456"
)
}
\seealso{
Use \link{create_test_single_snp_selector} to
create a single-SNP selector to be used in testing.

There are multiple SNP selectors,
see \link{create_snp_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
