% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_range_selector.R
\name{create_snp_range_selector}
\alias{create_snp_range_selector}
\title{Create a SNP range selector}
\usage{
create_snp_range_selector(snp_from, snp_to)
}
\arguments{
\item{snp_from}{a SNP (see \link{check_snp}) to start from}

\item{snp_to}{a SNP (see \link{check_snp}) to end at}
}
\description{
Create a SNP range selector,
to be used in, for example, \link{select_snps}
}
\examples{
create_snp_range_selector(
  snp_from = "rs123456",
  snp_to = "rs234567"
)
}
\seealso{
Use \link{create_test_snp_range_selector} to
create a SNP window selector to be used in testing.

There are multiple SNP selectors,
see \link{create_snps_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
