% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_select_args.R
\name{create_snp_select_args}
\alias{create_snp_select_args}
\title{Create the command-line arguments to select SNPs}
\usage{
create_snp_select_args(
  plink_bin_filenames,
  snp_selector,
  base_output_filename = plinkr::get_plinkr_tempfilename(),
  plink_options = create_plink_options()
)
}
\arguments{
\item{plink_bin_filenames}{the binary data files' names for \code{PLINK}
to work on,
as created by \link{create_plink_bin_filenames}.}

\item{snp_selector}{a SNP selector, a way to select one or more SNPs,
as can be checked by \link{check_snp_selector}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to select SNPs.
}
\examples{
create_snp_select_args(
  plink_bin_filenames = create_test_plink_bin_filenames(),
  snp_selector = create_test_single_snp_selector()
)
create_snp_select_args(
  plink_bin_filenames = create_test_plink_bin_filenames(),
  snp_selector = create_test_random_snp_selector()
)
create_snp_select_args(
  plink_bin_filenames = create_test_plink_bin_filenames(),
  snp_selector = create_test_snp_range_selector()
)
create_snp_select_args(
  plink_bin_filenames = create_test_plink_bin_filenames(),
  snp_selector = create_test_snp_window_selector()
)
}
\author{
Richèl J.C. Bilderbeek
}
