% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_selector.R
\name{create_snp_selector}
\alias{create_snp_selector}
\title{Create a SNP selector}
\usage{
create_snp_selector(snp_selector)
}
\arguments{
\item{snp_selector}{a SNP selector, a way to select one or more SNPs.
See \link{create_snp_selector} for all SNP selectors.}
}
\value{
Nothing.
}
\description{
Create a SNP selector.
This function is not meant to be used,
prefer the named specialized functions instead,
which are listed at the 'See Also' section.
}
\examples{
# Prefer using the specialized funtions over 'create_snp_selector'
create_snp_selector(create_test_random_snp_selector())
create_snp_selector(create_test_single_snp_selector())
create_snp_selector(create_test_snp_range_selector())
create_snp_selector(create_test_snp_window_selector())
}
\seealso{
There are multiple SNP selectors,
\itemize{
\item A random SNP selector: \link{create_random_snp_selector}
\item A single SNP selector: \link{create_single_snp_selector}
\item A single chromosome selector: \link{create_chromosome_selector}
\item A SNP range selector: \link{create_snp_range_selector}
\item A SNP window selector: \link{create_snp_window_selector}
}

There are multiple SNP selector checking functions,
see \link{check_snp_selector} for an overview.
}
\author{
Richèl J.C. Bilderbeek
}
