% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_window_selector.R
\name{create_snp_window_selector}
\alias{create_snp_window_selector}
\title{Create a SNP window selector}
\usage{
create_snp_window_selector(snp, window_kb)
}
\description{
Create a SNP window selector,
to be used in, for example, \link{select_snps}
}
\examples{
create_snp_window_selector(
  snp = "my_snp",
  window_kb = 12.34
)
}
\author{
Richèl J.C. Bilderbeek
}
