% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snvs.R
\name{create_snvs}
\alias{create_snvs}
\title{Create a table with single nucleotide variants (SNVs)}
\usage{
create_snvs(n_snps = 1, n_individuals = 1)
}
\arguments{
\item{n_snps}{the number of SNPs,
as can be checked by \link{check_n_snps}}

\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}
}
\value{
Nothing. Will \link{stop} if the input
\code{calc_phenotype_function} is invalid
}
\description{
Create a table with single nucleotide variants (SNVs),
in the form of a \link[tibble]{tibble} that contains the
two nucleotide calls for multiple individuals.
Each individual is represented by a row.
The first two columns are named \code{snv_1a}, \code{snv_1b} and
hold the genetic data for the first SNV of a diploid organism.
If there are more SNVs, columns continue with names,
\code{snv_2a}, \code{snv_2b}, \code{snv_3a}, \code{snv_3b}, etc.
In general, nucleotides are in uppercase.
}
\author{
Richèl J.C. Bilderbeek
}
