% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_assoc_params.R
\name{create_test_assoc_params}
\alias{create_test_assoc_params}
\title{Create testing parameters for the \link{assoc} function}
\usage{
create_test_assoc_params(
  data = create_test_plink_text_data(),
  confidence_interval = 0.95,
  maf = get_lowest_maf()
)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
}}

\item{confidence_interval}{confidence interval, a value between (and
excluding both) 0.0 and 1.0. The confidence interval helps assess
the certainty of an estimation: you can be 99 percent sure
a value is within the range of the 0.99 confidence interval}

\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the \code{PLINK} analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
\code{PLINK} \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}
}
\description{
Create parameters (as created by \link{create_assoc_params})
to detect associations with quantitative traits
(using \link{assoc}) used for testing.
}
\note{
This function is named after the \code{--assoc} PLINK flag.
}
\author{
Richèl J.C. Bilderbeek
}
