% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_assoc_qt_covar_data.R
\name{create_test_assoc_qt_covar_data}
\alias{create_test_assoc_qt_covar_data}
\title{Create a dataset for the \link{assoc_qt} function to be used in testing.}
\usage{
create_test_assoc_qt_covar_data(
  data = create_test_plink_text_data(),
  phenotype_data = plinkr::create_phenotype_data_table_from_data(data = data),
  cov_table = plinkr::create_cov_table_from_data(data = data)
)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
\item data to be used by \link{assoc},
as created by \link{create_assoc_data}
\item data to be used by \link{assoc_qt},
as created by \link{create_assoc_qt_data}
}}

\item{phenotype_data}{phenotype data in the form of either
a filename to a phenotype file (\code{phenotype_data_filename})
or a list with a phenotype table (\code{phenotype_data_table})}

\item{cov_table}{a table of covariates.

A \code{cov_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of covariate values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_cov_file} to read a covariates file.
Use \link{check_cov_table} to test if a covariates table is valid.}
}
\description{
Create a dataset (as created by \link{create_assoc_qt_params})
to detect associations with quantitative traits
(using \link{assoc_qt}) used for testing.
}
\note{
This function is named after the \code{--assoc} PLINK flag.
}
\author{
Richèl J.C. Bilderbeek
}
