% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_assoc_qt_params.R
\name{create_test_assoc_qt_params}
\alias{create_test_assoc_qt_params}
\title{Create testing parameters for the \link{assoc_qt} function}
\usage{
create_test_assoc_qt_params(maf = get_lowest_maf())
}
\arguments{
\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the \code{PLINK} analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
\code{PLINK} \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}
}
\description{
Create parameters (as created by \link{create_assoc_qt_params})
to detect associations with quantitative traits
(using \link{assoc_qt}) used for testing.
}
\author{
Richèl J.C. Bilderbeek
}
