% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_chromosome_selector.R
\name{create_test_chromosome_selector}
\alias{create_test_chromosome_selector}
\title{Create a selector to select all SNPs on a chromosome,
to be used in testing}
\usage{
create_test_chromosome_selector()
}
\description{
Create a selector to select all SNPs on a chromosome,
to be used in testing.
A SNP selector is used in, for example, \link{select_snps}
}
\seealso{
Use \link{create_chromosome_selector} to
create a regular chromosome selector.

There are multiple SNP selectors,
see \link{create_snp_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
