% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_pca_result_filenames.R
\name{create_test_pca_result_filenames}
\alias{create_test_pca_result_filenames}
\title{Create a \code{pca_result_filenames} to be used in testing.}
\usage{
create_test_pca_result_filenames()
}
\value{
a list with filenames
}
\description{
Create a \code{pca_result_filenames} to be used in testing.
}
\examples{
create_test_pca_result_filenames()
}
\seealso{
Use \link{pca} to do a principal component analysis.
}
\author{
Richèl J.C. Bilderbeek
}
