% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_ped_table.R
\name{create_test_ped_table}
\alias{create_test_ped_table}
\title{Create a test \code{.ped} table}
\usage{
create_test_ped_table()
}
\value{
a \code{.ped} table
}
\description{
Create a test \code{.ped} table
}
