% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_phenotype_data_table.R
\name{create_test_phenotype_data_table}
\alias{create_test_phenotype_data_table}
\title{Create a \code{phenotype_data_table} to be used in testing}
\usage{
create_test_phenotype_data_table()
}
\description{
Create a \code{phenotype_data_table} to be used in testing
}
\examples{
create_test_phenotype_data_filename()
}
\author{
Richèl J.C. Bilderbeek
}
