% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_plink2_bin_filenames.R
\name{create_test_plink2_bin_filenames}
\alias{create_test_plink2_bin_filenames}
\title{Create a \code{plink2_bin_filenames} to be used in testing.}
\usage{
create_test_plink2_bin_filenames(
  pgen_filename = get_plinkr_filename("select_snps_plink2.pgen"),
  psam_filename = get_plinkr_filename("select_snps_plink2.psam"),
  pvar_filename = get_plinkr_filename("select_snps_plink2.pvar")
)
}
\arguments{
\item{pgen_filename}{name of a \code{PLINK2} \code{.pgen} file
Use \link{read_plink2_pgen_file} to read a \code{PLINK2} \code{.pgen} file}

\item{psam_filename}{name of a \code{PLINK2} \code{.psam} file
Use \link{read_plink2_psam_file} to read a \code{PLINK2} \code{.psam} file}

\item{pvar_filename}{name of a \code{PLINK2} \code{.pvar} file
Use \link{read_plink2_pvar_file} to read a \code{PLINK2} \code{.pvar} file}
}
\value{
a list with the following elements:
\itemize{
\item \code{pgen_filename}: the \code{.pgen} filename
\item \code{psam_filename}: the \code{.psam} filename
\item \code{pvar_filename}: the \code{.pvar} filename
}
}
\description{
Create a \code{plink2_bin_filenames} to be used in testing.
}
\examples{
create_test_plink2_bin_filenames()
}
\seealso{
Use \link{create_plink2_bin_filenames} to creata a set
of \code{PLINK2} binary filenames to be used in testing.

\code{plinkr} has the following functions to create sets of filenames,
depending on the data:
\itemize{
\item \code{PLINK} text files: use \link{create_plink_text_filenames}
\item \code{PLINK} binary files: use \link{create_plink_bin_filenames}
\item \code{PLINK2} binary files: use \link{create_plink2_bin_filenames}
}
}
\author{
Richèl J.C. Bilderbeek
}
