% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_plink_bin_data.R
\name{create_test_plink_bin_data}
\alias{create_test_plink_bin_data}
\title{Create a set of \code{PLINK} binary data to be used in testing}
\usage{
create_test_plink_bin_data()
}
\description{
Create a set of \code{PLINK} binary data to be used in testing
}
\examples{
create_test_plink_bin_data()
}
\author{
Richèl J.C. Bilderbeek
}
