% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_plink_text_data.R
\name{create_test_plink_text_data}
\alias{create_test_plink_text_data}
\title{Create PLINK text data for testing}
\usage{
create_test_plink_text_data(
  ped_table = get_test_ped_table(),
  map_table = get_test_map_table()
)
}
\arguments{
\item{ped_table}{a 'pedigree' table.

A \code{ped_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item \code{within_family_id_father} Within-family ID of father
(\code{0} if father isn't in dataset)
\item \code{within_family_id_mother} Within-family ID of mother
(\code{0} if mother isn't in dataset)
\item \code{sex_code} Sex code
(\code{1} = male, \code{2} = female, \code{0} = unknown)
\item \code{case_control_code} Case control code
(\code{1} = control, \code{2} = case,
\code{9}/\code{0}/non-numeric = missing data if case/control)
\item \code{snv_[x][y]} Nucleotide for the \code{x}th variant
for haplotype \code{y} (\code{y} is either \code{a} or \code{b})
in the \code{.map file} (\code{0} = no call)
}

The \code{FID} and \code{IID} column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}.

Use \link{read_plink_ped_file} to read a \code{PLINK} \code{.ped} file.
Use \link{check_ped_table} to test if a pedigree table is valid.}

\item{map_table}{a genetic mapping table.

A \code{map_table} is a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{CHR}: the chromosome code or contig name.
The chromosome number cannot exceed 95
(\url{https://www.cog-genomics.org/plink/1.9/input#chr_set})
\item \code{SNP}: Variant identifier
\item \code{position_cm}: Position in morgans or centimorgans.
This value is optional. Zeroes denote it is unused
\item \code{BP}: Base-pair coordinat
}

Use \link{read_plink_map_file} to read a \code{PLINK} \code{.map} file.
Use \link{check_map_table} to test if a genetic mapping table is valid.}
}
\description{
Create PLINK text data for testing
}
\details{
These is the same tables as can be obtained by reading
the PLINK v1.7 example files called
\code{test.ped} and \code{test.map}.
}
\examples{
data <- create_test_plink_text_data()
is_plink_text_data(data)
is_plink_bin_data(data)
is_plink2_bin_data(data)
}
\author{
Richèl J.C. Bilderbeek
}
