% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_plink_text_filenames.R
\name{create_test_plink_text_filenames}
\alias{create_test_plink_text_filenames}
\title{Create a set of \code{plink_text_filenames} to be used in testing.}
\usage{
create_test_plink_text_filenames(
  map_filename = "test.map",
  ped_filename = "test.ped"
)
}
\arguments{
\item{map_filename}{name of a \code{PLINK} \code{.map} file
Use \link{read_plink_map_file} to read a \code{PLINK} \code{.map} file.}

\item{ped_filename}{name of a \code{PLINK} \code{.ped} file.
Use \link{read_plink_ped_file} to read a \code{PLINK} \code{.ped} file.}
}
\value{
a list with the following elements:
\itemize{
\item \code{map_filename}: the \code{.map} filename
\item \code{ped_filename}: the \code{.ped} filename
}
}
\description{
Create a set of \code{plink_text_filenames} to be used in testing.
}
\examples{
create_test_plink_text_filenames()
}
\seealso{
Use \link{create_plink_text_filenames} to creata a set
of \code{PLINK} text filenames.

\code{plinkr} has the following functions to create sets of filenames,
depending on the data:
\itemize{
\item \code{PLINK} text files: use \link{create_plink_text_filenames}
\item \code{PLINK} binary files: use \link{create_plink_bin_filenames}
\item \code{PLINK2} binary files: use \link{create_plink2_bin_filenames}
}
}
\author{
Richèl J.C. Bilderbeek
}
