% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_random_samples_selector.R
\name{create_test_random_samples_selector}
\alias{create_test_random_samples_selector}
\title{Create a random sample selector to be used in testing}
\usage{
create_test_random_samples_selector(n_samples = 2)
}
\arguments{
\item{n_samples}{the number of samples/individuals,
as can be checked by \link{check_n_samples}}
}
\description{
Create a random sample selector to be used in testing,
as can be used in, for example, \link{select_samples}
}
\examples{
create_test_random_samples_selector()
create_test_random_samples_selector(n_samples = 42)

check_random_samples_selector(create_test_random_samples_selector())
is_random_samples_selector(create_test_random_samples_selector())

check_sample_selector(create_test_random_samples_selector())
is_sample_selector(create_test_random_samples_selector())
}
\seealso{
Use \link{create_random_samples_selector} to
create a random sample selector.

There are multiple sample selector checking functions,
see \link{create_samples_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
