% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_random_snp_selector.R
\name{create_test_random_snp_selector}
\alias{create_test_random_snp_selector}
\title{Create a random SNP selector to be used in testing}
\usage{
create_test_random_snp_selector(n_snps = 2)
}
\arguments{
\item{n_snps}{the number of SNPs}
}
\description{
Create a random SNP selector to be used in testing,
as can be used in, for example, \link{select_snps}
}
\examples{
create_test_random_snp_selector()
create_test_random_snp_selector(n_snps = 42)

check_random_snp_selector(create_test_random_snp_selector())
is_random_snp_selector(create_test_random_snp_selector())

check_snp_selector(create_test_random_snp_selector())
is_snp_selector(create_test_random_snp_selector())
}
\seealso{
Use \link{create_random_snp_selector} to
create a random SNP selector.

There are multiple SNP selector checking functions,
see \link{check_snp_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
