% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_sim_qtl_params.R
\name{create_test_sim_qtl_params}
\alias{create_test_sim_qtl_params}
\title{Create a set of testing parameters for simulating
a quantitative trait}
\usage{
create_test_sim_qtl_params()
}
\description{
Create a set of testing parameters for simulating
a quantitative trait
}
