% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_simulate_qt_params.R
\name{create_test_simulate_qt_params}
\alias{create_test_simulate_qt_params}
\title{Create a set of testing parameters for simulating
a quantitative trait}
\usage{
create_test_simulate_qt_params(
  n_snps = 10,
  snp_label = "test_trait",
  allele_frequency_lower = 0.05,
  allele_frequency_upper = 0.95,
  additive_genetic_variance = 0.01,
  ratio_dominance_to_additive = 0
)
}
\arguments{
\item{n_snps}{Number of SNPs in this set}

\item{snp_label}{Label of this set of SNPs}

\item{allele_frequency_lower}{Lower allele frequency range}

\item{allele_frequency_upper}{Upper allele frequency range}

\item{additive_genetic_variance}{The additive genetic variance}

\item{ratio_dominance_to_additive}{The ratio of dominance to additive effects}
}
\description{
Create a set of parameters (see \link{create_simulate_qt_params})
for simulating a quantitative trait (see \link{simulate_qt})
to be used for testing.
}
\details{
This function is named after the \code{--simulate-qt} PLINK flag.
}
