% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_single_sample_selector.R
\name{create_test_single_sample_selector}
\alias{create_test_single_sample_selector}
\title{Create a sample selector to select a single sample, to be used in testing}
\usage{
create_test_single_sample_selector(fid = "adams", iid = "morticia")
}
\arguments{
\item{fid}{the family ID, which is called \code{FID} in \code{PLINK}.
Use \link{check_fid} to check if a \code{fid} is valid.}

\item{iid}{a within-family ID, as can be checked by \link{check_iid}}
}
\description{
Create a sample selector to select a single sample, to be used in testing.
A sample selector is used in, for example, \link{select_samples}
}
\examples{
create_test_single_sample_selector()

create_test_single_sample_selector(
  fid = "simpsons"
)
create_test_single_sample_selector(
  iid = "bart"
)
}
\seealso{
Use \link{create_single_sample_selector} to
create a regular single-sample selector.

There are multiple sample selectors,
see \link{create_samples_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
