% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_snp_selector.R
\name{create_test_snp_window_selector}
\alias{create_test_snp_window_selector}
\title{Create a SNP window selector}
\usage{
create_test_snp_window_selector(snp = "rs12345678", window_kb = 0.002)
}
\description{
Create a SNP window selector,
to be used in, for example, \link{select_snps}
}
\examples{
create_test_snp_window_selector()

create_test_snp_window_selector(
  snp = "my_snp"
)

create_test_snp_window_selector(
  window_kb = 12.34
)
}
\author{
Richèl J.C. Bilderbeek
}
