% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  log_filename,
  n_individuals,
  plink_folder,
  sim_filename,
  simfreq_filename,
  sim_qt_params,
  temp_sim_filename,
  url,
  verbose
)
}
\arguments{
\item{log_filename}{name of a PLINK \code{.log} file}

\item{n_individuals}{number of individuals}

\item{plink_folder}{folder where PLINK is installed}

\item{sim_filename}{name of a PLINK \code{.sim} file}

\item{simfreq_filename}{name of a PLINK \code{.simfreq} file}

\item{sim_qt_params}{the parameters for a quantitative
traits simulation, as can be created by
\link{create_sim_qt_params}}

\item{temp_sim_filename}{temporary file to store simulation
parameters, which is a PLINK \code{.sim} file}

\item{url}{PLINK download URL}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
