% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_random_snps.R
\name{filter_random_snps}
\alias{filter_random_snps}
\title{Create a \code{PLINK} dataset that has randomly selected SNPs}
\usage{
filter_random_snps(
  base_input_filename,
  base_output_filename,
  n_snps,
  verbose = FALSE
)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{n_snps}{the number of SNPs,
as can be checked by \link{check_n_snps}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
the names of the PLINK binary files created,
as can be checked by \link{check_plink_bin_filenames}
}
\description{
Create a \code{PLINK} dataset that has randomly selected SNPs.
}
\author{
Richèl J.C. Bilderbeek
}
