% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_from_bfiles.R
\name{freq_from_bfile}
\alias{freq_from_bfile}
\title{Let \code{PLINK}/\code{PLINK2}
calculate the allele frequencies.}
\usage{
freq_from_bfile(bfile, out)
}
\arguments{
\item{bfile}{the base filename of the binary files (i.e.
a \code{.bed}, \code{.bim} and \code{.fam} file).
This parameter is named after the \code{PLINK}
\code{--bfile} flag}

\item{out}{the base filename of the output files.
This parameter is named after the \code{PLINK}
\code{--out} flag}
}
\value{
a \link{list} with the following elements:
\itemize{
\item \code{frq_table} table with ?
\item \code{log} the log text
}
}
\description{
Let \code{PLINK}/\code{PLINK2}
calculate the allele frequencies
from PLINK1 binary filenames
}
\note{
This function is named after the \code{--freq} flag
used by \code{PLINK}
}
\author{
Richèl J.C. Bilderbeek
}
