% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_description.R
\name{get_data_description}
\alias{get_data_description}
\title{Get a description of the data}
\usage{
get_data_description(data)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
\item data to be used by \link{assoc},
as created by \link{create_assoc_data}
\item data to be used by \link{assoc_qt},
as created by \link{create_assoc_qt_data}
}}
}
\description{
Get a description of the data
}
\examples{
get_data_description(data = create_test_plink_text_data())
get_data_description(data = create_test_plink_bin_data())
get_data_description(data = create_test_plink2_bin_data())
get_data_description(data = create_test_plink_text_filenames())
get_data_description(data = create_test_plink_bin_filenames())
get_data_description(data = create_test_plink2_bin_filenames())
}
\author{
Richèl J.C. Bilderbeek
}
