% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_plink_exe_path.R
\name{get_default_plink_exe_path}
\alias{get_default_plink_exe_path}
\title{Get the path to the default PLINK executable as used by plinkr}
\usage{
get_default_plink_exe_path(
  plink_version = get_default_plink_version(),
  os = get_os()
)
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the \code{PLINK} version.
Use \link{get_plink_versions} to get all the supported \code{PLINK} versions.}

\item{os}{name of the operating system,
as returned by \link[rappdirs]{app_dir}}
}
\value{
path to the default PLINK executable  as used by plinkr
}
\description{
Get the path to the default PLINK executable as used by plinkr
}
\author{
Richèl J.C. Bilderbeek
}
