% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_plink_version.R
\name{get_default_plink_version}
\alias{get_default_plink_version}
\title{Get the default PLINK version as used by plinkr}
\usage{
get_default_plink_version()
}
\description{
Get the default PLINK version as used by plinkr
}
