% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heightest_maf.R
\name{get_heighest_maf}
\alias{get_heighest_maf}
\title{Get the highest value for a minor allelle frequency (MAF).}
\usage{
get_heighest_maf()
}
\description{
Get the highest value for a minor allelle frequency (MAF).
which is the highest value possible below 0.5
}
\examples{
get_heighest_maf()
}
\author{
Richèl J.C. Bilderbeek
}
