% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lowest_maf.R
\name{get_lowest_maf}
\alias{get_lowest_maf}
\title{Get the lowest non-zero value for a minor allelle frequency (MAF).}
\usage{
get_lowest_maf()
}
\description{
Get the lowest non-zero value for a minor allelle frequency (MAF),
which is the lowest representable non-zero floating-point value,
also known as \code{.Machine$double.xmin} (see \link{.Machine})
}
\author{
Richèl J.C. Bilderbeek
}
