% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minor_alelle_frequencies.R
\name{get_minor_alelle_frequencies}
\alias{get_minor_alelle_frequencies}
\title{Get the minor allele frequencies from genetic data}
\usage{
get_minor_alelle_frequencies(
  data,
  plink_options = create_plink_options(),
  temp_folder = plinkr::get_plinkr_tempfilename(),
  verbose = FALSE
)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
\item data to be used by \link{assoc},
as created by \link{create_assoc_data}
\item data to be used by \link{assoc_qt},
as created by \link{create_assoc_qt_data}
}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{temp_folder}{temporary folder to store results in}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
the minor allele frequecies
}
\description{
Get the minor allele frequencies from genetic data
}
\examples{
if (is_plink_installed()) {
  assoc_qt_data <- create_demo_assoc_qt_data()

  # PLINK text data
  get_minor_alelle_frequencies(assoc_qt_data$data)

  # PLINK binary data
  assoc_qt_data$data <- convert_plink_text_data_to_plink_bin_data(
    assoc_qt_data$data
  )
  get_minor_alelle_frequencies(assoc_qt_data$data)
}
}
\author{
Richèl J.C. Bilderbeek
}
