% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_samples.R
\name{get_n_samples}
\alias{get_n_samples}
\title{Get the number of samples from a table}
\usage{
get_n_samples(table)
}
\arguments{
\item{table}{a table}
}
\value{
the number of samples
}
\description{
Get the number of samples from a table
}
\seealso{
To obtain the number of samples/variants for a specific table:
\itemize{
\item \link{get_n_samples_from_bed_table} for a \code{.bed} table
\item \link{get_n_samples_from_bim_table} for a \code{.bim} table
\item \link{get_n_samples_from_fam_table} for a \code{.fam} table
\item \link{get_n_samples_from_map_table} for a \code{.map} table
\item \link{get_n_samples_from_ped_table} for a \code{.ped} table
\item \link{get_n_samples_from_pgen_table} for a \code{.pgen} table
\item \link{get_n_samples_from_psam_table} for a \code{.psam} table
\item \link{get_n_samples_from_pvar_table} for a \code{.pvar} table
}
}
