% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_samples.R
\name{get_n_samples_from_fam_table}
\alias{get_n_samples_from_fam_table}
\title{Get the number of samples from a \code{.fam} table}
\usage{
get_n_samples_from_fam_table(fam_table)
}
\arguments{
\item{fam_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{fam}: the family ID
\item \code{id}: the individual's ID
\item \code{pat}: ID of father
\item \code{mat}: ID of mother
\item \code{sex}: the gender
\item \code{pheno}: a phenotype
}

Use \link{get_test_fam_table} to get a \code{.fam} table as used in testing.
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.
Use \link{check_fam_table} to test if a \code{.fam} table is valid.}
}
\value{
the number of samples
}
\description{
Get the number of samples from a \code{.fam} table
}
