% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_samples.R
\name{get_n_samples_from_pgen_table}
\alias{get_n_samples_from_pgen_table}
\title{Get the number of samples from a \code{.pgen} table}
\usage{
get_n_samples_from_pgen_table(pgen_table)
}
\arguments{
\item{pgen_table}{an \link{array} that maps the individuals
to their SNPs, with as much rows as individuals, and as much
SNPs as columns. Optionally, the row names are the individuals' IDs,
where the column names are the SNP ID's}
}
\value{
the number of samples
}
\description{
Get the number of samples from a \code{.pgen} table
}
