% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_samples.R
\name{get_n_samples_from_psam_table}
\alias{get_n_samples_from_psam_table}
\title{Get the number of samples from a \code{.psam} table}
\usage{
get_n_samples_from_psam_table(psam_table)
}
\arguments{
\item{psam_table}{a \link[tibble]{tibble} with as much rows as
individuals. It has the following columns:
\itemize{
\item \code{FID}: the family ID
\item \code{IID}: the within-family ID
\item \code{SEX}: the sex
\item \code{PHENO1}: a phenotypic value
}

The uppercase column names are those as used by PLINK2.}
}
\value{
the number of samples
}
\description{
Get the number of samples from a \code{.psam} table
}
