% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_samples.R
\name{get_n_samples_from_pvar_table}
\alias{get_n_samples_from_pvar_table}
\title{Get the number of samples from a \code{.pvar} table}
\usage{
get_n_samples_from_pvar_table(pvar_table)
}
\arguments{
\item{pvar_table}{a \link[tibble]{tibble} with as much rows as
SNPs. It has the following columns:
\itemize{
\item \code{CHROM}: the chromosome
\item \code{POS}: the position on the chromosome
\item \code{ID}: the SNP ID
\item \code{REF}: the reference variant
\item \code{ALT}: the alternate variant
}

The uppercase column names are those as used by PLINK2.}
}
\value{
the number of samples
}
\description{
Get the number of samples from a \code{.pvar} table
}
