% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_snps.R
\name{get_n_snps}
\alias{get_n_snps}
\title{Get the number of SNPs from a table}
\usage{
get_n_snps(table)
}
\arguments{
\item{table}{a table}
}
\value{
the number of SNPs
}
\description{
Get the number of SNPs from a table
}
\seealso{
To obtain the number of SNPs/variants for a specific table:
\itemize{
\item \link{get_n_snps_from_bed_table} for a \code{.bed} table
\item \link{get_n_snps_from_bim_table} for a \code{.bim} table
\item \link{get_n_snps_from_map_table} for a \code{.map} table
\item \link{get_n_snps_from_ped_table} for a \code{.ped} table
\item \link{get_n_snps_from_pgen_table} for a \code{.pgen} table
\item \link{get_n_snps_from_pvar_table} for a \code{.pvar} table
}
}
