% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_snps.R
\name{get_n_snps_from_bim_table}
\alias{get_n_snps_from_bim_table}
\title{Get the number of SNPs from a \code{.bim} table}
\usage{
get_n_snps_from_bim_table(bim_table)
}
\arguments{
\item{bim_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{chr}: the chromosome number
\item \code{id}: the SNP ID
\item \code{posg}: the position
\item \code{pos}: the position
\item \code{ref}: something
\item \code{alt}: something
}

Use \link{get_test_bim_table} to get a \code{.bim} table as used in testing.
Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.
Use \link{check_bim_table} to test if a \code{.bim} table is valid.}
}
\value{
the number of SNPs
}
\description{
Get the number of SNPs from a \code{.bim} table
}
