% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Get the current operating system}
\usage{
get_os()
}
\value{
the name of the operating system, as returned by
\link[rappdirs]{app_dir()}
}
\description{
Get the current operating system
}
\examples{
get_os()
}
\author{
Richèl J.C. Bilderbeek
}
