% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phenotypes.R
\name{get_phenotypes}
\alias{get_phenotypes}
\title{Get the phenotypes that are named after their genetic background}
\usage{
get_phenotypes()
}
\description{
Get the phenotypes that are named after their genetic background
}
