% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink1_versions.R
\name{get_plink1_versions}
\alias{get_plink1_versions}
\title{Get the supported version of PLINK}
\usage{
get_plink1_versions()
}
\description{
Get the supported version of PLINK as string.
}
\details{
Use \link{create_plink_optionses} to obtain a list
of \code{plink_options}.
}
\examples{
get_plink1_versions()
}
\author{
Richèl J.C. Bilderbeek
}
