% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_download_url.R
\name{get_plink_download_url}
\alias{get_plink_download_url}
\title{Get the URL to download a version of PLINK from}
\usage{
get_plink_download_url(plink_version = get_default_plink_version())
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}}
}
\description{
Get the URL to download a version of PLINK from
}
