% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_download_url_mac.R
\name{get_plink_download_url_mac}
\alias{get_plink_download_url_mac}
\title{Get the URL to download a Mac version of PLINK}
\usage{
get_plink_download_url_mac(plink_version = get_default_plink_version())
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the PLINK version.
Use \link{get_plink_versions} to get all the supported PLINK versions.}
}
\value{
a URL
}
\description{
Get the URL to download a Mac version of PLINK
}
\author{
Richèl J.C. Bilderbeek
}
