% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_example_filename.R
\name{get_plink_example_filename}
\alias{get_plink_example_filename}
\title{Get the path to a PLINK example file}
\usage{
get_plink_example_filename(example_filename, plink_folder = get_plink_folder())
}
\arguments{
\item{example_filename}{name of the example file}

\item{plink_folder}{folder where PLINK is installed}
}
\value{
the path to a PLINK example file, if it exists.
Will \link{stop} if the file does not exist
}
\description{
Get the path to a PLINK example file
}
\examples{
if (is_plink_installed()) {
  get_plink_example_filename("test.map")
  get_plink_example_filename("test.ped")
}
}
