% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_example_filenames.R
\name{get_plink_example_filenames}
\alias{get_plink_example_filenames}
\title{Get the full paths to all PLINK example files}
\usage{
get_plink_example_filenames(plink_options = create_plink_options())
}
\arguments{
\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the full paths to all PLINK example files
}
\description{
Get the full paths to all PLINK example files
}
\examples{
if (is_plink_installed()) {
  # Default version
  get_plink_example_filenames()

  # v1.7
  get_plink_example_filenames(create_plink_v1_7_options())

  # v1.9
  get_plink_example_filenames(create_plink_v1_9_options())
}
}
\author{
Richèl J.C. Bilderbeek
}
