% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_exe_name.R
\name{get_plink_exe_name}
\alias{get_plink_exe_name}
\title{Get the name of the PLINK executable}
\usage{
get_plink_exe_name(plink_version = get_default_plink_version(), os = get_os())
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the \code{PLINK} version.
Use \link{get_plink_versions} to get all the supported \code{PLINK} versions.}

\item{os}{name of the operating system,
as returned by \link[rappdirs]{app_dir}}
}
\value{
path to the default PLINK executable  as used by plinkr
}
\description{
Get the name of the PLINK executable
}
\examples{
get_plink_exe_name()
get_plink_exe_name(plink_version = "2.0")
get_plink_exe_name(os = "win")
}
\author{
Richèl J.C. Bilderbeek
}
