% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_exe_path.R
\name{get_plink_exe_path}
\alias{get_plink_exe_path}
\title{Get the path to the default PLINK executable as used by plinkr}
\usage{
get_plink_exe_path(plink_folder = get_plink_folder())
}
\arguments{
\item{plink_folder}{folder where PLINK is installed}
}
\value{
path to the default PLINK executable  as used by plinkr
}
\description{
Get the path to the default PLINK executable as used by plinkr
}
