% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_folder.R
\name{get_plink_folder}
\alias{get_plink_folder}
\title{Get the path to the default PLINK folder as used by plinkr}
\usage{
get_plink_folder()
}
\value{
path to the default PLINK folder as used by plinkr
}
\description{
Get the path to the default PLINK folder as used by plinkr
}
