% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_help_text.R
\name{get_plink_help_text}
\alias{get_plink_help_text}
\title{Get the text of the PLINK help file}
\usage{
get_plink_help_text(plink_options = create_plink_options())
}
\value{
the text of the PLINK help file
}
\description{
Get the text of the PLINK help file
}
\examples{
if (is_plink_installed()) {

  # Default version
  get_plink_help_text()

  # PLINK v1.7
  get_plink_help_text(create_plink_v1_7_options())

  # PLINK v1.9
  get_plink_help_text(create_plink_v1_9_options())

  # Custom PLINK version, secretly v1.9
  plink_folder <- dirname(get_plink_exe_path(create_plink_v1_9_options()))
  get_plink_help_text(create_custom_plink_options(plink_folder))
}
}
\author{
Richèl J.C. Bilderbeek
}
