% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_version.R
\name{get_plink_version}
\alias{get_plink_version}
\title{Get the version of PLINK}
\usage{
get_plink_version(plink_options = create_plink_options())
}
\arguments{
\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the version of PLINK
}
\description{
Get the version of PLINK
}
\author{
Richèl J.C. Bilderbeek
}
