% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plinkr_filename.R
\name{get_plinkr_filename}
\alias{get_plinkr_filename}
\title{Get the path to a plinkr file}
\usage{
get_plinkr_filename(example_filename)
}
\arguments{
\item{example_filename}{name of the example file}
}
\value{
the path to a plinkr file, if it exists.
Will \link{stop} if the file does not exist
}
\description{
Get the path to a plinkr file
}
\examples{
get_plinkr_filename("plink.log")
}
