% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plinkr_folder.R
\name{get_plinkr_folder}
\alias{get_plinkr_folder}
\title{Get the folder where \link{plinkr} stores its temporary files}
\usage{
get_plinkr_folder()
}
\value{
path to the default \code{PLINK} folder as used by \link{plinkr}
}
\description{
Get the folder where \link{plinkr} stores its temporary files.
}
\details{
Use \link{get_plink_folder} to get the folder where \link{plinkr}
stores the versions of PLINK
}
\author{
Richèl J.C. Bilderbeek
}
