% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_ids_from_fam_file.R
\name{get_sample_ids_from_fam_file}
\alias{get_sample_ids_from_fam_file}
\title{Get the sample IDs from a \code{fam} file}
\usage{
get_sample_ids_from_fam_file(fam_filename)
}
\arguments{
\item{fam_filename}{name of a \code{PLINK} \code{.fam} file
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.}
}
\value{
the \code{sample_ids} (see \link{check_sample_ids})
}
\description{
Get the sample IDs from a \code{.fam} file
}
\examples{
fam_filename <- get_plinkr_filename("demo_assoc.fam")
get_sample_ids_from_fam_file(fam_filename)
}
\seealso{
use \link{get_sample_ids_from_fam_table} to extract
the \code{sample_ids} from a \code{.fam} table that is already loaded in memory
}
\author{
Richèl J.C. Bilderbeek
}
