% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_ids_from_fam_table.R
\name{get_sample_ids_from_fam_table}
\alias{get_sample_ids_from_fam_table}
\title{Get the sample IDs from a \code{.fam} file}
\usage{
get_sample_ids_from_fam_table(fam_table)
}
\arguments{
\item{fam_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{fam}: the family ID
\item \code{id}: the individual's ID
\item \code{pat}: ID of father
\item \code{mat}: ID of mother
\item \code{sex}: the gender
\item \code{pheno}: a phenotype
}

Use \link{get_test_fam_table} to get a \code{.fam} table as used in testing.
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.
Use \link{check_fam_table} to test if a \code{.fam} table is valid.}
}
\value{
the \code{sample_ids} (see \link{check_sample_ids})
}
\description{
Get the sample IDs from a \code{.fam} file
}
\examples{
fam_table <- get_test_fam_table()
get_sample_ids_from_fam_table(fam_table)
}
\seealso{
use \link{get_sample_ids_from_fam_file} to extract
the \code{sample_ids} from a \code{.fam} file
}
\author{
Richèl J.C. Bilderbeek
}
