% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_ids_from_phe_file.R
\name{get_sample_ids_from_phe_file}
\alias{get_sample_ids_from_phe_file}
\title{Get the sample IDs from a \code{.phe} file}
\usage{
get_sample_ids_from_phe_file(phe_filename)
}
\arguments{
\item{phe_filename}{name of a phenotype (\code{.phe}) file,
as checked by \link{check_phe_filename}.
Use \link{read_plink_phe_file} to read a raw phenotype file.}
}
\value{
the \code{sample_ids} (see \link{check_sample_ids})
}
\description{
Get the sample IDs from a \code{.phe} file
}
\examples{
phe_filename <- get_plinkr_filename("demo_assoc_qt.phe")
get_sample_ids_from_phe_file(phe_filename)
}
\seealso{
use \link{get_sample_ids_from_phe_table} to extract
the \code{sample_ids} from a \code{.phe} table that is already loaded in memory
}
\author{
Richèl J.C. Bilderbeek
}
