% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_ids_from_phe_table.R
\name{get_sample_ids_from_phe_table}
\alias{get_sample_ids_from_phe_table}
\title{Get the sample IDs from a \code{.phe} file}
\usage{
get_sample_ids_from_phe_table(phe_table)
}
\arguments{
\item{phe_table}{a table of phenotypes.

A \code{phe_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_phe_file} to read a phenotype file.
Use \link{check_phe_table} to test if a phenotype table is valid.}
}
\value{
the \code{sample_ids} (see \link{check_sample_ids})
}
\description{
Get the sample IDs from a \code{.phe} file
}
\examples{
phe_table <- get_test_phe_table()
get_sample_ids_from_phe_table(phe_table)
}
\seealso{
use \link{get_sample_ids_from_phe_file} to extract
the \code{sample_ids} from a \code{.phe} file
}
\author{
Richèl J.C. Bilderbeek
}
