% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_bed_table.R
\name{get_test_bed_table}
\alias{get_test_bed_table}
\title{Get a simple \code{.bed} table}
\usage{
get_test_bed_table(n_snps = 2)
}
\arguments{
\item{n_snps}{the number of SNPs}
}
\value{
a \code{.bed} table
}
\description{
Get a simple \code{.bed} table,
as can be used in testing.
}
\note{
The function has the word \code{get} in its name,
as getting the result is trivial.
When getting the result is non-trivial,
the word \code{create} is used.
}
\examples{
get_test_bed_table()
check_bed_table(get_test_bed_table())
is_bed_table(get_test_bed_table())
}
\author{
Richèl J.C. Bilderbeek
}
