% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_bim_table.R
\name{get_test_bim_table}
\alias{get_test_bim_table}
\title{Get a simple genetic mapping (\code{.bim}) table}
\usage{
get_test_bim_table()
}
\value{
a mapping (\code{.map}) table
}
\description{
Get a simple genetic mapping (\code{.map}) table,
as can be used in testing.
This is the same table as can be obtained by reading
the PLINK v1.7 example file called \code{test.map}.
}
\note{
The function has the word \code{get} in its name,
as getting the result is trivial. When getting the result is non-trivial,
the word \code{create} is used.
}
\examples{
get_test_bim_table()
}
\seealso{
use \link{create_demo_map_table} to get simulated
genetic mapping tables
}
\author{
Richèl J.C. Bilderbeek
}
