% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_eigenvec_table.R
\name{get_test_eigenvec_table}
\alias{get_test_eigenvec_table}
\title{Get a simple \code{.eigenvec} table}
\usage{
get_test_eigenvec_table()
}
\value{
a \code{.eigenvec} table
}
\description{
Get a simple \code{.eigenvec} table,
as can be used in testing.
}
\note{
The function has the word \code{get} in its name,
as getting the result is trivial. When getting the result is non-trivial,
the word \code{create} is used.
}
\examples{
get_test_eigenvec_table()
}
\author{
Richèl J.C. Bilderbeek
}
