% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_fam_table.R
\name{get_test_fam_table}
\alias{get_test_fam_table}
\title{Get a simple (\code{.fam}) table}
\usage{
get_test_fam_table()
}
\value{
a (\code{.fam}) table
}
\description{
Get a simple (\code{.fam}) table,
as can be used in testing.
}
\note{
The function has the word \code{get} in its name,
as getting the result is trivial. When getting the result is non-trivial,
the word \code{create} is used.
}
\examples{
get_test_fam_table()
}
\author{
Richèl J.C. Bilderbeek
}
