% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_psam_table.R
\name{get_test_psam_table}
\alias{get_test_psam_table}
\title{Get a simple \code{.psam} table}
\usage{
get_test_psam_table()
}
\value{
a \code{.psam} table
}
\description{
Get a simple \code{.psam} table,
as can be used in testing.
This is the same table as can be obtained by reading
the plinkr example file called
\code{toy_v1_9_after_make-bed_after_make-pgen.psam}.
}
\note{
The function has the word \code{get} in its name,
as getting the result is trivial. When getting the result is non-trivial,
the word \code{create} is used.
}
\examples{
get_test_psam_table()

# Same as:
plinkr::read_plink2_psam_file(
  plinkr::get_plinkr_filename("toy_v1_9_after_make-bed_after_make-pgen.psam")
)
}
\author{
Richèl J.C. Bilderbeek
}
