% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_pvar_table.R
\name{get_test_pvar_table}
\alias{get_test_pvar_table}
\title{Get a simple \code{.pvar} table}
\usage{
get_test_pvar_table()
}
\value{
a \code{.pvar} table
}
\description{
Get a simple \code{.pvar} table,
as can be used in testing.
This is the same table as can be obtained by reading
the plinkr example file called
\code{toy_v1_9_after_make-bed_after_make-pgen.pvar}.
}
\note{
The function has the word \code{get} in its name,
as getting the result is trivial. When getting the result is non-trivial,
the word \code{create} is used.
}
\examples{
get_test_pvar_table()
}
\author{
Richèl J.C. Bilderbeek
}
