% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_sample_ids.R
\name{get_test_sample_ids}
\alias{get_test_sample_ids}
\title{Create a \code{sample_ids} to be used in testing}
\usage{
get_test_sample_ids()
}
\value{
a \code{sample_ids}
}
\description{
Create a \code{sample_ids} to be used in testing
}
\examples{
check_sample_ids(get_test_sample_ids())
}
\author{
Richèl J.C. Bilderbeek
}
