% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_bed_table_attributes.R
\name{has_bed_table_attributes}
\alias{has_bed_table_attributes}
\title{Determine if the table has the attributes of a \code{bed_table}}
\usage{
has_bed_table_attributes(table)
}
\arguments{
\item{table}{a table}
}
\value{
TRUE if the table has the attributes of a \code{bed_table}
}
\description{
Determine if the table has the attributes of a \code{bed_table}
}
\author{
Richèl J.C. Bilderbeek
}
