% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_pgen_table_attributes.R
\name{has_pgen_table_attributes}
\alias{has_pgen_table_attributes}
\title{Determine if the table has the attributes of a \code{pgen_table}}
\usage{
has_pgen_table_attributes(table)
}
\value{
TRUE if the table has the attributes of a \code{pgen_table}
}
\description{
Determine if the table has the attributes of a \code{pgen_table}
}
\author{
Richèl J.C. Bilderbeek
}
