% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_plink_text_files.R
\name{has_plink_text_files}
\alias{has_plink_text_files}
\title{Determine if using \code{base_input_filename} there are \code{PLINK} text files}
\usage{
has_plink_text_files(base_input_filename)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}
}
\value{
TRUE if the \code{PLINK} text files exist
}
\description{
Determine if using \code{base_input_filename} there are \code{PLINK} text files.
}
\details{
This is done by searching the \code{PLINK} text files at:
\itemize{
\item \code{base_input_filename.map}
\item \code{base_input_filename.ped}
}

If both files exist, this function returns TRUE
}
\note{
\itemize{
\item the \code{PLINK} text files are not tested for validity
\item it is not checked if a log file is present
}
}
\examples{
# TRUE on PLINK v1.7 testing text data
has_plink_text_files(
  tools::file_path_sans_ext(get_plinkr_filename("test_v1_7.map"))
)
# FALSE on PLINK binary data
has_plink_text_files(
  tools::file_path_sans_ext(
    get_plinkr_filename("test_v1_7_after_make-bed.bed")
  )
)
# FALSE on PLINK2 binary data
has_plink_text_files(
  tools::file_path_sans_ext(
    get_plinkr_filename("toy_v1_9_after_make-bed_after_make-pgen.pgen")
  )
)
}
\author{
Richèl J.C. Bilderbeek
}
