% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_plink.R
\name{install_plink}
\alias{install_plink}
\title{Install a specific version of PLINK}
\usage{
install_plink(plink_options = create_plink_options())
}
\arguments{
\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
Nothing
}
\description{
Install a specific version of PLINK
}
\author{
Richèl J.C. Bilderbeek
}
