% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_plinks.R
\name{install_plinks}
\alias{install_plinks}
\title{Install multiple versions of PLINK}
\usage{
install_plinks(plink_optionses = create_plink_optionses(), os = get_os())
}
\arguments{
\item{os}{name of the operating system,
as returned by \link[rappdirs]{app_dir}}
}
\description{
Install multiple versions of PLINK
}
\author{
Richèl J.C. Bilderbeek
}
