% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_assoc_qt_data.R
\name{is_assoc_qt_data}
\alias{is_assoc_qt_data}
\title{Determine if the \code{assoc_qt_data} is indeed data used for a
quantitative trait analysis (as done by \link{assoc_qt})}
\usage{
is_assoc_qt_data(assoc_qt_data)
}
\arguments{
\item{assoc_qt_data}{data use to used by \link{assoc_qt}
(as created by \link{create_assoc_qt_data}),
which equals 'regular' data (see \link{check_data}) and a
phenotype table/file}
}
\value{
\link{TRUE} if the \code{assoc_qt_data} is indeed data used for a
quantitative trait analysis (as done by \link{assoc_qt})
}
\description{
Determine if the \code{assoc_qt_data} is indeed data used for a
quantitative trait analysis (as done by \link{assoc_qt})
}
\examples{
assoc_qt_data <- create_test_assoc_qt_data()
is_assoc_qt_data(assoc_qt_data)
}
\seealso{
\code{plinkr} has multiple functions to test if a type of data is valid:
\itemize{
\item \code{PLINK} text data: use \link{is_plink_text_data}
\item \code{PLINK} binary data: use \link{is_plink_bin_data}
\item \code{PLINK2} binary data: use \link{is_plink2_bin_data}
\item any type of \code{PLINK}/\code{PLINK2} data: use \link{is_data}
\item associative trait analysis data: use \link{is_assoc_qt_data}
}
}
\author{
Richèl J.C. Bilderbeek
}
