% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_bim_table.R
\name{is_bim_table}
\alias{is_bim_table}
\title{Determine if a \code{.bim} table is valid.}
\usage{
is_bim_table(bim_table, verbose = FALSE)
}
\arguments{
\item{bim_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{chr}: the chromosome number
\item \code{id}: the SNP ID
\item \code{posg}: the position
\item \code{pos}: the position
\item \code{ref}: something
\item \code{alt}: something
}

Use \link{get_test_bim_table} to get a \code{.bim} table as used in testing.
Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.
Use \link{check_bim_table} to test if a \code{.bim} table is valid.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the table is a valid \code{.bim} table
}
\description{
Determine if a \code{.bim} table is valid.
}
\examples{
is_bim_table(get_test_bim_table())
is_bim_table("nonsense")
}
\seealso{
use \link{check_bim_table} to check if a table
is a valid \code{.bim} table (i.e. \link{stop} if it is not)
}
\author{
Richèl J.C. Bilderbeek
}
