% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_chromosome_selector.R
\name{is_chromosome_selector}
\alias{is_chromosome_selector}
\title{Determine if this is a chromosome selector}
\usage{
is_chromosome_selector(chromosome_selector, verbose = FALSE)
}
\arguments{
\item{chromosome_selector}{a SNP selector (see \link{create_snp_selector})
that allows one to select a single chromsome,
as created by \link{create_chromosome_selector}
and checked by \link{check_chromosome_selector}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if this is a chromosome selector
}
\description{
Determine if this is a chromosome selector
}
\examples{
is_chromosome_selector(create_test_chromosome_selector())

is_chromosome_selector("nonsense")
}
\author{
Richèl J.C. Bilderbeek
}
