% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data.R
\name{is_data}
\alias{is_data}
\title{Determine if the \code{data} is indeed \code{PLINK} bin data}
\usage{
is_data(data, verbose = FALSE)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the \code{data} is indeed \code{PLINK} binary data
}
\description{
Determine if the \code{data} is indeed \code{PLINK} bin data
}
\examples{
data <- create_data(
  bed_table = get_test_bed_table(),
  bim_table = get_test_bim_table(),
  fam_table = get_test_fam_table()
)
is_data(data)
}
\seealso{
\code{plinkr} has multiple functions to test if a type of data is valid:
\itemize{
\item \code{PLINK} text data: use \link{is_plink_text_data}
\item \code{PLINK} binary data: use \link{is_plink_bin_data}
\item \code{PLINK2} binary data: use \link{is_plink2_bin_data}
\item any type of \code{PLINK}/\code{PLINK2} data: use \link{is_data}
\item associative trait analysis data: use \link{is_assoc_qt_data}
}
}
\author{
Richèl J.C. Bilderbeek
}
