% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_exe.R
\name{is_exe}
\alias{is_exe}
\title{Determine if a file can be executed}
\usage{
is_exe(filename)
}
\arguments{
\item{filename}{path to a filename}
}
\value{
TRUE if the file can be executed, FALSE otherwise
}
\description{
Determine if a file can be executed
}
