% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_fam_table.R
\name{is_fam_table}
\alias{is_fam_table}
\title{Determine if a \code{.fam} table is valid.}
\usage{
is_fam_table(fam_table, verbose = FALSE)
}
\arguments{
\item{fam_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{fam}: the family ID
\item \code{id}: the individual's ID
\item \code{pat}: ID of father
\item \code{mat}: ID of mother
\item \code{sex}: the gender
\item \code{pheno}: a phenotype
}

Use \link{get_test_fam_table} to get a \code{.fam} table as used in testing.
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.
Use \link{check_fam_table} to test if a \code{.fam} table is valid.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the table is a valid \code{.fam} table
}
\description{
Determine if a \code{.fam} table is valid.
}
\examples{
is_fam_table(get_test_fam_table())
is_fam_table(read_plink_fam_file(get_plinkr_filename("toy_v1_9.fam")))

is_fam_table("nonsense")
}
\seealso{
use \link{check_fam_table} to check if a table
is a valid \code{.fam} table (i.e. \link{stop} if it is not)
}
\author{
Richèl J.C. Bilderbeek
}
