% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_map_table.R
\name{is_map_table}
\alias{is_map_table}
\title{Determine if a \code{.map} table is valid.}
\usage{
is_map_table(map_table, verbose = FALSE)
}
\arguments{
\item{map_table}{a genetic mapping table.

A \code{map_table} is a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{CHR}: the chromosome code or contig name.
The chromosome number cannot exceed 95
(\url{https://www.cog-genomics.org/plink/1.9/input#chr_set})
\item \code{SNP}: Variant identifier
\item \code{position_cm}: Position in morgans or centimorgans.
This value is optional. Zeroes denote it is unused
\item \code{BP}: Base-pair coordinat
}

Use \link{read_plink_map_file} to read a \code{PLINK} \code{.map} file.
Use \link{check_map_table} to test if a genetic mapping table is valid.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the table is a valid \code{.map} table
}
\description{
Determine if a \code{.map} table is valid.
}
\examples{
is_map_table(get_test_map_table())
is_map_table(read_plink_map_file(get_plinkr_filename("toy_v1_9.map")))

is_map_table("nonsense")
}
\seealso{
use \link{check_map_table} to check if a table
is a valid \code{.map} table (i.e. \link{stop} if it is not)
}
\author{
Richèl J.C. Bilderbeek
}
