% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_marees_et_al_data_installed.R
\name{is_marees_et_al_data_installed}
\alias{is_marees_et_al_data_installed}
\title{Determine if the tutorial data of Marees et al., 2018 is installed.}
\usage{
is_marees_et_al_data_installed(plink_folder = get_plink_folder())
}
\arguments{
\item{plink_folder}{folder where \code{PLINK} is installed}
}
\value{
\code{TRUE} if the tutorial data of Marees et al., 2018
is installed, \code{FALSE} otherwise
}
\description{
Determine if the tutorial data of Marees et al., 2018 is installed.
}
\details{
\itemize{
\item Paper: \url{https://onlinelibrary.wiley.com/doi/10.1002/mpr.1608}
\item Code: \url{https://github.com/MareesAT/GWA_tutorial}
}

The code is re-done using plinkr in the vignette 'marees_et_al_2018'
}
\examples{
is_marees_et_al_data_installed()
is_marees_et_al_data_installed(plink_folder = "absent_folder")
}
\author{
Richèl J.C. Bilderbeek
}
