% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_ci.R
\name{is_on_ci}
\alias{is_on_ci}
\title{Determines if the environment is a continuous integration service}
\usage{
is_on_ci()
}
\value{
\link{TRUE} if run on AppVeyor or Travis CI, \link{FALSE} otherwise
}
\description{
Determines if the environment is a continuous integration service
}
\note{
It is possible to fake being on continuous integration service,
in this case GitHub Actions, using:\if{html}{\out{<div class="sourceCode r">}}\preformatted{Sys.setenv(GITHUB_ACTIONS = "I fake being on GitHub Actions")
is_on_ci() # Will be true
}\if{html}{\out{</div>}}

To undo this, do\if{html}{\out{<div class="sourceCode r">}}\preformatted{Sys.setenv(GITHUB_ACTIONS = "")
is_on_ci() # Will be false
}\if{html}{\out{</div>}}
}
\examples{
is_on_ci()
}
\author{
Richèl J.C. Bilderbeek
}
