% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_one_trait.R
\name{is_one_trait}
\alias{is_one_trait}
\title{Determine if the input is one trait}
\usage{
is_one_trait(trait)
}
\arguments{
\item{trait}{one trait with a clear genetic architecture and a known
minor allele frequency, as created by \link{create_trait}.
Use \link{is_one_trait} to detect if something is one trait}
}
\description{
Determine if the input is one trait.
Not a list of one trait.
Not a list of two traits.
One trait exactly.
}
\examples{
# Yes, this is one trait
is_one_trait(create_additive_trait())
is_one_trait(create_random_trait())

# No, these are not one trait
is_one_trait(list(create_demo_traits()))
is_one_trait(create_demo_traits())
is_one_trait("nonsense")
is_one_trait("")
is_one_trait(3.14)
is_one_trait(42)
is_one_trait(NA)
is_one_trait(NULL)
is_one_trait(Inf)
is_one_trait(c())
}
\author{
Richèl J.C. Bilderbeek
}
