% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ped_table.R
\name{is_ped_table}
\alias{is_ped_table}
\title{Determine if a \code{.ped} table is valid.}
\usage{
is_ped_table(ped_table, verbose = FALSE)
}
\arguments{
\item{ped_table}{a 'pedigree' table.

A \code{ped_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item \code{within_family_id_father} Within-family ID of father
(\code{0} if father isn't in dataset)
\item \code{within_family_id_mother} Within-family ID of mother
(\code{0} if mother isn't in dataset)
\item \code{sex_code} Sex code
(\code{1} = male, \code{2} = female, \code{0} = unknown)
\item \code{case_control_code} Case control code
(\code{1} = control, \code{2} = case,
\code{9}/\code{0}/non-numeric = missing data if case/control)
\item \code{snv_[x][y]} Nucleotide for the \code{x}th variant
for haplotype \code{y} (\code{y} is either \code{a} or \code{b})
in the \code{.map file} (\code{0} = no call)
}

The \code{FID} and \code{IID} column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}.

Use \link{read_plink_ped_file} to read a \code{PLINK} \code{.ped} file.
Use \link{check_ped_table} to test if a pedigree table is valid.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the table is a valid \code{.ped} table
}
\description{
Determine if a \code{.ped} table is valid.
}
\examples{
is_ped_table(get_test_ped_table())
is_ped_table(read_plink_ped_file(get_plinkr_filename("toy_v1_9.ped")))

is_ped_table("nonsense")
}
\seealso{
use \link{check_ped_table} to check if a table
is a valid \code{.ped} table (i.e. \link{stop} if it is not)
}
\author{
Richèl J.C. Bilderbeek
}
