% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pgen_table.R
\name{is_pgen_table}
\alias{is_pgen_table}
\title{Determine if a \code{.pgen} table is valid.}
\usage{
is_pgen_table(pgen_table, verbose = FALSE)
}
\arguments{
\item{pgen_table}{an \link{array} that maps the individuals
to their SNPs, with as much rows as individuals, and as much
SNPs as columns. Optionally, the row names are the individuals' IDs,
where the column names are the SNP ID's}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the table is a valid \code{.pgen} table
}
\description{
Determine if a \code{.pgen} table is valid.
}
\examples{
is_pgen_table(get_test_pgen_table())

is_pgen_table("nonsense")
}
\seealso{
use \link{check_pgen_table} to check if a table
is a valid \code{.pgen} table (i.e. \link{stop} if it is not)
}
\author{
Richèl J.C. Bilderbeek
}
