% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_phenotype_data_table.R
\name{is_phenotype_data_table}
\alias{is_phenotype_data_table}
\title{Determine if the \code{phenotype_data_table} is
indeed phenotypic data as a \link[tibble]{tibble}}
\usage{
is_phenotype_data_table(phenotype_data_table, verbose = FALSE)
}
\arguments{
\item{phenotype_data_table}{phenotype data
(as checked by \link{check_phenotype_data}))
in the form of a \link{list} with an in-memory phenotype table}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if if the \code{phenotype_data_table} is
indeed phenotypic data as a \link[tibble]{tibble}
}
\description{
Determine if the \code{phenotype_data_table} is
indeed phenotypic data as a \link[tibble]{tibble}
}
\author{
Richèl J.C. Bilderbeek
}
