% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plink2_bin_data.R
\name{is_plink2_bin_data}
\alias{is_plink2_bin_data}
\title{Determine if the \code{plink2_bin_data} is indeed \code{PLINK2} binary data}
\usage{
is_plink2_bin_data(plink2_bin_data)
}
\arguments{
\item{plink2_bin_data}{the genetic and phenotypic data for \code{PLINK2} to work
on, in \code{PLINK2} binary format, as created by \link{create_plink2_bin_data}.}
}
\description{
Determine if the \code{plink2_bin_data} is indeed \code{PLINK2} binary data
}
\author{
Richèl J.C. Bilderbeek
}
