% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plink_bin_data.R
\name{is_plink_bin_data}
\alias{is_plink_bin_data}
\title{Determine if the \code{plink_bin_data} is indeed \code{PLINK} bin data}
\usage{
is_plink_bin_data(plink_bin_data, verbose = FALSE)
}
\arguments{
\item{plink_bin_data}{the genetic and phenotypic data for \code{PLINK} to work
on, in \code{PLINK} binary format, as created by \link{create_plink_bin_data}.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the \code{plink_bin_data} is indeed \code{PLINK} binary data
}
\description{
Determine if the \code{plink_bin_data} is indeed \code{PLINK} bin data
}
\examples{
plink_bin_data <- create_plink_bin_data(
  bim_table = get_test_bim_table(),
  fam_table = get_test_fam_table(),
  bed_table = get_test_bed_table()
)
is_plink_bin_data(plink_bin_data)
}
\seealso{
\code{plinkr} has multiple functions to test if a type of data is valid:
\itemize{
\item \code{PLINK} text data: use \link{is_plink_text_data}
\item \code{PLINK} binary data: use \link{is_plink_bin_data}
\item \code{PLINK2} binary data: use \link{is_plink2_bin_data}
}
}
\author{
Richèl J.C. Bilderbeek
}
