% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plink_installed.R
\name{is_plink_installed}
\alias{is_plink_installed}
\title{Determine if PLINK is installed}
\usage{
is_plink_installed(plink_folder = get_plink_folder())
}
\value{
TRUE if PLINK is installed, FALSE otherwise
}
\description{
Determine if PLINK is installed
}
