% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plink_tutorial_data_installed.R
\name{is_plink_tutorial_data_installed}
\alias{is_plink_tutorial_data_installed}
\title{Determine if the \code{PLINK} tutorial data is installed.}
\usage{
is_plink_tutorial_data_installed()
}
\value{
\code{TRUE} if the \code{PLINK} tutorial data
is installed, \code{FALSE} otherwise
}
\description{
Determine if the \code{PLINK} tutorial data is installed.
The \code{PLINK} tutorial can be found at
\url{http://zzz.bwh.harvard.edu/plink/tutorial.shtml}
and is done in the \code{plink_tutorial} vignette
}
\author{
Richèl J.C. Bilderbeek
}
