% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_psam_table.R
\name{is_psam_table}
\alias{is_psam_table}
\title{Determine if a \code{.pvar} table is valid.}
\usage{
is_psam_table(psam_table, verbose = FALSE)
}
\arguments{
\item{psam_table}{a \link[tibble]{tibble} with as much rows as
individuals. It has the following columns:
\itemize{
\item \code{FID}: the family ID
\item \code{IID}: the within-family ID
\item \code{SEX}: the sex
\item \code{PHENO1}: a phenotypic value
}

The uppercase column names are those as used by PLINK2.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the table is a valid \code{.pvar} table
}
\description{
Determine if a \code{.pvar} table is valid.
}
\examples{
is_psam_table(get_test_psam_table())

is_psam_table("nonsense")
}
\seealso{
use \link{check_psam_table} to check if a table
is a valid \code{.pvar} table (i.e. \link{stop} if it is not)
}
\author{
Richèl J.C. Bilderbeek
}
