% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pvar_table.R
\name{is_pvar_table}
\alias{is_pvar_table}
\title{Determine if a \code{.pvar} table is valid.}
\usage{
is_pvar_table(pvar_table, verbose = FALSE)
}
\arguments{
\item{pvar_table}{a \link[tibble]{tibble} with as much rows as
SNPs. It has the following columns:
\itemize{
\item \code{CHROM}: the chromosome
\item \code{POS}: the position on the chromosome
\item \code{ID}: the SNP ID
\item \code{REF}: the reference variant
\item \code{ALT}: the alternate variant
}

The uppercase column names are those as used by PLINK2.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the table is a valid \code{.pvar} table
}
\description{
Determine if a \code{.pvar} table is valid.
}
\examples{
is_pvar_table(get_test_pvar_table())

is_pvar_table("nonsense")
}
\seealso{
use \link{check_pvar_table} to check if a table
is a valid \code{.pvar} table (i.e. \link{stop} if it is not)
}
\author{
Richèl J.C. Bilderbeek
}
