% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_random_samples_selector.R
\name{is_random_samples_selector}
\alias{is_random_samples_selector}
\title{Determine if this is a random sample selector}
\usage{
is_random_samples_selector(random_samples_selector, verbose = FALSE)
}
\arguments{
\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if this is a random sample selector
}
\description{
Determine if this is a random sample selector
}
\examples{
is_random_samples_selector(create_test_random_samples_selector())

is_random_samples_selector("nonsense")
}
\author{
Richèl J.C. Bilderbeek
}
