% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sim_qt_params.R
\name{is_sim_qt_params}
\alias{is_sim_qt_params}
\title{Determine if the input is a valid set of parameters to
simulate a quantitative trait}
\usage{
is_sim_qt_params(sim_qt_params)
}
\arguments{
\item{sim_qt_params}{the parameters for a quantitative
traits simulation, as can be created by
\link{create_sim_qt_params}}
}
\description{
Determine if the input is a valid set of parameters to
simulate a quantitative trait
}
