% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sim_qtl_params.R
\name{is_sim_qtl_params}
\alias{is_sim_qtl_params}
\title{Determine if the input is a valid set of parameters to
simulate a quantitative trait}
\usage{
is_sim_qtl_params(sim_qtl_params)
}
\description{
Determine if the input is a valid set of parameters to
simulate a quantitative trait
}
