% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_single_sample_selector.R
\name{is_single_sample_selector}
\alias{is_single_sample_selector}
\title{Determine if this is a single sample selector}
\usage{
is_single_sample_selector(single_sample_selector, verbose = FALSE)
}
\arguments{
\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if this is a single sample selector
}
\description{
Determine if this is a single sample selector
}
\examples{
is_single_sample_selector(create_test_single_sample_selector())

is_single_sample_selector("nonsense")
}
\author{
Richèl J.C. Bilderbeek
}
