% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_single_snp_selector.R
\name{is_single_snp_selector}
\alias{is_single_snp_selector}
\title{Determine if this is a single SNP selector}
\usage{
is_single_snp_selector(single_snp_selector, verbose = FALSE)
}
\arguments{
\item{single_snp_selector}{a SNP selector (see \link{check_snp_selector})
that allows one to select a single SNP,
as created by \link{create_single_snp_selector}
and checked by \link{check_single_snp_selector}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if this is a single SNP selector
}
\description{
Determine if this is a single SNP selector
}
\examples{
is_single_snp_selector(get_test_single_snp_selector())

is_single_snp_selector("nonsense")
}
\author{
Richèl J.C. Bilderbeek
}
