% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_snp_range_selector.R
\name{is_snp_range_selector}
\alias{is_snp_range_selector}
\title{Determine if this is a SNP range selector}
\usage{
is_snp_range_selector(snp_range_selector, verbose = FALSE)
}
\arguments{
\item{snp_range_selector}{a SNP selector (see \link{create_snps_selector}),
that allows one to select a range between two SNPs,
as created by \link{create_snp_range_selector}
and checked by \link{check_snp_range_selector}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if this is a SNP range selector
}
\description{
Determine if this is a SNP range selector
}
\examples{
is_snp_range_selector(create_test_snp_range_selector())

is_snp_range_selector("nonsense")
}
\author{
Richèl J.C. Bilderbeek
}
