% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_snp_selector.R
\name{is_snp_selector}
\alias{is_snp_selector}
\title{Determine if this is a SNP selector}
\usage{
is_snp_selector(snp_selector, verbose = FALSE)
}
\arguments{
\item{snp_selector}{a SNP selector, a way to select one or more SNPs.
See \link{create_snp_selector} for all SNP selectors.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if this is a SNP selector
}
\description{
Determine if this is a SNP selector
}
\examples{
is_snp_selector(create_test_random_snp_selector())
is_snp_selector(create_test_single_snp_selector())
is_snp_selector(create_test_snp_range_selector())
is_snp_selector(create_test_snp_window_selector())

is_snp_selector("nonsense")
}
\author{
Richèl J.C. Bilderbeek
}
