% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_snp_window_selector.R
\name{is_snp_window_selector}
\alias{is_snp_window_selector}
\title{Determine if this is a SNP window selector}
\usage{
is_snp_window_selector(snp_window_selector, verbose = FALSE)
}
\arguments{
\item{snp_window_selector}{a SNP selector (see \link{check_snp_selector}),
that allows one to select a focal SNP and a range/window of SNPs around it,
as created by \link{create_snp_window_selector}
and checked by \link{check_snp_window_selector}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if this is a SNP window selector
}
\description{
Determine if this is a SNP window selector
}
\examples{
is_snp_window_selector(create_test_snp_window_selector())

is_snp_window_selector("nonsense")
}
\author{
Richèl J.C. Bilderbeek
}
