% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bed.R
\name{make_bed}
\alias{make_bed}
\title{Convert to binary files}
\usage{
make_bed(
  base_input_filename,
  base_output_filename = file.path(dirname(base_input_filename), "output"),
  plink_options = create_plink_v1_9_options(),
  verbose = FALSE
)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
The full names of the created files
}
\description{
Convert to binary files
}
\note{
this function is named after the PLINK \code{--make-bed}
parameter flag
}
\author{
Richèl J.C. Bilderbeek
}
