% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bed_with_plink2.R
\name{make_bed_with_plink2}
\alias{make_bed_with_plink2}
\title{Convert \code{PLINK2} binary files to \code{PLINK} binary files using \code{PLINK2}}
\usage{
make_bed_with_plink2(
  base_input_filename,
  base_output_filename = file.path(dirname(base_input_filename), "output"),
  plink_options = create_plink_v2_0_options(),
  verbose = FALSE
)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
The full names of the created \code{PLINK} binary files,
which are a \code{.bed}, \code{.bim}, \code{.fam}
and \code{.log} file.
}
\description{
Convert \code{PLINK2} binary files to \code{PLINK} binary files using \code{PLINK2}
}
\note{
this function is named after the \code{PLINK}/\code{PLINK2} \code{--make-bed}
parameter flag
}
\author{
Richèl J.C. Bilderbeek
}
