% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pgen.R
\name{make_pgen}
\alias{make_pgen}
\title{Convert to \code{PLINK2} binary files}
\usage{
make_pgen(
  base_input_filename,
  base_output_filename = file.path(dirname(base_input_filename), "output"),
  plink_options = create_plink_v2_0_options(),
  verbose = FALSE
)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
The full names of the created \code{PLINK2} binary files,
which have the following four file extensions:
\itemize{
\item \code{.log}
\item \code{.pgen}
\item \code{.psam}
\item \code{.pvar}
}
}
\description{
Convert to \code{PLINK2} binary files
}
\note{
this function is named after the \code{PLINK} \code{--make-pgen}
parameter flag
}
\author{
Richèl J.C. Bilderbeek
}
