% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{missing}
\alias{missing}
\title{Let \code{PLINK}/\code{PLINK2}
calculate the missing rates summary statistic.}
\usage{
missing(bfile, out)
}
\value{
a \link[tibble]{tibble}
}
\description{
Let \code{PLINK}/\code{PLINK2}
calculate the missing rates summary statistic.
}
\note{
This function is named after the \code{--missing} flag
used by \code{PLINK}
}
\author{
Richèl J.C. Bilderbeek
}
