% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_from_bfiles.R
\name{missing_from_bfile}
\alias{missing_from_bfile}
\title{Let \code{PLINK}/\code{PLINK2}
calculate the missing rates summary statistic.}
\usage{
missing_from_bfile(bfile, out, plink_options = create_plink_options())
}
\arguments{
\item{bfile}{the base filename of the binary files (i.e.
a \code{.bed}, \code{.bim} and \code{.fam} file).
This parameter is named after the \code{PLINK}
\code{--bfile} flag}

\item{out}{the base filename of the output files.
This parameter is named after the \code{PLINK}
\code{--out} flag}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
a \link{list} with the following elements:
\itemize{
\item \code{imiss_table} table with the missing genotype data per individual
\item \code{lmiss_table} table with the missing genotype data per SNP
\item \code{log} the log text
}
}
\description{
Let \code{PLINK}/\code{PLINK2}
calculate the missing rates summary statistic
from PLINK1 binary filenames
}
\note{
This function is named after the \code{--missing} flag
used by \code{PLINK}
}
\author{
Richèl J.C. Bilderbeek
}
