% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_from_bfiles.R
\name{missing_from_bfiles}
\alias{missing_from_bfiles}
\title{Let \code{PLINK}/\code{PLINK2}
calculate the missing rates summary statistic.}
\usage{
missing_from_bfiles(bfile, out)
}
\value{
a \link[tibble]{tibble}
}
\description{
Let \code{PLINK}/\code{PLINK2}
calculate the missing rates summary statistic
from PLINK1 binary filenames
}
\note{
This function is named after the \code{--missing} flag
used by \code{PLINK}
}
\author{
Richèl J.C. Bilderbeek
}
