% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Do a principal component analysis}
\usage{
pca(
  base_input_filename,
  base_output_filename,
  plink_options = create_plink_v1_9_options(),
  verbose = FALSE
)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a list of filenames that are created.

The list has elements:
\itemize{
\item \code{bed_filename}: the \code{.bed} filename,
as can be read by \link{read_plink_bed_file}
\item \code{bim_filename}: the \code{.bim} filename,
as can be read by \link{read_plink_bim_file}
\item \code{eigenval_filename}: the \code{.eigenval} filename,
as can be read by \link{read_plink_eigenval_file}
\item \code{eigenvec_filename}: the \code{.eigenvec} filename,
as can be read by \link{read_plink_eigenvec_file}
\item \code{fam_filename}: the \code{.fam} filename,
as can be read by \link{read_plink_fam_file}
\item \code{log_filename}: the \code{.log} filename,
as can be read by \link{read_plink_log_file}
}
}
\description{
Do a principal component analysis
}
\note{
this function is named after the \code{PLINK}/\code{PLINK2} \code{--pca}
parameter flag
}
\author{
Richèl J.C. Bilderbeek
}
