% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkr_report.R
\name{plinkr_report}
\alias{plinkr_report}
\title{Create a \link{plinkr} report, to be used when reporting bugs}
\usage{
plinkr_report(plink_optionses = plinkr::create_plink_optionses())
}
\arguments{
\item{plink_optionses}{a list of one or more options to run PLINK,
as created by \link{create_plink_optionses}.
The reduplicated plural was used to express this is a list
of \code{plink_options}, instead of one set of \code{plink_options}}
}
\description{
Create a \link{plinkr} report, to be used when reporting bugs
}
\examples{
plinkr_report()
}
\author{
Richèl J.C. Bilderbeek
}
