% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca_results.R
\name{plot_pca_results}
\alias{plot_pca_results}
\title{Plot the result of a call to \link{pca}}
\usage{
plot_pca_results(pca_result_filenames)
}
\value{
a \link[ggplot2]{ggplot2}
}
\description{
Plot the result of a call to \link{pca}.
}
\details{
This will show the first two components of a principal
component analysis.
The x axis shows the first principal component,
The y axis shows the second.
The axes' labels show how percent of the variants is explained
by that principal component
}
\examples{
plot_pca_results(create_test_pca_result_filenames())
}
\author{
Richèl J.C. Bilderbeek
}
