% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink2_bin_data.R
\name{read_plink2_bin_data}
\alias{read_plink2_bin_data}
\title{Read \code{PLINK2} binary files based on a shared base input filename}
\usage{
read_plink2_bin_data(base_input_filename, verbose = FALSE)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Read \code{PLINK2} binary files based on a shared base input filename
}
\author{
Richèl J.C. Bilderbeek
}
