% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink2_pgen_file.R
\name{read_plink2_pgen_file}
\alias{read_plink2_pgen_file}
\title{Read a \code{PLINK2} \code{.pgen} file}
\usage{
read_plink2_pgen_file(pgen_filename, names_loci, names_ind, verbose = FALSE)
}
\arguments{
\item{pgen_filename}{name of a \code{PLINK2} \code{.pgen} file
Use \link{read_plink2_pgen_file} to read a \code{PLINK2} \code{.pgen} file}

\item{names_loci}{see \link[genio]{read_bed}}

\item{names_ind}{see \link[genio]{read_bed}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{.pgen} table
}
\description{
Read a \code{PLINK2} \code{.pgen} file
}
\examples{
pgen_filename <- get_plinkr_filename(
  "test_v1_7_after_make-bed_after_make-pgen.pgen"
)
read_plink2_pgen_file(
  pgen_filename = pgen_filename,
  names_loci = paste0("snp_", 1:2),
  names_ind = paste0("person_", letters[1:6])
)
}
\seealso{
use \link{read_plink2_pgen_file_from_files} to read
a \code{PLINK2} \code{.pgen} file with column and rownames obtained from
the two other \code{PLINK2} data files
}
\author{
Richèl J.C. Bilderbeek
}
