% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink2_pgen_file_from_files.R
\name{read_plink2_pgen_file_from_files}
\alias{read_plink2_pgen_file_from_files}
\title{Read a \code{PLINK}/\code{PLINK2} \code{.bed} file
using a \code{.bim} and \code{.fam} file.}
\usage{
read_plink2_pgen_file_from_files(
  pgen_filename,
  psam_filename,
  pvar_filename,
  verbose = FALSE
)
}
\arguments{
\item{pgen_filename}{name of a \code{PLINK2} \code{.pgen} file
Use \link{read_plink2_pgen_file} to read a \code{PLINK2} \code{.pgen} file}

\item{psam_filename}{name of a \code{PLINK2} \code{.psam} file
Use \link{read_plink2_psam_file} to read a \code{PLINK2} \code{.psam} file}

\item{pvar_filename}{name of a \code{PLINK2} \code{.pvar} file
Use \link{read_plink2_pvar_file} to read a \code{PLINK2} \code{.pvar} file}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Read a \code{PLINK}/\code{PLINK2} \code{.bed} file
using a \code{.bim} and \code{.fam} file.
These two accessory files are used to extract the names
of the individuals and the names of the loci.
Use \link{read_plink_pgen_file} to read a \code{.bed} file
without those two accessory files.
}
\examples{
read_plink_pgen_file_from_files(
  pgen_filename = get_plinkr_filename("demo_assoc.bed"),
  psam_filename = get_plinkr_filename("demo_assoc.bim"),
  pvar_filename = get_plinkr_filename("demo_assoc.fam")
)
}
\author{
Richèl J.C. Bilderbeek
}
