% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink2_psam_file.R
\name{read_plink2_psam_file}
\alias{read_plink2_psam_file}
\title{Read a \code{PLINK2} \code{.psam} file}
\usage{
read_plink2_psam_file(psam_filename)
}
\arguments{
\item{psam_filename}{name of a \code{PLINK2} \code{.psam} file
Use \link{read_plink2_psam_file} to read a \code{PLINK2} \code{.psam} file}
}
\description{
Read a \code{PLINK2} \code{.psam} file
}
\author{
Richèl J.C. Bilderbeek
}
