% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink2_pvar_file.R
\name{read_plink2_pvar_file}
\alias{read_plink2_pvar_file}
\title{Read a \code{PLINK2} \code{.pvar} file}
\usage{
read_plink2_pvar_file(pvar_filename)
}
\arguments{
\item{pvar_filename}{name of a \code{PLINK2} \code{.pvar} file
Use \link{read_plink2_pvar_file} to read a \code{PLINK2} \code{.pvar} file}
}
\description{
Read a \code{PLINK2} \code{.pvar} file
}
\note{
this function is inspired by
\url{https://github.com/simingz/ctwas/blob/main/R/ctwas_read_data.R#L62}
}
\author{
Richèl J.C. Bilderbeek
}
