% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink2_qassoc_file.R
\name{read_plink2_qassoc_file}
\alias{read_plink2_qassoc_file}
\title{Read a \code{PLINK2} \link{assoc_qt} file}
\usage{
read_plink2_qassoc_file(qassoc_filename)
}
\arguments{
\item{qassoc_filename}{name of a \code{PLINK} \code{.qassoc} file
Use \link{read_plink_qassoc_file} to read a \code{PLINK} \code{.qassoc} file.}
}
\value{
a tibble
}
\description{
Read a \code{PLINK2} \link{assoc_qt} file
}
\examples{
read_plink2_qassoc_file(
  qassoc_filename = get_plinkr_filename("run1.qassoc")
)
}
\author{
Richèl J.C. Bilderbeek
}
