% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_assoc_file.R
\name{read_plink_assoc_file}
\alias{read_plink_assoc_file}
\title{Read a PLINK \code{.assoc} file}
\usage{
read_plink_assoc_file(assoc_filename)
}
\arguments{
\item{assoc_filename}{name of a PLINK \code{.assoc} file
Use \link{read_plink_assoc_file} to read a PLINK \code{.assoc} file.}
}
\value{
a tibble
}
\description{
Read a PLINK \code{.assoc} file
}
\examples{
read_plink_assoc_file(
  assoc_filename = get_plinkr_filename("run1.assoc")
)
}
