% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_bed_file.R
\name{read_plink_bed_file}
\alias{read_plink_bed_file}
\title{Read a \code{PLINK}/\code{PLINK2} \code{.bed} file}
\usage{
read_plink_bed_file(bed_filename, names_loci, names_ind, verbose = FALSE)
}
\arguments{
\item{bed_filename}{name of a \code{PLINK} \code{.bed} file
Use \link{read_plink_bed_file} to read a \code{PLINK} \code{.bed} file.}

\item{names_loci}{see \link[genio]{read_bed}}

\item{names_ind}{see \link[genio]{read_bed}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Read a \code{PLINK}/\code{PLINK2} \code{.bed} file
}
\author{
Richèl J.C. Bilderbeek
}
