% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_bed_file_from_files.R
\name{read_plink_bed_file_from_files}
\alias{read_plink_bed_file_from_files}
\title{Read a \code{PLINK}/\code{PLINK2} \code{.bed} file
using a \code{.bim} and \code{.fam} file.}
\usage{
read_plink_bed_file_from_files(
  bed_filename,
  bim_filename,
  fam_filename,
  verbose = FALSE
)
}
\arguments{
\item{bed_filename}{name of a \code{PLINK} \code{.bed} file
Use \link{read_plink_bed_file} to read a \code{PLINK} \code{.bed} file.}

\item{bim_filename}{name of a \code{PLINK} \code{.bim} file
Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.}

\item{fam_filename}{name of a \code{PLINK} \code{.fam} file
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Read a \code{PLINK}/\code{PLINK2} \code{.bed} file
using a \code{.bim} and \code{.fam} file.
These two accessory files are used to extract the names
of the individuals and the names of the loci.
Use \link{read_plink_bed_file} to read a \code{.bed} file
without those two accessory files.
}
\examples{
read_plink_bed_file_from_files(
  bed_filename = get_plinkr_filename("demo_assoc.bed"),
  bim_filename = get_plinkr_filename("demo_assoc.bim"),
  fam_filename = get_plinkr_filename("demo_assoc.fam")
)
}
\author{
Richèl J.C. Bilderbeek
}
