% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_bim_file.R
\name{read_plink_bim_file}
\alias{read_plink_bim_file}
\title{Read a \code{PLINK}/\code{PLINK2} \code{.bim} file}
\usage{
read_plink_bim_file(bim_filename, verbose = FALSE)
}
\arguments{
\item{bim_filename}{name of a \code{PLINK} \code{.bim} file
Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Read a \code{PLINK}/\code{PLINK2} \code{.bim} file
}
\author{
Richèl J.C. Bilderbeek
}
