% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_cov_file.R
\name{read_plink_cov_file}
\alias{read_plink_cov_file}
\title{Read a \code{PLINK} covariates file}
\usage{
read_plink_cov_file(cov_filename)
}
\arguments{
\item{cov_filename}{name of a covariates (\code{.cov}) file.
Use \link{read_plink_cov_file} to read a covariates file.}
}
\value{
a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item \code{Px} One or more columns with phenotypic values,
where \code{x} is the index of the trait:
the first column is called \code{P1}, the second \code{P2},
etcetera.
The \code{FID}, \code{IID} and \code{Px} column names
match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}
}
}
\description{
Read a \code{PLINK} covariates file
}
\note{
The filename extension used in the \code{PLINK}
tutorial is \code{.covar}
}
\author{
Richèl J.C. Bilderbeek
}
