% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_eigenval_file.R
\name{read_plink_eigenval_file}
\alias{read_plink_eigenval_file}
\title{Read a \code{PLINK}/\code{PLINK2} \code{.eigenval} file}
\usage{
read_plink_eigenval_file(eigenval_filename)
}
\value{
a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{eigenval}: the Eigenvalue
}
}
\description{
Read a \code{PLINK}/\code{PLINK2} \code{.eigenval} file.
}
\details{
By default, \link{pca} (the function to do principal
component analysis) uses 20 Eigenvalues
}
\examples{
read_plink_eigenval_file(
  eigenval_filename = get_plinkr_filename("pca.eigenval")
)
}
\seealso{
use \link{get_test_eigenval_table} to get an
Eigenvalue table, to be used in testing
}
\author{
Richèl J.C. Bilderbeek
}
