% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_fam_file.R
\name{read_plink_fam_file}
\alias{read_plink_fam_file}
\title{Read a \code{PLINK}/\code{PLINK2} \code{.fam} file}
\usage{
read_plink_fam_file(fam_filename, verbose = FALSE)
}
\arguments{
\item{fam_filename}{name of a \code{PLINK} \code{.fam} file
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Read a \code{PLINK}/\code{PLINK2} \code{.fam} file
}
\author{
Richèl J.C. Bilderbeek
}
