% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_frq_file.R
\name{read_plink_frq_file}
\alias{read_plink_frq_file}
\title{Read a PLINK \code{.frq} file}
\usage{
read_plink_frq_file(frq_filename)
}
\value{
a \link[tibble]{tibble} with column names:
\itemize{
\item \code{CHR} ?
\item \code{SNP} ?
\item \code{N_MISS} ?
\item \code{N_GENO} ?
\item \code{F_MISS} ?
}

These column names are the names used by \code{PLINK}.
}
\description{
Read a PLINK \code{.frq} file
}
\author{
Richèl J.C. Bilderbeek
}
