% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_frq_strat_file.R
\name{read_plink_frq_strat_file}
\alias{read_plink_frq_strat_file}
\title{Read a PLINK \code{.frq.strat} file}
\usage{
read_plink_frq_strat_file(frq_strat_filename)
}
\arguments{
\item{frq_strat_filename}{name of a \code{PLINK} \code{.frq.strat} file
Use \link{read_plink_frq_strat_file}
to read a \code{PLINK} \code{.frq.strat} file.}
}
\value{
a \link[tibble]{tibble} with column names:
\itemize{
\item \code{CHR} ?
\item \code{SNP} ?
\item \code{CLST} ?
\item \code{A1} ?
\item \code{A2} ?
\item \code{MAF} ?
\item \code{MAC} ?
\item \code{NCHROBS} ?
}

These column names are the names used by \code{PLINK}.
}
\description{
Read a PLINK \code{.frq.strat} file
}
\examples{
frq_strat_filename <- get_plinkr_filename("miss_stat.frq.strat")
read_plink_frq_strat_file(frq_strat_filename)
}
\author{
Richèl J.C. Bilderbeek
}
