% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_imiss_file.R
\name{read_plink_imiss_file}
\alias{read_plink_imiss_file}
\title{Read a PLINK \code{.imiss} file}
\usage{
read_plink_imiss_file(imiss_filename)
}
\arguments{
\item{imiss_filename}{name of a \code{PLINK} \code{.imiss} file
Use \link{read_plink_imiss_file} to read a \code{PLINK} \code{.imiss} file.}
}
\value{
a \link[tibble]{tibble} with column names:
\itemize{
\item \code{FIF} family ID
\item \code{IID} Within-family ID
\item \code{MISS_PHENO} ?
\item \code{N_MISS} ?
\item \code{N_GENO} ?
\item \code{F_MISS} ?
}

These column names are the names used by \code{PLINK}.
}
\description{
Read a PLINK \code{.imiss} file
}
\author{
Richèl J.C. Bilderbeek
}
