% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_ld_file.R
\name{read_plink_ld_file}
\alias{read_plink_ld_file}
\title{Read a PLINK \code{.ld} file}
\usage{
read_plink_ld_file(ld_filename)
}
\value{
a \link[tibble]{tibble} with column names:
\itemize{
\item \code{CHR_A}	Chromosome code for first variant
\item \code{BP_A}	Base-pair coordinate of first variant
\item \code{SNP_A}	ID of first variant
\item (\code{MAF_A}: Allele 1 frequency for first variant. Requires \code{with-freqs})
\item \code{CHR_B}	Chromosome code for second variant
\item \code{BP_B}	Base-pair coordinate of second variant
\item \code{SNP_B}	ID of second variant
\item (\code{PHASE}: In-phase allele pairs. Requires \verb{in-phase})
\item (\code{MAF_B}	Allele 1 frequency for second variant. Requires \code{with-freqs})
\item \code{R} or \code{R2}: for \code{R}: the correlation coefficient,
for \code{R2}, the squared correlation coefficient
}

These column names are the names used by \code{PLINK}, the description is
taken from the PLINK 1.9 File format reference at
\url{https://www.cog-genomics.org/plink/1.9/formats#ld}
}
\description{
Read a PLINK \code{.ld} file
}
\examples{
if (1 == 2) {
  ld_filename <- get_plinkr_filename("miss_stat.ld")
  read_plink_ld_file(ld_filename)
}
}
\author{
Richèl J.C. Bilderbeek
}
