% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_lmiss_file.R
\name{read_plink_lmiss_file}
\alias{read_plink_lmiss_file}
\title{Read a PLINK \code{.lmiss} file}
\usage{
read_plink_lmiss_file(lmiss_filename)
}
\arguments{
\item{lmiss_filename}{name of a \code{PLINK} \code{.lmiss} file
Use \link{read_plink_lmiss_file} to read a \code{PLINK} \code{.lmiss} file.}
}
\value{
a \link[tibble]{tibble} with column names:
\itemize{
\item \code{CHR} ?
\item \code{SNP} ?
\item \code{N_MISS} ?
\item \code{N_GENO} ?
\item \code{F_MISS} ?
}

These column names are the names used by \code{PLINK}.
}
\description{
Read a PLINK \code{.lmiss} file
}
\author{
Richèl J.C. Bilderbeek
}
