% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_log_file.R
\name{read_plink_log_file}
\alias{read_plink_log_file}
\title{Read a PLINK \code{.log} file}
\usage{
read_plink_log_file(log_filename)
}
\arguments{
\item{log_filename}{name of a \code{PLINK} \code{.log} file}
}
\description{
Read a PLINK \code{.log} file
}
\author{
Richèl J.C. Bilderbeek
}
