% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_map_file.R
\name{read_plink_map_file}
\alias{read_plink_map_file}
\title{Read a PLINK \code{.map} file}
\usage{
read_plink_map_file(map_filename)
}
\arguments{
\item{map_filename}{name of a PLINK \code{.map} file}
}
\value{
a [tibble]{tibble} with the following columns:
\itemize{
\item \code{chromosome_code}: the chromosome code or contig name
\item \code{variant_id}: Variant identifier
\item \code{position}: Position in morgans or centimorgans.
This value is options. Zeroes denote it is unused
\item \code{coordinat}: Base-pair coordinat
}
}
\description{
Read a variant information file
accompanying a \code{.ped} text pedigree +
genotype table.
}
