% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_phenotype_file.R
\name{read_plink_phe_file}
\alias{read_plink_phe_file}
\title{Read a \code{PLINK} raw phenotype file}
\usage{
read_plink_phe_file(phe_filename)
}
\arguments{
\item{phe_filename}{name of a raw phenotype file.
Use \link{read_plink_phe_file} to read a raw phenotype file.}
}
\value{
a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item \code{Px} One or more columns with phenotypic values,
where \code{x} is the index of the trait:
the first column is called \code{P1}, the second \code{P2},
etcetera.
The \code{FID}, \code{IID} and \code{Px} column names
match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}
}
}
\description{
Read a \code{PLINK} raw phenotype file.
}
\note{
The filename extension used in the \code{PLINK}
tutorial is \code{.phe}, where the \code{PLINK}
'Getting started' uses \code{.raw}
}
\examples{
read_plink_phe_file(
  phe_filename = get_plinkr_filename("pheno.raw")
)
}
\author{
Richèl J.C. Bilderbeek
}
