% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_phenotype_file.R
\name{read_plink_phenotype_file}
\alias{read_plink_phenotype_file}
\title{Read a PLINK raw phenotype file}
\usage{
read_plink_phenotype_file(phenotype_filename)
}
\arguments{
\item{phenotype_filename}{name of a raw phenotype file.
Use \link{read_plink_phenotype_file} to read a raw phenotype file.}
}
\value{
a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

The \code{FID} and \code{IID} column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}
}
\description{
Read a PLINK raw phenotype file
}
\examples{
read_plink_phenotype_file(
  phenotype_filename = get_plinkr_filename("pheno.raw")
)
}
\author{
Richèl J.C. Bilderbeek
}
