% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_qassoc_files.R
\name{read_plink_qassoc_files}
\alias{read_plink_qassoc_files}
\title{Read one or more \code{PLINK} \code{.qassoc} file}
\usage{
read_plink_qassoc_files(qassoc_filenames)
}
\arguments{
\item{qassoc_filenames}{name of one or more
\code{PLINK} \code{.qassoc} files.
Use \link{read_plink_qassoc_files} to read one or more
\code{PLINK} \code{.qassoc} files.}
}
\value{
a \link[tibble]{tibble}
}
\description{
Read one or more \code{PLINK} \code{.qassoc} file
}
\examples{
read_plink_qassoc_files(
  qassoc_filenames = get_plinkr_filename("run1.qassoc")
)
}
\author{
Richèl J.C. Bilderbeek
}
