% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_simfreq_file.R
\name{read_plink_simfreq_file}
\alias{read_plink_simfreq_file}
\title{Read a PLINK \code{.simfreq} file}
\usage{
read_plink_simfreq_file(simfreq_filename)
}
\arguments{
\item{simfreq_filename}{name of a \code{PLINK} \code{.simfreq} file}
}
\value{
a \link[tibble]{tibble} with column names:
\itemize{
\item \code{n_snps} Number of SNPs in set
(always 1 in autogenerated file)
\item \code{label} Label of this set of SNPs
\item \code{allele_frequency_lower_bound}
Reference allele frequency lower bound
\item \code{allele_frequency_upper_bound}
Reference allele frequency upper bound
(equal to lower bound in autogenerated file)
\item \code{additive_genetic_variance} Additive genetic variance for each SNP
\item \code{dominance_deviation} Dominance deviation
}
}
\description{
Read a PLINK \code{.simfreq} file
}
\seealso{
File format reference at
\url{https://www.cog-genomics.org/plink/1.9/formats#simfreq}
}
\author{
Richèl J.C. Bilderbeek
}
