% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qassoc_file.R
\name{read_qassoc_file}
\alias{read_qassoc_file}
\title{Read a \code{.qassoc} file}
\usage{
read_qassoc_file(qassoc_filename)
}
\value{
a tibble
}
\description{
Read a \code{.qassoc} file
}
\examples{
read_qassoc_file(qassoc_filename = get_plinkr_filename("run1.qassoc"))
}
