% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_fids.R
\name{rename_fids}
\alias{rename_fids}
\title{Rename all the family IDs within the data}
\usage{
rename_fids(data, fids)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
\item data to be used by \link{assoc},
as created by \link{create_assoc_data}
\item data to be used by \link{assoc_qt},
as created by \link{create_assoc_qt_data}
}}

\item{fids}{one or more family IDs (which are called \code{FID}s in \code{PLINK}).
Use \link{check_fids} to check if the elements of \code{fids} are valid.}
}
\value{
the data with renamed family IDs
}
\description{
Rename all the family IDs within the data
}
\examples{
# Use PLINK text data
data_before <- create_test_plink_text_data()
new_fids <- as.character(1 + data_before$ped_table$FID)
# the .ped table has the new FIDs
data_after <- rename_fids(
  data = data_before,
  fids = new_fids
)

# Use data used by plink::assoc_qt
data_before <- create_test_assoc_qt_data()
new_fids <- as.character(1 + data_before$data$ped_table$FID)
# the .ped and .phe tables have the new FIDs
data_after <- rename_fids(
  data = data_before,
  fids = new_fids
)
}
\author{
Richèl J.C. Bilderbeek
}
