% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_iids.R
\name{rename_iids}
\alias{rename_iids}
\title{Rename all the within-family IDs within the data}
\usage{
rename_iids(data, iids)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
}}

\item{iids}{one or more within-family IDs,
as can be checked by \link{check_iids}}
}
\value{
the data with renamed family IDs
}
\description{
Rename all the within-family IDs within the data
}
\examples{
# Use PLINK text data
data_before <- create_test_plink_text_data()
new_iids <- as.character(1 + data_before$ped_table$IID)
# the .ped table has the new iids
data_after <- rename_iids(
  data = data_before,
  iids = new_iids
)

# Use data used by plink::assoc_qt
data_before <- create_test_assoc_qt_data()
new_iids <- as.character(1 + data_before$data$ped_table$IID)
# the .ped and .phe tables have the new iids
data_after <- rename_iids(
  data = data_before,
  iids = new_iids
)
}
\author{
Richèl J.C. Bilderbeek
}
