% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plink.R
\name{run_plink}
\alias{run_plink}
\title{Run PLINK from arguments}
\usage{
run_plink(args, plink_folder = get_plink_folder())
}
\arguments{
\item{args}{arguments for the PLINK executable}

\item{plink_folder}{folder where PLINK is installed}
}
\value{
the text that PLINK returns
}
\description{
Run PLINK from arguments
}
