% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plink_v1_7.R
\name{run_plink_v1_7}
\alias{run_plink_v1_7}
\title{Run PLINK v1.7 from arguments}
\usage{
run_plink_v1_7(
  args,
  plink_options = create_plink_v1_7_options(),
  verbose = FALSE
)
}
\arguments{
\item{args}{arguments for the \code{PLINK} executable}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{set to TRUE for more output}
}
\value{
the text that PLINK returns
}
\description{
Run PLINK v1.7 from arguments
}
\examples{
if (is_plink_installed()) {

  run_plink_v1_7("--help")

}
}
\author{
Richèl J.C. Bilderbeek
}
