% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plink_v2_0.R
\name{run_plink_v2_0}
\alias{run_plink_v2_0}
\title{Run PLINK v2.0 from arguments}
\usage{
run_plink_v2_0(
  args,
  plink_options = create_plink_v2_0_options(),
  verbose = FALSE
)
}
\arguments{
\item{args}{arguments for the PLINK executable}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{set to TRUE for more output}
}
\value{
the text that PLINK returns
}
\description{
Run PLINK v2.0 from arguments
}
\examples{
if (is_plink_installed()) {

  run_plink_v2_0("--help")

}
}
\author{
Richèl J.C. Bilderbeek
}
