% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_str_split.R
\name{safe_str_split}
\alias{safe_str_split}
\title{Call \link[stringr]{str_split} safely
stringi::stri_split_regex \strong{sometimes} gives an
'embedded nul in string' error.
This has been not been reported, yet a similar report is at
https://github.com/gagolews/stringi/issues/458 .
Until then, use the workaround}
\usage{
safe_str_split(string, pattern, simplify = FALSE)
}
\arguments{
\item{string}{one or more strings}

\item{pattern}{regex, see \link[stringr]{str_split}}

\item{simplify}{simplify, see \link[stringr]{str_split}}
}
\description{
Call \link[stringr]{str_split} safely
stringi::stri_split_regex \strong{sometimes} gives an
'embedded nul in string' error.
This has been not been reported, yet a similar report is at
https://github.com/gagolews/stringi/issues/458 .
Until then, use the workaround
}
