% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_str_trim.R
\name{safe_str_trim}
\alias{safe_str_trim}
\title{Call \link[stringr]{str_trim} safely
stringi::str_trim \strong{sometimes} gives an 'embedded nul in string' error.
This has been reported at https://github.com/gagolews/stringi/issues/458 .
Until then, just try multiple times :-)}
\usage{
safe_str_trim(string)
}
\arguments{
\item{string}{one or more strings}
}
\description{
Call \link[stringr]{str_trim} safely
stringi::str_trim \strong{sometimes} gives an 'embedded nul in string' error.
This has been reported at https://github.com/gagolews/stringi/issues/458 .
Until then, just try multiple times :-)
}
