% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_bim_table_to_file.R
\name{save_bim_table_to_file}
\alias{save_bim_table_to_file}
\title{Save a \code{.bim} table to file.}
\usage{
save_bim_table_to_file(bim_table, bim_filename, verbose = FALSE)
}
\arguments{
\item{bim_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{chr}: the chromosome number
\item \code{id}: the SNP ID
\item \code{posg}: the position
\item \code{pos}: the position
\item \code{ref}: something
\item \code{alt}: something
}

Use \link{get_test_bim_table} to get a \code{.bim} table as used in testing.
Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.
Use \link{check_bim_table} to test if a \code{.bim} table is valid.}

\item{bim_filename}{name of a \code{PLINK} \code{.bim} file
Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Save a \code{.bim} table to file
}
\examples{
bim_table <- get_test_bim_table()
bim_filename <- get_plinkr_tempfilename(fileext = ".bim")
save_bim_table_to_file(
  bim_table = bim_table,
  bim_filename = bim_filename
)
file.remove(bim_filename)
}
\author{
Richèl J.C. Bilderbeek
}
