% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cov_table_to_file.R
\name{save_cov_table}
\alias{save_cov_table}
\title{Save a \code{.cov} table to file}
\usage{
save_cov_table(cov_table, cov_filename)
}
\arguments{
\item{cov_table}{a table of covariates.

A \code{cov_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of covariate values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_cov_file} to read a covariates file.
Use \link{check_cov_table} to test if a covariates table is valid.}

\item{cov_filename}{name of a covariates (\code{.cov}) file.
Use \link{read_plink_cov_file} to read a covariates file.}
}
\description{
Save a \code{.cov} table to file
}
\author{
Richèl J.C. Bilderbeek
}
