% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fam_table.R
\name{save_fam_table}
\alias{save_fam_table}
\title{Save a \code{.fam} table to file.}
\usage{
save_fam_table(fam_table, fam_filename, verbose = FALSE)
}
\arguments{
\item{fam_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{fam}: the family ID
\item \code{id}: the individual's ID
\item \code{pat}: ID of father
\item \code{mat}: ID of mother
\item \code{sex}: the gender
\item \code{pheno}: a phenotype
}

Use \link{get_test_fam_table} to get a \code{.fam} table as used in testing.
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.
Use \link{check_fam_table} to test if a \code{.fam} table is valid.}

\item{fam_filename}{name of a \code{PLINK} \code{.fam} file
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Save a \code{.fam} table to file
}
\examples{
fam_table <- get_test_fam_table()
fam_filename <- get_plinkr_tempfilename(fileext = ".fam")
save_fam_table(
  fam_table = fam_table,
  fam_filename = fam_filename
)
file.remove(fam_filename)
}
\author{
Richèl J.C. Bilderbeek
}
