% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_map_table.R
\name{save_map_table}
\alias{save_map_table}
\title{Save a \code{.map} table to file.}
\usage{
save_map_table(map_table, map_filename, plink_options = create_plink_options())
}
\arguments{
\item{map_table}{a genetic mapping table.

A \code{map_table} is a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{CHR}: the chromosome code or contig name.
The chromosome number cannot exceed 95
(\url{https://www.cog-genomics.org/plink/1.9/input#chr_set})
\item \code{SNP}: Variant identifier
\item \code{position_cm}: Position in morgans or centimorgans.
This value is optional. Zeroes denote it is unused
\item \code{BP}: Base-pair coordinat
}

Use \link{read_plink_map_file} to read a \code{PLINK} \code{.map} file.
Use \link{check_map_table} to test if a genetic mapping table is valid.}

\item{map_filename}{name of a \code{PLINK} \code{.map} file
Use \link{read_plink_map_file} to read a \code{PLINK} \code{.map} file.}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\description{
Save a \code{.map} table to file
}
\author{
Richèl J.C. Bilderbeek
}
