% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_map_table_to_file.R
\name{save_map_table_to_file}
\alias{save_map_table_to_file}
\title{Save a \code{.map} table to file.}
\usage{
save_map_table_to_file(map_table, map_filename)
}
\arguments{
\item{map_table}{a genetic mapping table.

A \code{map_table} is a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{CHR}: the chromosome code or contig name
\item \code{SNP}: Variant identifier
\item \code{position_cm}: Position in morgans or centimorgans.
This value is optional. Zeroes denote it is unused
\item \code{BP}: Base-pair coordinat
}

Use \link{read_plink_map_file} to read a PLINK \code{.map} file.
Use \link{check_map_table} to test if a genetic mapping table is valid.}

\item{map_filename}{name of a PLINK \code{.map} file
Use \link{read_plink_map_file} to read a PLINK \code{.map} file.}
}
\description{
Save a \code{.map} table to file
}
\author{
Richèl J.C. Bilderbeek
}
