% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ped_table_to_file.R
\name{save_ped_table_to_file}
\alias{save_ped_table_to_file}
\title{Save a \code{.ped} table to file.}
\usage{
save_ped_table_to_file(ped_table, ped_filename)
}
\arguments{
\item{ped_table}{a 'pedigree' table.

A \code{ped_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{family_id} The family ID (\code{FID})
\item \code{within_family_id} Within-family ID
(\code{IID}, cannot be zero)
\item \code{within_family_id_father} Within-family ID of father
(\code{0} if father isn't in dataset)
\item \code{within_family_id_mother} Within-family ID of mother
(\code{0} if mother isn't in dataset)
\item \code{sex_code} Sex code
(\code{1} = male, \code{2} = female, \code{0} = unknown)
\item \code{case_control_code} Case control code
(\code{1} = control, \code{2} = case,
\code{9}/\code{0}/non-numeric = missing data if case/control)
\item \code{snv_[x][y]} Nucleotide for the \code{x}th variant
for haplotype \code{y} (\code{y} is either \code{a} or \code{b})
in the \code{.map file} (\code{0} = no call)
}

Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.
Use \link{check_ped_table} to test if a pedigree table is valid.}

\item{ped_filename}{name of a PLINK \code{.ped} file.
Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.}
}
\description{
Save a \code{.ped} table to file to be used by PLINK.
Most columns have a single space between them.
The exceptions, where there are two spaces between the columns are:
\itemize{
\item between \code{sex_code} and \code{case_control_code}
\item between \code{case_control_code} and the first SNP
\item between each pair of SNPs
}
}
