% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pgen_table_to_file.R
\name{save_pgen_table}
\alias{save_pgen_table}
\title{Save a \code{.pgen} table to file.}
\usage{
save_pgen_table(pgen_table, pgen_filename, verbose = FALSE)
}
\arguments{
\item{pgen_table}{a \link{matrix} that maps the individuals
to their SNPs. TODO}

\item{pgen_filename}{name of a \code{PLINK2} \code{.pgen} file
Use \link{read_plink2_pgen_file} to read a \code{PLINK2} \code{.pgen} file}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Save a \code{.pgen} table to file
}
\examples{
pgen_table <- get_test_pgen_table()
pgen_filename <- get_plinkr_tempfilename(fileext = ".pgen")
if (1 == 2) {
  save_pgen_table(
    pgen_table = pgen_table,
    pgen_filename = pgen_filename
  )
  file.remove(pgen_filename)
}
}
\author{
Richèl J.C. Bilderbeek
}
