% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_phe_table_to_file.R
\name{save_phe_table_to_file}
\alias{save_phe_table_to_file}
\title{Save a \code{.phe} table to file}
\usage{
save_phe_table_to_file(phe_table, phe_filename)
}
\arguments{
\item{phe_table}{a table of phenotypes.

A \code{phe_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_phe_file} to read a phenotype file.
Use \link{check_phe_table} to test if a phenotype table is valid.}

\item{phe_filename}{name of a phenotype (\code{.phe}) file.
Use \link{read_plink_phe_file} to read a raw phenotype file.}
}
\description{
Save a \code{.phe} table to file
}
\author{
Richèl J.C. Bilderbeek
}
