% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_phenotype_data_table.R
\name{save_phenotype_data_table}
\alias{save_phenotype_data_table}
\title{Save a \code{phenotype_data} (in the form of a \code{phenotype_data_table}) to file}
\usage{
save_phenotype_data_table(phenotype_data_table, phe_filename)
}
\arguments{
\item{phenotype_data_table}{phenotype data
(as checked by \link{check_phenotype_data}))
in the form of a \link{list} with an in-memory phenotype table}

\item{phe_filename}{name of a phenotype (\code{.phe}) file,
as checked by \link{check_phe_filename}.
Use \link{read_plink_phe_file} to read a raw phenotype file.}
}
\value{
a \code{phenotype_data} in the form of a \code{phenotype_data_filename}
}
\description{
Save a \code{phenotype_data} (in the form of a \code{phenotype_data_table}) to file,
returning a \code{phenotype_data} in the form of a \code{phenotype_data_filename}
}
\author{
Richèl J.C. Bilderbeek
}
