% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plink2_bin_data.R
\name{save_plink2_bin_data}
\alias{save_plink2_bin_data}
\title{Save \code{PLINK2} binary data to files.}
\usage{
save_plink2_bin_data(
  plink2_bin_data,
  base_input_filename = get_plinkr_tempfilename(),
  plink_options = create_plink_v2_0_options(),
  verbose = FALSE
)
}
\arguments{
\item{plink2_bin_data}{the genetic and phenotypic data for \code{PLINK2} to work
on, in \code{PLINK2} binary format, as created by \link{create_plink2_bin_data}.}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{plink2_bin_filenames}, which is list with the following elements:
\itemize{
\item \code{pgen_filename}: full paths to the \code{.pgen} created
\item \code{psam_filename}: full paths to the \code{.psam} created
\item \code{pvar_filename}: full paths to the \code{.pvar} created
}
}
\description{
Save \code{PLINK2} binary data to files.
}
\seealso{
to save \code{PLINK}/\code{PLINK2} data use:
\itemize{
\item \link{save_plink_text_data}: to save \code{PLINK} text data
\item \link{save_plink_bin_data}: to save \code{PLINK} binary data
\item \link{save_plink2_bin_data}: to save \code{PLINK2} binary data
}
}
\author{
Richèl J.C. Bilderbeek
}
