% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plink_bin_data.R
\name{save_plink_bin_data}
\alias{save_plink_bin_data}
\title{Save \code{PLINK} binary data to files.}
\usage{
save_plink_bin_data(
  plink_bin_data,
  base_input_filename = get_plinkr_tempfilename(),
  verbose = FALSE
)
}
\arguments{
\item{plink_bin_data}{the genetic and phenotypic data for \code{PLINK} to work
on, in \code{PLINK} binary format, as created by \link{create_plink_bin_data}.}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
the full paths to the \code{.bed}, \code{.bim} and \code{.fam} files created
}
\description{
Save \code{PLINK} binary data to files.
}
\seealso{
to save \code{PLINK}/\code{PLINK2} data use:
\itemize{
\item \link{save_plink_text_data}: to save \code{PLINK} text data
\item \link{save_plink_bin_data}: to save \code{PLINK} binary data
\item \link{save_plink2_bin_data}: to save \code{PLINK2} binary data
}
}
\author{
Richèl J.C. Bilderbeek
}
