% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plink_text_data.R
\name{save_plink_text_data}
\alias{save_plink_text_data}
\title{Save \code{PLINK} text data to files.}
\usage{
save_plink_text_data(
  plink_text_data,
  base_input_filename = get_plinkr_tempfilename(),
  verbose = FALSE
)
}
\arguments{
\item{plink_text_data}{the genetic and phenotypic data for \code{PLINK} to work
on, in \code{PLINK} text format, as created by \link{create_plink_text_data}.}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
the full paths to the \code{.bed}, \code{.bim} and \code{.fam} files created
}
\description{
Save \code{PLINK} text data to files.
}
\examples{
#
# Example 1: using plinkr::create_test_plink_text_data
#

# Folder and base filename to save to
base_input_filename <- file.path(
  get_plinkr_tempfilename(),
  "save_plink_text_data"
)

# Create PLINK text data
plink_text_data <- create_test_plink_text_data()

# Save. Convenient: filenames are returned!
plink_text_filenames <- save_plink_text_data(
  plink_text_data = plink_text_data,
  base_input_filename = base_input_filename
)

# Cleanup
file.remove(plink_text_filenames$map_filename)
file.remove(plink_text_filenames$ped_filename)
unlink(dirname(base_input_filename), recursive = TRUE)

#
# Example 2: using plinkr::create_demo_assoc_qt_data
#

# Folder and base filename to save to
base_input_filename <- file.path(
  get_plinkr_tempfilename(),
  "save_plink_text_with_demo_assoc_qt_data"
)

# Create PLINK text data
assoc_qt_data <- create_demo_assoc_qt_data()

# Save. Convenient: filenames are returned!
plink_text_filenames <- save_plink_text_data(
  plink_text_data = create_test_plink_text_data(),
  base_input_filename = base_input_filename
)

# Cleanup
file.remove(plink_text_filenames$map_filename)
file.remove(plink_text_filenames$ped_filename)
unlink(dirname(base_input_filename), recursive = TRUE)
}
\seealso{
to save \code{PLINK}/\code{PLINK2} data use:
\itemize{
\item \link{save_plink_text_data}: to save \code{PLINK} text data
\item \link{save_plink_bin_data}: to save \code{PLINK} binary data
\item \link{save_plink2_bin_data}: to save \code{PLINK2} binary data
}
}
\author{
Richèl J.C. Bilderbeek
}
