% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_psam_table_to_file.R
\name{save_psam_table_to_file}
\alias{save_psam_table_to_file}
\title{Save a \code{.psam} table to file.}
\usage{
save_psam_table_to_file(psam_table, psam_filename, verbose = FALSE)
}
\arguments{
\item{psam_table}{a \link[tibble]{tibble} with as much rows as
individuals. It has the following columns:
\itemize{
\item \code{FID}: the family ID
\item \code{IID}: the within-family ID
\item \code{SEX}: the sex
\item \code{PHENO1}: a phenotypic value
}

The uppercase column names are those as used by PLINK2.}

\item{psam_filename}{name of a \code{PLINK2} \code{.psam} file
Use \link{read_plink2_psam_file} to read a \code{PLINK2} \code{.psam} file}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Save a \code{.psam} table to file
}
\examples{
psam_table <- get_test_psam_table()
psam_filename <- get_plinkr_tempfilename(fileext = ".psam")
save_psam_table_to_file(
  psam_table = psam_table,
  psam_filename = psam_filename
)
file.remove(psam_filename)
}
\author{
Richèl J.C. Bilderbeek
}
