% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pvar_table.R
\name{save_pvar_table}
\alias{save_pvar_table}
\title{Save a \code{.pvar} table to file.}
\usage{
save_pvar_table(pvar_table, pvar_filename, verbose = FALSE)
}
\arguments{
\item{pvar_table}{a \link[tibble]{tibble} with as much rows as
SNPs. It has the following columns:
\itemize{
\item \code{CHROM}: the chromosome
\item \code{POS}: the position on the chromosome
\item \code{ID}: the SNP ID
\item \code{REF}: the reference variant
\item \code{ALT}: the alternate variant
}

The uppercase column names are those as used by PLINK2.}

\item{pvar_filename}{name of a \code{PLINK2} \code{.pvar} file
Use \link{read_plink2_pvar_file} to read a \code{PLINK2} \code{.pvar} file}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Save a \code{.pvar} table to file
}
\examples{
pvar_table <- get_test_pvar_table()
pvar_filename <- get_plinkr_tempfilename(fileext = ".pvar")
save_pvar_table(
  pvar_table = pvar_table,
  pvar_filename = pvar_filename
)
file.remove(pvar_filename)
}
\author{
Richèl J.C. Bilderbeek
}
