% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_qassoc_table.R
\name{save_qassoc_table}
\alias{save_qassoc_table}
\title{Save a \code{.qassoc} table to file.}
\usage{
save_qassoc_table(
  qassoc_table,
  qassoc_filename,
  plink_options = create_plink_options()
)
}
\arguments{
\item{qassoc_table}{the table that hold the result of an association
with a quantitative trait, as created by \code{PLINK}/\code{PLINK2}
and stored as a \code{.qassoc} file (for \code{PLINK}) or
\verb{[basename].<phenotype>.glm.linear} (for \code{PLINK2}).}

\item{qassoc_filename}{name of a \code{PLINK} \code{.qassoc} file
Use \link{read_plink_qassoc_file} to read a \code{PLINK} \code{.qassoc} file.}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\description{
Save a \code{.qassoc} table to file
}
\author{
Richèl J.C. Bilderbeek
}
