% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_simulate_qt_params_to_file.R
\name{save_simulate_qt_params_to_file}
\alias{save_simulate_qt_params_to_file}
\title{Save the parameters for a quantitative traits simulation
to file}
\usage{
save_simulate_qt_params_to_file(simulate_qt_params, sim_filename)
}
\arguments{
\item{simulate_qt_params}{the parameters for a quantitative
traits simulation, as can be created by
\link{create_simulate_qt_params}}

\item{sim_filename}{name of a PLINK \code{.sim} file}
}
\description{
This function is named after the \code{--simulate-qt} PLINK flag.
}
