% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_samples.R
\name{select_samples}
\alias{select_samples}
\title{Select samples/individuals}
\usage{
select_samples(
  data,
  sample_selector,
  sample_ids_filename = plinkr::get_plinkr_tempfilename(pattern = "sample_ids_",
    fileext = ".txt"),
  base_output_filename = plinkr::get_plinkr_tempfilename(),
  plink_options = plinkr::create_plink_options(),
  verbose = FALSE
)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
\item data to be used by \link{assoc},
as created by \link{create_assoc_data}
\item data to be used by \link{assoc_qt},
as created by \link{create_assoc_qt_data}
}}

\item{sample_selector}{a sample/individual selector,
a way to select one or more samples/individuals.
See \link{create_samples_selector} for all sample selectors.}

\item{sample_ids_filename}{name of a file to store \code{sample_ids}
(see \link{check_sample_ids}) to}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{plink_bin_data} (see \link{check_plink_bin_data})
}
\description{
Select samples/individuals from data
}
\examples{
if (is_plink_installed()) {
  # Use PLINK files
  plink_bin_filenames <- create_plink_bin_filenames(
    bed_filename = get_plinkr_filename("select_snps.bed"),
    bim_filename = get_plinkr_filename("select_snps.bim"),
    fam_filename = get_plinkr_filename("select_snps.fam")
  )

  # Select a single sample
  single_sample_selector <- create_single_sample_selector(
    fid = "C",
    iid = "1"
  )
  plink_bin_data <- select_samples(
    data = plink_bin_filenames,
    sample_selector = single_sample_selector
  )

  # Select 2 random samples
  random_samples_selector <- create_random_samples_selector(
    n_samples = 2
  )
  plink_bin_data <- select_samples(
    data = plink_bin_filenames,
    sample_selector = random_samples_selector
  )
}
}
\seealso{
To select SNPs, use \link{select_snps}

There are multiple sample selector checking functions,
see \link{create_samples_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
