% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_qt.R
\name{simulate_qt}
\alias{simulate_qt}
\title{Simulate a quantitative trait}
\usage{
simulate_qt(
  simulate_qt_params,
  n_individuals,
  temp_sim_filename = get_plinkr_tempfilename(),
  plink_options = create_plink_options(),
  verbose = FALSE
)
}
\arguments{
\item{simulate_qt_params}{the parameters for a quantitative
traits simulation, as can be created by
\link{create_simulate_qt_params}}

\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}

\item{temp_sim_filename}{temporary file to store simulation
parameters, which is a \code{PLINK} \code{.sim} file}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Simulate a quantitative trait
}
\note{
This function is named after the \code{--simulate-qt} PLINK flag.
}
\author{
Richèl J.C. Bilderbeek
}
