% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_plink.R
\name{uninstall_plink}
\alias{uninstall_plink}
\title{Uninstall PLINK}
\usage{
uninstall_plink(plink_folder = get_plink_folder())
}
\arguments{
\item{plink_folder}{folder where PLINK is installed}
}
\value{
Nothing
}
\description{
Uninstall PLINK
}
