% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_plink_tutorial_data.R
\name{uninstall_plink_tutorial_data}
\alias{uninstall_plink_tutorial_data}
\title{Uninstall the \code{PLINK} tutorial data}
\usage{
uninstall_plink_tutorial_data(plink_folder = get_plink_folder())
}
\arguments{
\item{plink_folder}{folder where \code{PLINK} is installed}
}
\value{
Nothing. Will \link{stop} if the tutorial data is already uninstalled
}
\description{
Uninstall the \code{PLINK} tutorial data.
}
\details{
The data file will be removed from the \code{plink_folder}
and the \code{plink_tutorial} subfolder will be removed.
}
\seealso{
use \link{install_plink_tutorial_data} to install
the \code{PLINK} tutorial data
}
\author{
Richèl J.C. Bilderbeek
}
