% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_plinks.R
\name{uninstall_plinks}
\alias{uninstall_plinks}
\title{Uninstall multiple versions of PLINK}
\usage{
uninstall_plinks(plink_optionses = create_plink_optionses())
}
\description{
Uninstall multiple versions of PLINK
}
\author{
Richèl J.C. Bilderbeek
}
