% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{assoc}
\alias{assoc}
\title{Let \code{PLINK}/\code{PLINK2} detect an association for a binary trait.}
\usage{
assoc(assoc_params, plink_options = create_plink_options(), verbose = FALSE)
}
\arguments{
\item{assoc_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc}),
as can be created by \link{create_assoc_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \link{list} with the following columns:
\itemize{
\item \code{assoc_table}: a \link[tibble]{tibble} with associations
found by \code{PLINK}.
See \link{read_plink_assoc_file} for an explanation of the
column names.
\item \code{log}: the log file as text as created by \code{PLINK}
}
}
\description{
Let \code{PLINK}/\code{PLINK2} detect an association for a binary trait.
}
\note{
This function is named after the \code{--assoc} flag
used by \code{PLINK}
}
\examples{
if (is_plink_installed()) {
  assoc(create_demo_assoc_params())
}
}
\author{
Richèl J.C. Bilderbeek
}
