% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{assoc}
\alias{assoc}
\title{Let PLINK detect an association with one binary traits.}
\usage{
assoc(assoc_params, plink_options = create_plink_options(), verbose = FALSE)
}
\arguments{
\item{assoc_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc}),
as can be created by \link{create_assoc_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{CHR}: Chromosome number
\item \code{SNP}: SNP identifier
\item \code{BP}: Physical position (base-pair)
\item \code{A1}: Minor allele name (based on whole sample)
\item \code{F_A}: Frequency of this allele in cases
\item \code{F_U}: Frequency of this allele in controls
\item \code{A2}: Major allele name
\item \code{CHISQ}: Basic allelic test chi-square (1df)
\item \code{P}: Asymptotic p-value for this test
\item \code{OR}: Estimated odds ratio (for A1, i.e. A2 is reference)
\item \code{SE}: Standard error of the estimated log(odds ratio)
\item \code{L95}: Lower bound of the 95\% confidence interval of the odds ratio
\item \code{U95}: Upper bound of the 95\% confidence interval of the odds ratio
}

The table with have as much rows as the number of SNPs

Note that parameters in uppercase are named as such by PLINK.
}
\description{
Let PLINK detect an association with one binary traits.
}
\note{
This function is named after the \code{--assoc} flag used by PLINK
}
\examples{
if (is_plink_installed()) {
  assoc(create_demo_assoc_params())
}
}
\author{
Richèl J.C. Bilderbeek
}
