% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_from_bfile.R
\name{assoc_from_bfile}
\alias{assoc_from_bfile}
\title{Let \code{PLINK}/\code{PLINK2}
do a binary-trait/case-control association}
\usage{
assoc_from_bfile(bfile, out)
}
\arguments{
\item{bfile}{the base filename of the binary files (i.e.
a \code{.bed}, \code{.bim} and \code{.fam} file).
This parameter is named after the \code{PLINK}
\code{--bfile} flag}

\item{out}{the base filename of the output files.
This parameter is named after the \code{PLINK}
\code{--out} flag}
}
\value{
a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{CHR}: Chromosome number
\item \code{SNP}: SNP identifier
\item \code{BP}: Physical position (base-pair)
\item \code{A1}: Minor allele name (based on whole sample)
\item \code{F_A}: Frequency of this allele in cases
\item \code{F_U}: Frequency of this allele in controls
\item \code{A2}: Major allele name
\item \code{CHISQ}: Basic allelic test chi-square (1df)
\item \code{P}: Asymptotic p-value for this test
\item \code{OR}: Estimated odds ratio (for A1, i.e. A2 is reference)
\item \code{SE}: Standard error of the estimated log(odds ratio)
\item \code{L95}: Lower bound of the 95\% confidence interval of the odds ratio
\item \code{U95}: Upper bound of the 95\% confidence interval of the odds ratio
}

The table with have as much rows as the number of SNPs

Note that parameters in uppercase are named as such by PLINK.
}
\description{
Let \code{PLINK}/\code{PLINK2}
do a binary-trait/case-control association
from PLINK1 binary filenames
}
\note{
This function is named after the \code{--assoc} flag
used by \code{PLINK}
}
\author{
Richèl J.C. Bilderbeek
}
