% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_qt.R
\name{assoc_qt}
\alias{assoc_qt}
\title{Let \code{PLINK}/\code{PLINK2} detect an association
with one or more quantitative traits.}
\usage{
assoc_qt(
  assoc_qt_params,
  plink_options = create_plink_options(),
  verbose = FALSE
)
}
\arguments{
\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
A \code{assoc_qt_result}, which is a list
which is a \link{list} with elements:
\itemize{
\item \code{qassoc_table}: the quantitative analysis results table,
as can be checked by \link{check_qassoc_table}
\item \code{log}: the text from the log file created by \code{PLINK}/\code{PLINK2}
when doing \link{assoc_qt}
}
}
\description{
Let \code{PLINK}/\code{PLINK2} detect an association
with one or more quantitative traits.
}
\note{
This function is named after the \code{--assoc} flag used by PLINK.
}
\examples{
if (is_plink_installed()) {
  assoc_qt(create_demo_assoc_qt_data())
}
}
\seealso{
Use \link{assoc} to do a case-control association
}
\author{
Richèl J.C. Bilderbeek
}
