% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_qt.R
\name{assoc_qt}
\alias{assoc_qt}
\title{Let PLINK detect an association with a quantitative trait}
\usage{
assoc_qt(assoc_qt_params)
}
\description{
This function is named after the \code{--assoc} flag used by PLINK
}
