% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_qt_nth_trait.R
\name{assoc_qt_nth_trait}
\alias{assoc_qt_nth_trait}
\title{Let PLINK detect an association with the n-th quantitative trait}
\usage{
assoc_qt_nth_trait(
  assoc_qt_params,
  n,
  plink_options = create_plink_options(),
  verbose = FALSE
)
}
\arguments{
\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}

\item{n}{the nth trait. Use '1' for the first trait}

\item{verbose}{set to TRUE for more output}
}
\description{
Let PLINK detect an association with the n-th quantitative trait
}
\note{
This function is named after the \code{--assoc} flag used by PLINK
}
\author{
Richèl J.C. Bilderbeek
}
