% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assoc_qt_covar_data.R
\name{check_assoc_qt_covar_data}
\alias{check_assoc_qt_covar_data}
\title{Check if the \code{assoc_qt_covar_data} are valid, will stop otherwise}
\usage{
check_assoc_qt_covar_data(assoc_qt_covar_data)
}
\arguments{
\item{assoc_qt_covar_data}{data to do an association analysis
for a quantitative trait with covariates
(i.e. using \link{assoc_qt_covar}),
as can be created by \link{create_assoc_qt_covar_data}}
}
\value{
Nothing.
}
\description{
Check if the \code{assoc_qt_covar_data} are valid, will stop otherwise
}
\note{
This function is named after the \code{--assoc} PLINK flag.
}
\author{
Richèl J.C. Bilderbeek
}
