% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assoc_qt_data.R
\name{check_assoc_qt_data}
\alias{check_assoc_qt_data}
\title{Check if the \code{assoc_qt_data} is valid}
\usage{
check_assoc_qt_data(assoc_qt_data)
}
\arguments{
\item{assoc_qt_data}{data use to used by \link{assoc_qt}
(as created by \link{create_assoc_qt_data}),
which equals 'regular' data (see \link{check_data}) and a
phenotype table/file}
}
\value{
nothing.
}
\description{
Check if the \code{assoc_qt_data} is valid.
}
\details{
\code{assoc_qt_data} is a list combined of:
\itemize{
\item 'regular' data, as checked by \link{check_data}
\item phenotype data, as checked by \link{check_phenotype_data}
}

Will \link{stop} if the \code{assoc_qt_data} is invalid
}
\seealso{
\code{plinkr} has the following functions to check if in-memory \code{data}
is valid:
\itemize{
\item \code{PLINK} text data: use \link{check_plink_text_data}
\item \code{PLINK} binary data: use \link{check_plink_bin_data}
\item \code{PLINK2} binary data: use \link{check_plink2_bin_data}
\item any \code{PLINK}/\code{PLINK2} data: use \link{check_data}
\item quantitative trait analysis data: use \link{check_assoc_qt_data}
}
}
\author{
Richèl J.C. Bilderbeek
}
