% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assoc_qt_params.R
\name{check_assoc_qt_params}
\alias{check_assoc_qt_params}
\title{Check if the \code{assoc_qt_params} are valid, will stop otherwise}
\usage{
check_assoc_qt_params(assoc_qt_params)
}
\arguments{
\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}
}
\value{
Nothing.
}
\description{
This function is named after the \code{--assoc-qt} PLINK flag.
}
