% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bed_table.R
\name{check_bed_table}
\alias{check_bed_table}
\title{Check if a \code{.bed} table is valid.}
\usage{
check_bed_table(bed_table)
}
\arguments{
\item{bed_table}{a table that maps the SNPs to the individuals,
of which the column names are the names of the individuals,
the row names are the names of the SNPs,
and the values are the SNP variant.
Use \link{get_test_bed_table} to get a \code{.bed} table as used in testing.
Use \link{read_plink_bed_file} to read a \code{PLINK} \code{.bed} file.
Use \link{check_bed_table} to test if a \code{.bed} table is valid.}
}
\value{
Nothing. Will \link{stop} if the \code{bed_table} is invalid
}
\description{
Check if a \code{.bed} table is valid.
Will \link{stop} if not
}
\examples{
# Simple .bed table
check_bed_table(get_test_bed_table())

# Read a .bed table from file, one needs a .bim and .fam for that as well
bim_table <- read_plink_bim_file(get_plinkr_filename("toy_data.bim"))
fam_table <- read_plink_fam_file(get_plinkr_filename("toy_data.fam"))
bed_table <- read_plink_bed_file(
  bed_filename = get_plinkr_filename("toy_data.bed"),
  names_loci = bim_table$id,
  names_ind = fam_table$id
)
check_bed_table(bed_table)
}
\author{
Richèl J.C. Bilderbeek
}
