% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_calc_phenotype_function.R
\name{check_calc_phenotype_function}
\alias{check_calc_phenotype_function}
\title{Check if the \code{calc_phenotype_function} is valid.
Check if the \code{calc_phenotype_function}
is valid.}
\usage{
check_calc_phenotype_function(calc_phenotype_function)
}
\arguments{
\item{calc_phenotype_function}{a function that calculate the phenotypes
from genotypes. The input is the genetic data as a \link[tibble]{tibble},
in which each row is an individual and the columns are the SNVs.
The first two columns are named \code{snv_1a}, \code{snv_1b} and
hold the genetic data for the first SNV of a diploid organism.
If there are more SNVs, columns continue with names,
\code{snv_2a}, \code{snv_2b}, \code{snv_3a}, \code{snv_3b}, etc.
Nucleotides are in uppercase.
The output must be the phenotypic values, as a numeric vector,
which has the same length as the number of individuals.
Use \link{check_calc_phenotype_function}
to check a \code{calc_phenotype_function}.}
}
\value{
Nothing. Will \link{stop} if the input
\code{calc_phenotype_function} is invalid
}
\description{
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
