% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_confidence_interval.R
\name{check_confidence_interval}
\alias{check_confidence_interval}
\title{Check if the \code{confidence_interval} is valid.
Check if the \code{confidence_interval} is valid,
which is when it is a number between
0.0 and 1.0 (i.e. excluding 0.0 and 1.0).}
\usage{
check_confidence_interval(confidence_interval)
}
\arguments{
\item{confidence_interval}{confidence interval, a value between (and
excluding both) 0.0 and 1.0. The confidence interval helps assess
the certainty of an estimation: you can be 99 percent sure
a value is within the range of the 0.99 confidence interval}
}
\value{
Nothing. Will \link{stop} if the input
is not a number between
0.0 and 1.0
}
\description{
Will \link{stop} if not
}
\examples{
check_confidence_interval(0.01)
check_confidence_interval(0.95)
check_confidence_interval(0.99)
}
\author{
Richèl J.C. Bilderbeek
}
