% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check if the data is valid}
\usage{
check_data(data)
}
\description{
Check if the data is valid.
The data can be in PLINK text, PLINK binary or PLINK2 binary format.
Will \link{stop} if the data is not in any of these formats
}
