% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check if the data is valid}
\usage{
check_data(data)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
\item data to be used by \link{assoc},
as created by \link{create_assoc_data}
\item data to be used by \link{assoc_qt},
as created by \link{create_assoc_qt_data}
}}
}
\value{
nothing.
}
\description{
Check if the data is valid.
The data can be in PLINK text, PLINK binary or PLINK2 binary format.
Will \link{stop} if the data is not in any of these formats
}
\examples{
check_data(data = create_test_plink_text_data())
check_data(data = create_test_plink_bin_data())
check_data(data = create_test_plink2_bin_data())
check_data(data = create_test_plink_text_filenames())
check_data(data = create_test_plink_bin_filenames())
check_data(data = create_test_plink2_bin_filenames())
}
\seealso{
\code{plinkr} has the following functions to check if in-memory \code{data}
is valid:
\itemize{
\item \code{PLINK} text data: use \link{check_plink_text_data}
\item \code{PLINK} binary data: use \link{check_plink_bin_data}
\item \code{PLINK2} binary data: use \link{check_plink2_bin_data}
\item any \code{PLINK}/\code{PLINK2} data: use \link{check_data}
\item quantitative trait analysis data: use \link{check_assoc_qt_data}
}
}
\author{
Richèl J.C. Bilderbeek
}
