% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_equal_number_of_snvs.R
\name{check_equal_number_of_snvs}
\alias{check_equal_number_of_snvs}
\title{Internal function}
\usage{
check_equal_number_of_snvs(data)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
\item data to be used by \link{assoc},
as created by \link{create_assoc_data}
\item data to be used by \link{assoc_qt},
as created by \link{create_assoc_qt_data}
}}
}
\description{
Check if the data has the same number of
SNVs in its tables.
Will \link{stop} if not.
}
\examples{
check_equal_number_of_snvs(create_test_plink_text_data())
check_equal_number_of_snvs(create_test_plink_bin_data())
check_equal_number_of_snvs(create_test_plink2_bin_data())
}
\author{
Richèl J.C. Bilderbeek
}
