% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_equal_number_of_snvs.R
\name{check_equal_number_of_snvs}
\alias{check_equal_number_of_snvs}
\title{Internal function, called by \link{check_assoc_qt_params}}
\usage{
check_equal_number_of_snvs(assoc_qt_params)
}
\arguments{
\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}
}
\description{
Check if the \code{assoc_qt_params} has the same number of
SNVs in the genetic mapping and the pedigree table.
Will \link{stop} if not.
}
