% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_maf.R
\name{check_maf}
\alias{check_maf}
\title{Check if the \code{maf} is valid.
Check if the \code{maf}, an abbreviation of 'Minor Allele
Frequency' is valid, which is when it is a number between
0.0 and 0.5 (i.e. excluding 0.0 and 0.5).}
\usage{
check_maf(maf)
}
\arguments{
\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the \code{PLINK} analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
\code{PLINK} \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}
}
\value{
Nothing. Will \link{stop} if the input
is not number between
0.0 and 0.5
}
\description{
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
