% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mafs.R
\name{check_mafs}
\alias{check_mafs}
\title{Check the \code{mafs}, the minor allele frequencies.}
\usage{
check_mafs(mafs)
}
\arguments{
\item{mafs}{one or more minor allele frequencies.
These allele frequencies must be ordered decreasingly,
i.e. the MAF is at the first position, where the
even rarer alleles are at the second and third positions.

Note that \code{PLINK} cannot handle triallelic nor
quadallelic SNPs: \code{PLINK} will give a warning that it
is setting the rarest alleles to missing.}
}
\description{
Check the \code{mafs}, the minor allele frequencies,
will \link{stop} if these are invalid.
}
\examples{
check_mafs(0.25)
check_mafs(c(0.1, 0.01))
check_mafs(c(0.1, 0.01, 0.001))
}
\author{
Richèl J.C. Bilderbeek
}
