% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ped_table.R
\name{check_ped_table}
\alias{check_ped_table}
\title{Check if a \code{.ped} is valid. Will \link{stop} if not}
\usage{
check_ped_table(ped_table)
}
\arguments{
\item{ped_table}{a 'pedigree' table.

A \code{ped_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{family_id} The family ID (\code{FID})
\item \code{within_family_id} Within-family ID
(\code{IID}, cannot be zero)
\item \code{within_family_id_father} Within-family ID of father
(\code{0} if father isn't in dataset)
\item \code{within_family_id_mother} Within-family ID of mother
(\code{0} if mother isn't in dataset)
\item \code{sex_code} Sex code
(\code{1} = male, \code{2} = female, \code{0} = unknown)
\item \code{case_control_code} Case control code
(\code{1} = control, \code{2} = case,
\code{9}/\code{0}/non-numeric = missing data if case/control)
\item \code{snv_[x][y]} Nucleotide for the \code{x}th variant
for haplotype \code{y} (\code{y} is either \code{a} or \code{b})
in the \code{.map file} (\code{0} = no call)
}

Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.
Use \link{check_ped_table} to test if a pedigree table is valid.}
}
\value{
Nothing.
Will \link{stop} if the pedigree table is invalid.
}
\description{
Check if a \code{.ped} is valid. Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
