% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ped_table.R
\name{check_ped_table}
\alias{check_ped_table}
\title{Check if a \code{.ped} table is valid. Will \link{stop} if not}
\usage{
check_ped_table(ped_table)
}
\arguments{
\item{ped_table}{a 'pedigree' table.

A \code{ped_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item \code{within_family_id_father} Within-family ID of father
(\code{0} if father isn't in dataset)
\item \code{within_family_id_mother} Within-family ID of mother
(\code{0} if mother isn't in dataset)
\item \code{sex_code} Sex code
(\code{1} = male, \code{2} = female, \code{0} = unknown)
\item \code{case_control_code} Case control code
(\code{1} = control, \code{2} = case,
\code{9}/\code{0}/non-numeric = missing data if case/control)
\item \code{snv_[x][y]} Nucleotide for the \code{x}th variant
for haplotype \code{y} (\code{y} is either \code{a} or \code{b})
in the \code{.map file} (\code{0} = no call)
}

The \code{FID} and \code{IID} column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}.

Use \link{read_plink_ped_file} to read a \code{PLINK} \code{.ped} file.
Use \link{check_ped_table} to test if a pedigree table is valid.}
}
\value{
Nothing.
Will \link{stop} if the pedigree table is invalid.
}
\description{
Check if a \code{.ped} table is valid. Will \link{stop} if not
}
\examples{
check_ped_table(get_test_ped_table())
check_ped_table(read_plink_ped_file(get_plink_example_filename("toy.ped")))
}
\author{
Richèl J.C. Bilderbeek
}
