% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phenotype_table.R
\name{check_phenotype_table}
\alias{check_phenotype_table}
\title{Check if a phenotype table is valid.}
\usage{
check_phenotype_table(phenotype_table)
}
\arguments{
\item{phenotype_table}{a table of phenotypes.

A \code{phenotype_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{family_id} The family ID (\code{FID})
\item \code{within_family_id} Within-family ID
(\code{IID}, cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

Use \link{read_plink_phenotype_file} to read a phenotype file.
Use \link{check_phenotype_table} to test if a phenotype table is valid.}
}
\value{
Nothing. Will \link{stop} if the \code{phenotype_table} is invalid
}
\description{
Check if a phenotype table is valid.
Will \link{stop} if not
}
