% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phenotypes.R
\name{check_phenotypes}
\alias{check_phenotypes}
\title{Check if the phenotypes are indeed phenotypes with a clear genetic setup.}
\usage{
check_phenotypes(phenotypes)
}
\arguments{
\item{phenotypes}{one ore more phenotypes,
named after their genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive
(the nucleotides used are A and T as these are in the word 'additive'):
\itemize{
\item \code{AA} 1.0
\item \code{AT} 0.5
\item \code{TT} 0.0
}
}}
}
\value{
Nothing
Will \link{stop} if the phenotypes are not
in the supported list of phenotypes with a clear genetic setup,
as can be obtained with \link{get_phenotypes}
}
\description{
Check if the phenotypes are indeed phenotypes with a clear genetic setup.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
