% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink_bin_data.R
\name{check_plink_bin_data}
\alias{check_plink_bin_data}
\title{Check if the \code{PLINK} bin data is valid.}
\usage{
check_plink_bin_data(plink_bin_data)
}
\arguments{
\item{plink_bin_data}{the in-memory binary data for \code{PLINK} to work on,
as created by \link{create_plink_bin_data}.}
}
\value{
nothing
}
\description{
Check if the \code{PLINK} bin data is valid.
Will \link{stop} if not.
}
\details{
Valid \code{PLINK} binary data has the following properties:
\itemize{
\item The \code{bed_table} must be valid, as checked by \link{check_bed_table}
\item The \code{bim_table} must be valid, as checked by \link{check_bim_table}
\item The \code{fam_table} must be valid, as checked by \link{check_fam_table}
\item The number of SNVs must match,
as checked by \link{check_equal_number_of_snvs}
\item The column names of the \code{bed_table} must match the \code{id}s in the
\code{fam_table}
}
}
\examples{
plink_bin_data <- create_test_plink_bin_data()
check_plink_bin_data(plink_bin_data)
}
\seealso{
\code{plinkr} has the following functions to check if in-memory \code{data}
is valid:
\itemize{
\item \code{PLINK} text data: use \link{check_plink_text_data}
\item \code{PLINK} binary data: use \link{check_plink_bin_data}
\item \code{PLINK2} binary data: use \link{check_plink2_bin_data}
\item any \code{PLINK}/\code{PLINK2} data: use \link{check_data}
\item quantitative trait analysis data: use \link{check_assoc_qt_data}
}
}
\author{
Richèl J.C. Bilderbeek
}
