% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink_is_installed.R
\name{check_plink_is_installed}
\alias{check_plink_is_installed}
\title{Check if PLINK is installed}
\usage{
check_plink_is_installed(
  plink_version = get_default_plink_version(),
  plink_folder = get_plink_folder()
)
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}}

\item{plink_folder}{folder where PLINK is installed}
}
\value{
Nothing. Will \link{stop} if PLINK is not installed.
}
\description{
Check if PLINK is installed. Will \link{stop} if not
}
