% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink_text_data.R
\name{check_plink_text_data}
\alias{check_plink_text_data}
\title{Check if the \code{PLINK} text data is valid.}
\usage{
check_plink_text_data(plink_text_data)
}
\value{
nothing
}
\description{
Check if the \code{PLINK} text data is valid.
Will \link{stop} if not.
}
\examples{
plink_text_data <- create_plink_text_data(
  ped_table = get_test_ped_table(),
  map_table = get_test_map_table()
)
check_plink_text_data(plink_text_data)
}
\author{
Richèl J.C. Bilderbeek
}
