% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plink_version.R
\name{check_plink_version}
\alias{check_plink_version}
\title{Check if the PLINK version is valid, will stop if not}
\usage{
check_plink_version(plink_version)
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the PLINK version.
Use \link{get_plink_versions} to get all the supported PLINK versions.}
}
\description{
Check if the PLINK version is valid, will stop if not
}
