% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/check_plink_version_and_data_can_work_together.R
\name{check_plink_version_and_data_can_work_together}
\alias{check_plink_version_and_data_can_work_together}
\title{Check that the \code{PLINK}/\code{PLINK2} version can work together with
the data (\code{PLINK} text data, \code{PLINK} binary data, \code{PLINK2} binary data).}
\usage{
check_plink_version_and_data_can_work_together(data, plink_options)
}
\description{
Check that the \code{PLINK}/\code{PLINK2} version can work together with
the data (\code{PLINK} text data, \code{PLINK} binary data, \code{PLINK2} binary data).
Will \link{stop} if not
}
\details{
\tabular{lll}{
   PLINK version \tab data \tab expected \cr
   PLINK \tab PLINK1 text data \tab OK \cr
   PLINK \tab PLINK1 bin data \tab OK \cr
   PLINK \tab PLINK2 bin data \tab Error \cr
   PLINK2 \tab PLINK1 text data \tab Error \cr
   PLINK2 \tab PLINK1 bin data \tab Error \cr
   PLINK2 \tab PLINK2 bin data \tab OK \cr
}
}
