% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_random_samples_selector.R
\name{check_random_samples_selector}
\alias{check_random_samples_selector}
\title{Check if a random-sample selector is valid.}
\usage{
check_random_samples_selector(random_samples_selector)
}
\arguments{
\item{random_samples_selector}{a sample selector
(see \link{create_samples_selector})
that allows one to select one or more random samples,
as created by \link{create_random_samples_selector}
and checked by \link{check_random_samples_selector}}
}
\value{
Nothing.
}
\description{
Check if a random-sample selector is valid.
Will \link{stop} if not.
}
\examples{
check_random_samples_selector(create_test_random_samples_selector())
}
\seealso{
There are multiple sample selector checking functions,
see \link{create_samples_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
