% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_single_snp_selector.R
\name{check_single_snp_selector}
\alias{check_single_snp_selector}
\title{Check if a single-SNP selector is valid.}
\usage{
check_single_snp_selector(single_snp_selector)
}
\arguments{
\item{single_snp_selector}{a SNP selector
(see \link{create_snp_selector} for all SNP selectors)
that allows one to select a single SNP,
as created by \link{create_single_snp_selector}
and checked by \link{check_single_snp_selector}}
}
\value{
Nothing.
}
\description{
Check if a single-SNP selector is valid.
Will \link{stop} if not.
}
\examples{
check_single_snp_selector(create_test_single_snp_selector())
}
\seealso{
There are multiple SNP selector checking functions,
see \link{create_snp_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
