% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_snp.R
\name{check_snp}
\alias{check_snp}
\title{Checks that the \code{snp} is valid}
\usage{
check_snp(snp)
}
\arguments{
\item{snp}{a SNP, e.g. \code{rs12345678},
as can be checked by \link{check_snp}}
}
\value{
nothing
}
\description{
Checks that the \code{snp} is valid.
Calls \link{stop} if not
}
\examples{
check_snp("rs1234")
}
\author{
Richèl J.C. Bilderbeek
}
