% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_snp_range_selector.R
\name{check_snp_range_selector}
\alias{check_snp_range_selector}
\title{Check if a SNP range selector is valid.}
\usage{
check_snp_range_selector(snp_range_selector)
}
\arguments{
\item{snp_range_selector}{a SNP selector (see \link{create_snps_selector}),
that allows one to select a range between two SNPs,
as created by \link{create_snp_range_selector}
and checked by \link{check_snp_range_selector}}
}
\value{
Nothing.
}
\description{
Check if a SNP range selector is valid.
Will \link{stop} if not.
}
\examples{
check_snp_range_selector(create_test_snp_range_selector())
}
\seealso{
There are multiple SNP selector checking functions,
see \link{create_snps_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
