% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_snp_selector.R
\name{check_snp_selector}
\alias{check_snp_selector}
\title{Check if SNP selector is valid.}
\usage{
check_snp_selector(snp_selector)
}
\arguments{
\item{snp_selector}{a SNP selector, a way to select one or more SNPs.
See \link{create_snp_selector} for all SNP selectors.}
}
\value{
Nothing.
}
\description{
Check if SNP selector is valid.
Will \link{stop} if not.
}
\examples{
check_snp_selector(create_test_random_snp_selector())
check_snp_selector(create_test_single_snp_selector())
check_snp_selector(create_test_snp_range_selector())
check_snp_selector(create_test_snp_window_selector())
}
\seealso{
There are multiple SNP selectors,
see \link{create_snp_selector} for an overview.

There are multiple SNP selector checking functions:
\itemize{
\item A random SNP selector: \link{check_random_snp_selector}
\item A single SNP selector: \link{check_single_snp_selector}
\item A SNP range selector: \link{check_snp_range_selector}
\item A SNP window selector: \link{check_snp_window_selector}
}
}
\author{
Richèl J.C. Bilderbeek
}
