% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_snp_window_selector.R
\name{check_snp_window_selector}
\alias{check_snp_window_selector}
\title{Check if SNP windows selector is valid.}
\usage{
check_snp_window_selector(snp_window_selector)
}
\value{
Nothing.
}
\description{
Check if SNP windows selector is valid.
Will \link{stop} if not.
}
\examples{
check_snp_window_selector(create_test_snp_window_selector())
}
\author{
Richèl J.C. Bilderbeek
}
