% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_verbose.R
\name{check_verbose}
\alias{check_verbose}
\title{Check if the \code{verbose} is valid.
Check if the \code{verbose} is valid,
which is when it is either \link{TRUE} or \link{FALSE}.}
\usage{
check_verbose(verbose)
}
\arguments{
\item{verbose}{set to TRUE for more output}
}
\value{
Nothing. Will \link{stop} if the input
is not either \link{TRUE} or \link{FALSE}.
}
\description{
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
