% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_pgen_table_to_bed_table.R
\name{convert_pgen_table_to_bed_table}
\alias{convert_pgen_table_to_bed_table}
\title{Convert a \code{PLINK2} \code{.pgen} table to a \code{PLINK} \code{.bed} table}
\usage{
convert_pgen_table_to_bed_table(pgen_table, verbose = FALSE)
}
\arguments{
\item{pgen_table}{an \link{array} that maps the individuals
to their SNPs, with as much rows as individuals, and as much
SNPs as columns. Optionally, the row names are the individuals' IDs,
where the column names are the SNP ID's}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{PLINK} \code{.bed} table
}
\description{
Convert a \code{PLINK2} \code{.pgen} table to a \code{PLINK} \code{.bed} table.
To do so, simply transpose a \code{.pgen} table,
see \url{https://github.com/chrchang/plink-ng/issues/195}
}
\author{
Richèl J.C. Bilderbeek
}
