% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_plink_bin_data_to_ped_table.R
\name{convert_plink_bin_data_to_ped_table}
\alias{convert_plink_bin_data_to_ped_table}
\title{Convert \code{PLINK} data to a \code{PLINK} \code{.ped} table}
\usage{
convert_plink_bin_data_to_ped_table(plink_bin_data, verbose = FALSE)
}
\arguments{
\item{plink_bin_data}{the in-memory binary data for \code{PLINK} to work on,
as created by \link{create_plink_bin_data}.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{PLINK} \code{.ped} table
}
\description{
Convert \code{PLINK} data to a \code{PLINK} \code{.ped} table
}
\author{
Richèl J.C. Bilderbeek
}
