% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_additive_trait.R
\name{create_additive_trait}
\alias{create_additive_trait}
\title{Create a trait in which the phenotype is additive
and the minor allele frequency is known.}
\usage{
create_additive_trait(
  mafs = 0.25,
  n_snps = 1,
  base_phenotype_value = 10,
  phenotype_increase = 0.5
)
}
\arguments{
\item{mafs}{one or more minor allele frequencies.
These allele frequencies must be ordered decreasingly,
i.e. the MAF is at the first position, where the
even rarer alleles are at the second and third positions.

Note that \code{PLINK} cannot handle triallelic nor
quadallelic SNPs: \code{PLINK} will give a warning that it
is setting the rarest alleles to missing.}

\item{n_snps}{the number of SNPs}

\item{base_phenotype_value}{the base phenotypic value for an additive trait,
i.e. the phenotypic value for homozygotes of the common allele}

\item{phenotype_increase}{the phenotypic value increase per rare allele,
for an additive trait.}
}
\description{
Create a trait in which the phenotype is additive
and the minor allele frequency is known.
The phenotype is fully and perfectly determined by the genotype,
see \link{calc_additive_phenotype_values} for the exact phenotypic values
}
\examples{
create_additive_trait()
create_additive_trait(mafs = 0.01)
create_additive_trait(n_snps = 2)
create_additive_trait(base_phenotype_value = 3.14)
create_additive_trait(phenotype_increase = 2.71)
}
\author{
Richèl J.C. Bilderbeek
}
