% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_qt_args.R
\name{create_assoc_qt_args}
\alias{create_assoc_qt_args}
\title{Create the command-line arguments to call PLINK or PLINK2
to do a quantitative trait analysis}
\usage{
create_assoc_qt_args(assoc_qt_params, plink_options = create_plink_options())
}
\arguments{
\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the command-line arguments
}
\description{
Create the command-line arguments to call PLINK or PLINK2
to do a quantitative trait analysis
}
\author{
Richèl J.C. Bilderbeek
}
