% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_qt_covar_params.R
\name{create_assoc_qt_covar_params}
\alias{create_assoc_qt_covar_params}
\title{Create parameters for the \link{assoc_qt_covar} function}
\usage{
create_assoc_qt_covar_params(
  ped_table,
  map_table,
  phe_table,
  cov_table,
  maf = get_lowest_maf(),
  base_input_filename = file.path(get_plinkr_tempfilename(), "assoc_input"),
  base_output_filename = file.path(get_plinkr_tempfilename(), "assoc_output")
)
}
\arguments{
\item{ped_table}{a 'pedigree' table.

A \code{ped_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item \code{within_family_id_father} Within-family ID of father
(\code{0} if father isn't in dataset)
\item \code{within_family_id_mother} Within-family ID of mother
(\code{0} if mother isn't in dataset)
\item \code{sex_code} Sex code
(\code{1} = male, \code{2} = female, \code{0} = unknown)
\item \code{case_control_code} Case control code
(\code{1} = control, \code{2} = case,
\code{9}/\code{0}/non-numeric = missing data if case/control)
\item \code{snv_[x][y]} Nucleotide for the \code{x}th variant
for haplotype \code{y} (\code{y} is either \code{a} or \code{b})
in the \code{.map file} (\code{0} = no call)
}

The \code{FID} and \code{IID} column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}.

Use \link{read_plink_ped_file} to read a \code{PLINK} \code{.ped} file.
Use \link{check_ped_table} to test if a pedigree table is valid.}

\item{map_table}{a genetic mapping table.

A \code{map_table} is a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{CHR}: the chromosome code or contig name.
The chromosome number cannot exceed 95
(\url{https://www.cog-genomics.org/plink/1.9/input#chr_set})
\item \code{SNP}: Variant identifier
\item \code{position_cm}: Position in morgans or centimorgans.
This value is optional. Zeroes denote it is unused
\item \code{BP}: Base-pair coordinat
}

Use \link{read_plink_map_file} to read a \code{PLINK} \code{.map} file.
Use \link{check_map_table} to test if a genetic mapping table is valid.}

\item{phe_table}{a table of phenotypes.

A \code{phe_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_phe_file} to read a phenotype file.
Use \link{check_phe_table} to test if a phenotype table is valid.}

\item{cov_table}{a table of covariates.

A \code{cov_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of covariate values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_cov_file} to read a covariates file.
Use \link{check_cov_table} to test if a covariates table is valid.}

\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the \code{PLINK} analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
\code{PLINK} \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}
}
\description{
Create parameters to detect associations with quantitative traits
and covariates
(using \link{assoc_qt_covar}).
Use \link{create_test_assoc_qt_covar_params})
to create a set of parameters
that is used in testing.
}
\note{
This function is named after the \code{--assoc} and
\code{--covar} \code{PLINK} flags.
}
\author{
Richèl J.C. Bilderbeek
}
