% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_qt_params.R
\name{create_assoc_qt_params}
\alias{create_assoc_qt_params}
\title{Create parameters for the \link{assoc_qt} function}
\usage{
create_assoc_qt_params(ped_table, map_table, phenotype_table, maf)
}
\arguments{
\item{ped_table}{a table obtained by reading a PLINK \code{.ped} file
Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.}

\item{map_table}{a table obtained by reading a PLINK \code{.map} file
Use \link{read_plink_map_file} to read a PLINK \code{.map} file.}

\item{phenotype_table}{a table of phenotypes.
Use \link{read_plink_phenotype_file} to read a phenotype file.}

\item{maf}{minor allele frequency}
}
\description{
Create parameters for the \link{assoc_qt} function
}
