% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_assoc_qt_params.R
\name{create_assoc_qt_params}
\alias{create_assoc_qt_params}
\title{Create parameters for the \link{assoc_qt} function}
\usage{
create_assoc_qt_params(
  data,
  phe_table,
  maf = get_lowest_maf(),
  base_input_filename = file.path(get_plinkr_tempfilename(), "assoc_input"),
  base_output_filename = file.path(get_plinkr_tempfilename(), "assoc_output")
)
}
\arguments{
\item{data}{PLINK or PLINK2 text or binary data:
\itemize{
\item PLINK text data:
a \code{.ped} table (see \link{read_plink_ped_file})
and a \code{.map} (see \link{read_plink_map_file}),
as created by \link{create_plink_text_data}.
\item PLINK binary data:
a \code{.bed} table (see \link{read_plink_bed_file})
a \code{.bim} table (see \link{read_plink_bim_file})
and a \code{.fam} (see \link{read_plink_fam_file}),
as created by \link{create_plink_bin_data}.
\item PLINK2 binary data: use \link{create_plink2_bin_data}
}}

\item{phe_table}{a table of phenotypes.

A \code{phe_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_phe_file} to read a phenotype file.
Use \link{check_phe_table} to test if a phenotype table is valid.}

\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the \code{PLINK} analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
\code{PLINK} \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}
}
\description{
Create parameters to detect associations with quantitative traits
(using \link{assoc_qt}).
Use \link{create_test_assoc_qt_params}) to create a set of parameters
that is used in testing.
}
\note{
This function is named after the \code{--assoc} \code{PLINK} flag.
Where \code{--PLINK} uses phenotypic values to determine if an association
is about a binary/case-control trait or a quantitative trait,
\link{plinkr} uses functions named \link{assoc} and \link{assoc_qt}.
}
\author{
Richèl J.C. Bilderbeek
}
