% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cov_table_from_data.R
\name{create_cov_table_from_data}
\alias{create_cov_table_from_data}
\title{Create a single-covariate table from pedigree (\code{.ped}) table}
\usage{
create_cov_table_from_data(data)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
}}
}
\value{
a single-covariate \link[tibble]{tibble}, with columns:
\itemize{
\item \code{FID} the family ID
\item \code{IID} the within-family ID
\item \code{C1} the covariate, which is zero by default
}

The column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#covno}
}
\description{
Create a single-covariate table from pedigree (\code{.ped}) table.
The individuals' family and within-family ID's are copied.
}
\examples{
create_cov_table_from_data(data = create_test_plink_text_data())
}
\author{
Richèl J.C. Bilderbeek
}
