% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_custom_trait.R
\name{create_custom_trait}
\alias{create_custom_trait}
\title{Create a custom trait in which the phenotype is
determined in a custom way
and the minor allele frequency is known.}
\usage{
create_custom_trait(
  phenotype = "custom",
  mafs = 0.25,
  n_snps = 1,
  calc_phenotype_function = calc_random_phenotype_values
)
}
\arguments{
\item{phenotype}{one phenotype, named after its genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive
(the nucleotides used are A and T as these are in the word 'additive'):
\itemize{
\item \code{AA} 11.0
\item \code{AT} 10.5
\item \code{TT} 10.0
}
}}

\item{mafs}{one or more minor allele frequencies.
These allele frequencies must be ordered decreasingly,
i.e. the MAF is at the first position, where the
even rarer alleles are at the second and third positions.

Note that \code{PLINK} cannot handle triallelic nor
quadallelic SNPs: \code{PLINK} will give a warning that it
is setting the rarest alleles to missing.}

\item{n_snps}{the number of SNPs}

\item{calc_phenotype_function}{a function that calculate the phenotypes
from genotypes. The input is the genetic data as a \link[tibble]{tibble},
in which each row is an individual and the columns are the SNVs.
The first two columns are named \code{snv_1a}, \code{snv_1b} and
hold the genetic data for the first SNV of a diploid organism.
If there are more SNVs, columns continue with names,
\code{snv_2a}, \code{snv_2b}, \code{snv_3a}, \code{snv_3b}, etc.
Nucleotides are in uppercase.
The output must be the phenotypic values, as a numeric vector,
which has the same length as the number of individuals.
Use \link{check_calc_phenotype_function}
to check a \code{calc_phenotype_function}.}
}
\description{
Create a custom trait in which the phenotype is
determined in a custom way
and the minor allele frequency is known.
}
\examples{
# Create a custom trait that calculates the phenotypes randomly,
# regardless of the genotype.
# This is the same as using 'create_random_trait'
create_custom_trait(
  phenotype = "random",
  calc_phenotype_function = calc_random_phenotype_values
)

# Create a custom trait that calculates the phenotypes additively
# This is the same as using 'create_additive_trait'
create_custom_trait(
  phenotype = "additive",
  calc_phenotype_function = calc_additive_phenotype_values
)

# Create a custom trait that calculates the phenotypes from
# an epistatic interaction
# This is the same as using 'create_epistate_trait'
create_custom_trait(
  phenotype = "epistatic",
  calc_phenotype_function = calc_epistatic_phenotype_values,
  n_snps = 2
)

# A trivial trait, that shows the genotypes worked on
create_custom_trait(
  phenotype = "debug",
  calc_phenotype_function = function(snvs) {

    # Show the input
    message(paste0(knitr::kable(snvs), collapse = "\n"))

    # Return as much 1s as individuals
    rep(1, nrow(snvs))
  }
)

}
\author{
Richèl J.C. Bilderbeek
}
