% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_assoc_qt_params.R
\name{create_demo_assoc_qt_params}
\alias{create_demo_assoc_qt_params}
\title{Create a set of \code{assoc_qt} parameters
to demonstrate PLINKs \code{--assoc_qt} functionality}
\usage{
create_demo_assoc_qt_params(n_snps = 1, phenotypes = "random", maf = 0.05)
}
\arguments{
\item{n_snps}{the number of SNPs}

\item{phenotypes}{one ore more phenotypes,
named after their genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive
(the nucleotides used are A and T as these are in the word 'additive'):
\itemize{
\item \code{AA} 1.0
\item \code{AT} 0.5
\item \code{TT} 0.0
}
}}

\item{maf}{minor allele frequency}
}
\description{
This function creates:
\itemize{
\item Individuals that have all combinations of SNPs exactly once,
as created by the \link{create_demo_ped_table} function
\item All SNPs that are on different chromosomes
as created by the \link{create_demo_map_table} function
\item Traits that follow different genotype to phenotype mapping:
\itemize{
\item \code{control}: a trait that is random
\item \code{additive}: a trait that is perfectly additive,
each \code{A} denotes an increase of 0.1,
each \code{C} denotes an increase of 0.2,
each \code{G} denotes an increase of 0.3,
each \code{T} denotes an increase of 0.4
}
}
}
\note{
This function is named after the \code{--assoc-qt} PLINK flag.
}
