% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_assoc_qt_params.R
\name{create_demo_assoc_qt_params}
\alias{create_demo_assoc_qt_params}
\title{Create a set of \code{assoc_qt} parameters
to demonstrate PLINKs \code{--assoc_qt} functionality}
\usage{
create_demo_assoc_qt_params(
  n_individuals = 4,
  phenotypes = get_phenotypes(),
  mafs = rep(0.25, length(phenotypes)),
  maf = get_lowest_maf()
)
}
\arguments{
\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}

\item{phenotypes}{one ore more phenotypes,
named after their genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive
(the nucleotides used are A and T as these are in the word 'additive'):
\itemize{
\item \code{AA} 1.0
\item \code{AT} 0.5
\item \code{TT} 0.0
}
}}

\item{mafs}{one or more minor allele frequencies}

\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the PLINK analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
PLINK \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}
}
\description{
This function creates:
\itemize{
\item Individuals that have all combinations of SNPs exactly once,
as created by the \link{create_demo_ped_table} function
\item All SNPs that are on different chromosomes
as created by the \link{create_demo_map_table} function
\item Traits that follow different genotype to phenotype mapping:
\itemize{
\item \code{control}: a trait that is random
\item \code{additive}: a trait that is perfectly additive.
The trait value is \code{10.0},
where each \code{T} increases it by 0.5
}
}
}
\note{
This function is named after the \code{--assoc-qt} PLINK flag.
}
\author{
Richèl J.C. Bilderbeek
}
