% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_assoc_qt_params.R
\name{create_demo_assoc_qt_params}
\alias{create_demo_assoc_qt_params}
\title{Create a set of \code{assoc_qt} parameters
to demonstrate PLINKs \code{--assoc_qt} functionality}
\usage{
create_demo_assoc_qt_params(n_individuals = 4, traits = create_demo_traits())
}
\arguments{
\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}

\item{traits}{one or more traits
with a clear genetic architecture and a known minor allele frequency,
as, for example, created by \link{create_demo_traits}.}
}
\description{
This function creates:
\itemize{
\item Individuals that have all combinations of SNPs exactly once,
as created by the \link{create_demo_ped_table} function
\item All SNPs that are on different chromosomes
as created by the \link{create_demo_map_table} function
\item Traits that follow different genotype to phenotype mapping:
\itemize{
\item \code{control}: a trait that is random
\item \code{additive}: a trait that is perfectly additive.
The trait value is \code{10.0},
where each \code{T} increases it by 0.5
}
}
}
\note{
This function is named after the \code{--assoc-qt} PLINK flag.
}
\examples{
# Default
create_demo_assoc_qt_params()

# Add more individuals
create_demo_assoc_qt_params(n_individuals = 1000)

# Use a random trait
create_demo_assoc_qt_params(
  traits = create_random_trait()
)

# Use an additive trait
create_demo_assoc_qt_params(
  traits = create_additive_trait()
)

# Use an additive and random trait
create_demo_assoc_qt_params(
  traits = list(create_additive_trait(), create_random_trait())
)

# Use three random traits
create_demo_assoc_qt_params(
  traits = rep(list(create_random_trait()), 3)
)

# Use two additive traits with different minor allele frequencies
create_demo_assoc_qt_params(
  traits = list(
    create_additive_trait(maf = 0.01),
    create_additive_trait(maf = 0.10)
  )
)
}
\author{
Richèl J.C. Bilderbeek
}
