% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_assoc_qt_params.R
\name{create_demo_assoc_qt_params}
\alias{create_demo_assoc_qt_params}
\title{Create a set of \code{assoc_qt} parameters
to demonstrate PLINKs \code{--assoc_qt} functionality}
\usage{
create_demo_assoc_qt_params()
}
\description{
This function creates:
\itemize{
\item Individuals that have all combinations of SNPs exactly once,
as created by the \link{create_demo_ped_table} function
\item All SNPs that are on different chromosomes
as created by the \link{create_demo_map_table} function
\item Traits that follow different genotype to phenotype mapping:
\itemize{
\item \code{control}: a trait that is random
\item \code{additive}: a trait that is perfectly additive,
each \code{A} denotes an increase of 0.1,
each \code{C} denotes an increase of 0.2,
each \code{G} denotes an increase of 0.3,
each \code{T} denotes an increase of 0.4
}
}
}
\note{
This function is named after the \code{--assoc-qt} PLINK flag.
}
