% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_map_table.R
\name{create_demo_map_table}
\alias{create_demo_map_table}
\title{Create a demo genetic mapping \code{.map} table}
\usage{
create_demo_map_table(traits = create_demo_traits())
}
\arguments{
\item{traits}{one or more traits
with a clear genetic architecture and a known minor allele frequency,
as, for example, created by \link{create_demo_traits}.}
}
\description{
Create a demo genetic mapping \code{.map} table.
It is simply a map of SNPs that are located on different chromosomes.
}
\details{
Each SNP is positioned on a different chromosome.
As PLINK allows for a maximum of 95 chromosomes,
the 96th SNP is placed on the first chromosome again,
after which the chromosomes are travered again.
}
\examples{
# Default demonstration genetic mapping table
create_demo_map_table()

# Random, 1 SNP
create_demo_map_table(traits = create_random_trait())

# Random, 2 SNPs
create_demo_map_table(traits = create_random_trait(n_snps = 2))
}
\author{
Richèl J.C. Bilderbeek
}
