% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_ped_table.R
\name{create_demo_ped_table}
\alias{create_demo_ped_table}
\title{Create a demo pedigree \code{.ped} table}
\usage{
create_demo_ped_table(n_snps = 1)
}
\arguments{
\item{n_snps}{the number of SNPs}
}
\description{
Create a demo pedigree \code{.ped} table,
which should allow the individuals to have all combinations of
biallelic SNPs exactly once.
As we assume two nucleotides and a diploid organism,
these are 4 individuals.
}
