% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_demo_ped_table.R
\name{create_demo_ped_table}
\alias{create_demo_ped_table}
\title{Create a demo pedigree \code{.ped} table}
\usage{
create_demo_ped_table(n_individuals = 4, phenotypes = get_phenotypes())
}
\arguments{
\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}

\item{phenotypes}{one ore more phenotypes,
named after their genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive
(the nucleotides used are A and T as these are in the word 'additive'):
\itemize{
\item \code{AA} 1.0
\item \code{AT} 0.5
\item \code{TT} 0.0
}
}}
}
\description{
Create a demo pedigree \code{.ped} table,
which should allow the individuals to have all combinations of
biallelic SNPs exactly once.
As we assume two nucleotides and a diploid organism,
these are 4 individuals.
}
\examples{
create_demo_ped_table()
}
\author{
Richèl J.C. Bilderbeek
}
