% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epistatic_trait.R
\name{create_epistatic_trait}
\alias{create_epistatic_trait}
\title{Create a trait in which the phenotype is epistatic
and the minor allele frequency is known.}
\usage{
create_epistatic_trait(
  mafs = 0.25,
  n_snps = 2,
  regular_phenotype_value = 20,
  epistatic_phenotype_value = 21
)
}
\arguments{
\item{mafs}{one or more minor allele frequencies.
These allele frequencies must be ordered decreasingly,
i.e. the MAF is at the first position, where the
even rarer alleles are at the second and third positions.

Note that \code{PLINK} cannot handle triallelic nor
quadallelic SNPs: \code{PLINK} will give a warning that it
is setting the rarest alleles to missing.}

\item{n_snps}{the number of SNPs}

\item{regular_phenotype_value}{the regular phenotypic value}

\item{epistatic_phenotype_value}{the phenotypic value when the
epistatic phenotype is expressed}
}
\description{
Create a trait in which the phenotype is epistatic
(i.e. determined by multiple SNPs)
and the minor allele frequency is known.
}
\details{
The phenotype is fully and perfectly determined by the genotype,
see \link{calc_epistatic_phenotype_values} for the exact phenotypic values
}
\examples{
create_epistatic_trait()
create_epistatic_trait(maf = 0.01)
create_epistatic_trait(n_snps = 3)
create_epistatic_trait(regular_phenotype_value = 3.14)
create_epistatic_trait(epistatic_phenotype_value = 2.71)
}
\author{
Richèl J.C. Bilderbeek
}
