% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phenotype_table_from_ped_table.R
\name{create_phenotype_table_from_ped_table}
\alias{create_phenotype_table_from_ped_table}
\title{Create a single-phenotype table from a \code{.ped} table}
\usage{
create_phenotype_table_from_ped_table(ped_table)
}
\arguments{
\item{ped_table}{a table obtained by reading a PLINK \code{.ped} file
Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.}
}
\value{
a single-phenotype \link[tibble]{tibble}, with columns:
\itemize{
\item \code{family_id}
\item \code{within_family_id}
\item \code{.phenotype_value}
}
}
\description{
Create a single-phenotype table from a \code{.ped} table
}
