% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plink_optionses.R
\name{create_plink_optionses}
\alias{create_plink_optionses}
\title{Create the PLINK options for all supported versions of PLINK}
\usage{
create_plink_optionses(plink_folder = get_plink_folder(), os = get_os())
}
\arguments{
\item{plink_folder}{folder where \code{PLINK} is installed}

\item{os}{name of the operating system,
as returned by \link[rappdirs]{app_dir}}
}
\value{
a list of multiple PLINK options
}
\description{
Create the PLINK options for all supported versions of PLINK.
}
\details{
Use \link{get_plink_versions} to get all the supported version
as strings
}
\examples{
create_plink_optionses()
create_plink_optionses(plink_folder = get_plink_folder())
}
\author{
Richèl J.C. Bilderbeek
}
