% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plink_v1_7_options.R
\name{create_plink_v1_7_options}
\alias{create_plink_v1_7_options}
\title{Create the options to run PLINK v1.7}
\usage{
create_plink_v1_7_options(plink_folder = get_plink_folder(), os = get_os())
}
\arguments{
\item{plink_folder}{folder where \code{PLINK} is installed}

\item{os}{name of the operating system,
as returned by \link[rappdirs]{app_dir}}
}
\value{
the PLINK options,
as can be checked by \link{check_plink_options}
}
\description{
Create the options to run PLINK v1.7
}
\examples{
create_plink_v1_7_options()
}
\author{
Richèl J.C. Bilderbeek
}
