% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_trait.R
\name{create_random_trait}
\alias{create_random_trait}
\title{Create a trait with no relation between genotype and phenotype
and a known minor allele frequency.}
\usage{
create_random_trait(mafs = 0.25)
}
\arguments{
\item{mafs}{one or more minor allele frequencies}
}
\description{
Create a trait with no relation between genotype and phenotype
and a known minor allele frequency.
}
\examples{
create_random_trait()
create_random_trait(maf = 0.01)
}
\author{
Richèl J.C. Bilderbeek
}
