% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_snp_window_selector.R
\name{create_snp_window_selector}
\alias{create_snp_window_selector}
\title{Create a SNP window selector}
\usage{
create_snp_window_selector(snp, window_kb)
}
\arguments{
\item{snp}{a SNP, e.g. \code{rs12345678},
as can be checked by \link{check_snp}}

\item{window_kb}{the window (i.e. the amount of base pairs around
a focal SNP) in kilobases, as checked by \link{check_window_kb}}
}
\description{
Create a SNP window selector,
to be used in, for example, \link{select_snps}
}
\examples{
create_snp_window_selector(
  snp = "my_snp",
  window_kb = 12.34
)

if (is_plink_installed()) {

  plink_bin_filenames <- create_plink_bin_filenames(
    bed_filename = get_plinkr_filename("select_snps.bed"),
    bim_filename = get_plinkr_filename("select_snps.bim"),
    fam_filename = get_plinkr_filename("select_snps.fam")
  )

  # Selects a window of 3, i.e. the variant before,
  # and after the focal SNP
  snp_window_selector <- create_snp_window_selector(
    snp = "snp_5",
    window_kb = 0.003
  )
  new_plink_bin_data <- select_snps(
    data = plink_bin_filenames,
    snp_selector = snp_window_selector
  )
}
}
\seealso{
Use \link{create_test_snp_window_selector} to
create a SNP window selector to be used in testing.

There are multiple SNP selectors,
see \link{create_snps_selector} for an overview
}
\author{
Richèl J.C. Bilderbeek
}
