% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_assoc_qt_params.R
\name{create_test_assoc_qt_params}
\alias{create_test_assoc_qt_params}
\title{Create testing parameters for the \link{assoc_qt} function}
\usage{
create_test_assoc_qt_params(
  ped_table = get_test_ped_table(),
  map_table = get_test_map_table(),
  phenotype_table = plinkr::create_phenotype_table_from_ped_table(ped_table =
    ped_table)
)
}
\arguments{
\item{ped_table}{a table obtained by reading a PLINK \code{.ped} file
Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.}

\item{map_table}{a table obtained by reading a PLINK \code{.map} file
Use \link{read_plink_map_file} to read a PLINK \code{.map} file.}

\item{phenotype_table}{a table of phenotypes.
Use \link{read_plink_phenotype_file} to read a phenotype file.}
}
\description{
Create parameters (as created by \link{create_assoc_qt_params})
to detect associations with quantitative traits
(using \link{assoc_qt}) used for testing.
}
\details{
This function is named after the \code{--assoc-qt} PLINK flag.
}
