% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_assoc_qt_params.R
\name{create_test_assoc_qt_params}
\alias{create_test_assoc_qt_params}
\title{Create testing parameters for the \link{assoc_qt} function}
\usage{
create_test_assoc_qt_params(
  ped_table = get_test_ped_table(),
  map_table = get_test_map_table(),
  phenotype_table = plinkr::create_phenotype_table_from_ped_table(ped_table =
    ped_table),
  maf = get_lowest_maf()
)
}
\arguments{
\item{ped_table}{a 'pedigree' table.

A \code{ped_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{family_id} The family ID (\code{FID})
\item \code{within_family_id} Within-family ID
(\code{IID}, cannot be zero)
\item \code{within_family_id_father} Within-family ID of father
(\code{0} if father isn't in dataset)
\item \code{within_family_id_mother} Within-family ID of mother
(\code{0} if mother isn't in dataset)
\item \code{sex_code} Sex code
(\code{1} = male, \code{2} = female, \code{0} = unknown)
\item \code{case_control_code} Case control code
(\code{1} = control, \code{2} = case,
\code{9}/\code{0}/non-numeric = missing data if case/control)
\item \code{snv_[x][y]} Nucleotide for the \code{x}th variant
for haplotype \code{y} (\code{y} is either \code{a} or \code{b})
in the \code{.map file} (\code{0} = no call)
}

Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.
Use \link{check_ped_table} to test if a pedigree table is valid.}

\item{map_table}{a genetic mapping table.

A \code{map_table} is a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{CHR}: the chromosome code or contig name
\item \code{SNP}: Variant identifier
\item \code{position_cm}: Position in morgans or centimorgans.
This value is optional. Zeroes denote it is unused
\item \code{BP}: Base-pair coordinat
}

Use \link{read_plink_map_file} to read a PLINK \code{.map} file.
Use \link{check_map_table} to test if a genetic mapping table is valid.}

\item{phenotype_table}{a table of phenotypes.

A \code{phenotype_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{family_id} The family ID (\code{FID})
\item \code{within_family_id} Within-family ID
(\code{IID}, cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

Use \link{read_plink_phenotype_file} to read a phenotype file.
Use \link{check_phenotype_table} to test if a phenotype table is valid.}

\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the PLINK analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_af}

The parameter name \code{maf} is named after the
PLINK \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}
}
\description{
Create parameters (as created by \link{create_assoc_qt_params})
to detect associations with quantitative traits
(using \link{assoc_qt}) used for testing.
}
\note{
This function is named after the \code{--assoc-qt} PLINK flag.
}
