% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_snp_range_selector.R
\name{create_test_snp_range_selector}
\alias{create_test_snp_range_selector}
\title{Create a SNP range selector}
\usage{
create_test_snp_range_selector(snp_from = "rs123456", snp_to = "rs234567")
}
\description{
Create a SNP range selector,
to be used in, for example, \link{select_snps}
}
\examples{
create_test_snp_range_selector()

create_test_snp_range_selector(
  snp_from = "rs123456",
  snp_to = "rs234567"
)
}
\seealso{
Use \link{create_test_snp_range_selector} to
create a SNP window selector to be used in testing.

There are multiple SNP selectors:
\itemize{
\item a single SNP: see \link{create_single_snp_selector}
\item a SNP range: see \link{create_snp_range_selector}
\item a window around a SNP: see \link{create_snp_window_selector}
}
}
\author{
Richèl J.C. Bilderbeek
}
