% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trait.R
\name{create_trait}
\alias{create_trait}
\title{Create a trait with a clear genetic architecture and a known
minor allele frequency.}
\usage{
create_trait(phenotype, mafs = 0.25)
}
\arguments{
\item{phenotype}{one phenotype, named after its genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive
(the nucleotides used are A and T as these are in the word 'additive'):
\itemize{
\item \code{AA} 11.0
\item \code{AT} 10.5
\item \code{TT} 10.0
}
}}

\item{mafs}{one or more minor allele frequencies}
}
\description{
Create a trait with a clear genetic architecture and a known
minor allele frequency.
}
\examples{
create_trait(phenotype = "random")
create_trait(phenotype = "random", maf = 0.01)
}
\author{
Richèl J.C. Bilderbeek
}
