% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trait.R
\name{create_trait}
\alias{create_trait}
\title{Create a trait with a clear genetic architecture and a known
minor allele frequency.}
\usage{
create_trait(
  phenotype,
  mafs = 0.25,
  n_snps = 1,
  calc_phenotype_function = calc_random_phenotype_values
)
}
\arguments{
\item{phenotype}{one phenotype, named after its genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive
(the nucleotides used are A and T as these are in the word 'additive'):
\itemize{
\item \code{AA} 11.0
\item \code{AT} 10.5
\item \code{TT} 10.0
}
}}

\item{mafs}{one or more minor allele frequencies.
These allele frequencies must be ordered decreasingly,
i.e. the MAF is at the first position, where the
even rarer alleles are at the second and third positions.

Note that \code{PLINK} cannot handle triallelic nor
quadallelic SNPs: \code{PLINK} will give a warning that it
is setting the rarest alleles to missing.}

\item{n_snps}{the number of SNPs,
as can be checked by \link{check_n_snps}}

\item{calc_phenotype_function}{a function that calculate the phenotypes
from genotypes. The input is the genetic data as a \link[tibble]{tibble},
in which each row is an individual and the columns are the SNVs.
The first two columns are named \code{snv_1a}, \code{snv_1b} and
hold the genetic data for the first SNV of a diploid organism.
If there are more SNVs, columns continue with names,
\code{snv_2a}, \code{snv_2b}, \code{snv_3a}, \code{snv_3b}, etc.
Nucleotides are in uppercase.
The output must be the phenotypic values, as a numeric vector,
which has the same length as the number of individuals.
Use \link{check_calc_phenotype_function}
to check a \code{calc_phenotype_function}.}
}
\description{
Create a trait with a clear genetic architecture and a known
minor allele frequency.
}
\details{
Prefer to use the specialized functions:
\itemize{
\item \link{create_additive_trait}
\item \link{create_custom_trait}
\item \link{create_epistatic_trait}
\item \link{create_random_trait}
\item \link{create_random_case_control_trait}
}
}
\examples{
create_trait(phenotype = "random")
create_trait(phenotype = "random", maf = 0.01)
create_trait(phenotype = "random", n_snps = 2)
create_trait(
  phenotype = "random",
  calc_phenotype_function = calc_random_phenotype_values
)
}
\author{
Richèl J.C. Bilderbeek
}
