% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trait.R
\name{create_trait}
\alias{create_trait}
\title{Create a trait with a clear genetic architecture and a known
minor allele frequency.}
\usage{
create_trait(phenotype, mafs = 0.25, n_snps = 1)
}
\arguments{
\item{phenotype}{one phenotype, named after its genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive
(the nucleotides used are A and T as these are in the word 'additive'):
\itemize{
\item \code{AA} 11.0
\item \code{AT} 10.5
\item \code{TT} 10.0
}
}}

\item{mafs}{one or more minor allele frequencies.
These allele frequencies must be ordered decreasingly,
i.e. the MAF is at the first position, where the
even rarer alleles are at the second and third positions.

Note that \code{PLINK} cannot handle triallelic nor
quadallelic SNPs: \code{PLINK} will give a warning that it
is setting the rarest alleles to missing.}

\item{n_snps}{the number of SNPs}
}
\description{
Create a trait with a clear genetic architecture and a known
minor allele frequency.
}
\examples{
create_trait(phenotype = "random")
create_trait(phenotype = "random", maf = 0.01)
create_trait(phenotype = "random", n_snps = 2)
}
\author{
Richèl J.C. Bilderbeek
}
