% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  add_noweb,
  assoc_filename,
  assoc_params,
  assoc_qt_params,
  example_filename,
  log_filename,
  maf,
  mafs,
  map_filename,
  map_table,
  n_individuals,
  n_snps,
  ped_filename,
  ped_table,
  phenotype,
  phenotype_filename,
  phenotype_table,
  phenotypes,
  plink_folder,
  plink_version,
  plink_versions,
  qassoc_filename,
  sim_filename,
  simfreq_filename,
  simulate_qt_params,
  snvs,
  temp_sim_filename,
  trait,
  traits,
  url,
  verbose
)
}
\arguments{
\item{add_noweb}{add the \code{--noweb} option, which prevents
that PLINK will check if it is in its latest version.
By default, this value is set to true,
hence plinkr will skip this check.
Use \link{get_plink_version} to get the current PLINK version.}

\item{assoc_filename}{name of a PLINK \code{.assoc} file
Use \link{read_plink_assoc_file} to read a PLINK \code{.assoc} file.}

\item{assoc_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc}),
as can be created by \link{create_assoc_params}}

\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}

\item{example_filename}{name of the example file}

\item{log_filename}{name of a PLINK \code{.log} file}

\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the PLINK analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
PLINK \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}

\item{mafs}{one or more minor allele frequencies.
These allele frequencies must be ordered decreasingly,
i.e. the MAF is at the first position, where the
even rarer alleles are at the second and third positions.

Note that \code{PLINK} cannot handle triallelic nor
quadallelic SNPs: \code{PLINK} will give a warning that it
is setting the rarest alleles to missing.}

\item{map_filename}{name of a PLINK \code{.map} file
Use \link{read_plink_map_file} to read a \code{PLINK} \code{.map} file.}

\item{map_table}{a genetic mapping table.

A \code{map_table} is a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{CHR}: the chromosome code or contig name
\item \code{SNP}: Variant identifier
\item \code{position_cm}: Position in morgans or centimorgans.
This value is optional. Zeroes denote it is unused
\item \code{BP}: Base-pair coordinat
}

Use \link{read_plink_map_file} to read a PLINK \code{.map} file.
Use \link{check_map_table} to test if a genetic mapping table is valid.}

\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}

\item{n_snps}{the number of SNPs}

\item{ped_filename}{name of a PLINK \code{.ped} file.
Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.}

\item{ped_table}{a 'pedigree' table.

A \code{ped_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{family_id} The family ID (\code{FID})
\item \code{within_family_id} Within-family ID
(\code{IID}, cannot be zero)
\item \code{within_family_id_father} Within-family ID of father
(\code{0} if father isn't in dataset)
\item \code{within_family_id_mother} Within-family ID of mother
(\code{0} if mother isn't in dataset)
\item \code{sex_code} Sex code
(\code{1} = male, \code{2} = female, \code{0} = unknown)
\item \code{case_control_code} Case control code
(\code{1} = control, \code{2} = case,
\code{9}/\code{0}/non-numeric = missing data if case/control)
\item \code{snv_[x][y]} Nucleotide for the \code{x}th variant
for haplotype \code{y} (\code{y} is either \code{a} or \code{b})
in the \code{.map file} (\code{0} = no call)
}

Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.
Use \link{check_ped_table} to test if a pedigree table is valid.}

\item{phenotype}{one phenotype, named after its genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive
(the nucleotides used are A and T as these are in the word 'additive'):
\itemize{
\item \code{AA} 11.0
\item \code{AT} 10.5
\item \code{TT} 10.0
}
}}

\item{phenotype_filename}{name of a raw phenotype file.
Use \link{read_plink_phenotype_file} to read a raw phenotype file.}

\item{phenotype_table}{a table of phenotypes.

A \code{phenotype_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{family_id} The family ID (\code{FID})
\item \code{within_family_id} Within-family ID
(\code{IID}, cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

Use \link{read_plink_phenotype_file} to read a phenotype file.
Use \link{check_phenotype_table} to test if a phenotype table is valid.}

\item{phenotypes}{one ore more phenotypes,
named after their genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive
(the nucleotides used are A and T as these are in the word 'additive'):
\itemize{
\item \code{AA} 11.0
\item \code{AT} 10.5
\item \code{TT} 10.0
}
}}

\item{plink_folder}{folder where PLINK is installed}

\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the PLINK version.
Use \link{get_plink_versions} to get all the supported PLINK versions.}

\item{plink_versions}{one or more versions of PLINK,
e.g. as can be obtained using \link{get_plink_versions}}

\item{qassoc_filename}{name of a PLINK \code{.qassoc} file
Use \link{read_plink_qassoc_file} to read a PLINK \code{.qassoc} file.}

\item{sim_filename}{name of a PLINK \code{.sim} file}

\item{simfreq_filename}{name of a PLINK \code{.simfreq} file}

\item{simulate_qt_params}{the parameters for a quantitative
traits simulation, as can be created by
\link{create_simulate_qt_params}}

\item{snvs}{a \link[tibble]{tibble} that contains the
two nucleotide calls for multiple individuals.
Each column is a haplotypes, hence, for a diploid organism,
there are two columns.
Each individual is represented by a row.
Nucleotides are in uppercase.
Per SNV table, there can be only two different nucleotides,
as this is how a SNP works; a SNP from A to C is a different one (and
has a different SNP ID) than a SNP that encodes a mutation
for A to G.}

\item{temp_sim_filename}{temporary file to store simulation
parameters, which is a PLINK \code{.sim} file}

\item{trait}{one trait with a clear genetic architecture and a known
minor allele frequency, as created by \link{create_trait}.
Use \link{is_one_trait} to detect if something is one trait}

\item{traits}{one or more traits
with a clear genetic architecture and a known minor allele frequency,
as, for example, created by \link{create_demo_traits}.}

\item{url}{PLINK download URL}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
