% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  add_noweb,
  allow_no_sex,
  args,
  assoc_adjusted_filename,
  assoc_data,
  assoc_filename,
  assoc_params,
  assoc_qt_covar_data,
  assoc_qt_covar_params,
  assoc_qt_data,
  assoc_qt_params,
  assoc_qt_result,
  assoc_table,
  base_input_filename,
  base_output_filename,
  base_output_plink1_filename,
  base_output_plink2_filename,
  base_phenotype_value,
  bed_filename,
  bed_table,
  bfile,
  bim_filename,
  bim_table,
  calc_phenotype_function,
  chromosome_number,
  chromosome_selector,
  confidence_interval,
  cov_filename,
  cov_table,
  data,
  epistatic_phenotype_value,
  example_filename,
  fam_filename,
  fam_table,
  fid,
  fids,
  frq_filename,
  frq_strat_filename,
  iid,
  iids,
  ld_filename,
  ld_window_r2,
  imiss_filename,
  lmiss_filename,
  log_filename,
  maf,
  mafs,
  map_filename,
  map_table,
  n_individuals,
  n_phenotypes,
  n_samples,
  n_snps,
  n_snps_per_phenotype,
  os,
  out,
  ped_filename,
  ped_table,
  pgen_filename,
  pgen_table,
  phenotype,
  phe_filename,
  phenotype_increase,
  phe_table,
  phenotypes,
  phenotype_data,
  phenotype_data_filename,
  phenotype_data_table,
  phenotype_value_dominant,
  phenotype_value_recessive,
  plink_bin_data,
  plink_bin_filenames,
  plink_exe_path,
  plink_filenames,
  plink_folder,
  plink_options,
  plink1_options,
  plink2_options,
  plink_optionses,
  plink_text_data,
  plink_text_filenames,
  plink_version,
  plink_versions,
  plinkr_folder,
  plink2_bin_data,
  plink2_bin_filenames,
  psam_filename,
  psam_table,
  pvar_filename,
  pvar_table,
  qassoc_filename,
  qassoc_filenames,
  qassoc_table,
  random_samples_selector,
  random_snps_selector,
  regular_phenotype_value,
  sample_ids,
  sample_ids_filename,
  sample_selector,
  sim_filename,
  simfreq_filename,
  simulate_qt_params,
  single_sample_selector,
  single_snp_selector,
  snp,
  snp_from,
  snp_range_selector,
  snp_selector,
  snp_to,
  snp_window_selector,
  snvs,
  temp_folder,
  temp_sim_filename,
  trait,
  traits,
  url,
  verbose,
  window_kb
)
}
\arguments{
\item{add_noweb}{add the \code{--noweb} option, which prevents
that \code{PLINK} will check if it is in its latest version.
By default, this value is set to true,
hence \link{plinkr} will skip this check.
Use \link{get_plink_version} to get the current \code{PLINK} version.}

\item{allow_no_sex}{set this to \link{TRUE}
to retain phenotype values for missing-sex samples.
This parameter is named after the \code{PLINK} \code{--allow-no-sex} flag}

\item{args}{arguments for the \code{PLINK} or \code{PLINK2} executable.
Use \link{check_plink_args} to check if these are valid}

\item{assoc_adjusted_filename}{name of a \code{PLINK}
\code{.assoc.adjusted} file.
Use \link{read_plink_assoc_adjusted_file}
to read a \code{PLINK} \code{.assoc.adjusted} file.}

\item{assoc_data}{data to use for an association analysis
for a quantitative trait (i.e. using \link{assoc}),
as can be created by \link{create_assoc_data}}

\item{assoc_filename}{name of a \code{PLINK} \code{.assoc} file.
Use \link{read_plink_assoc_file} to read a \code{PLINK} \code{.assoc} file.}

\item{assoc_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc}),
as can be created by \link{create_assoc_params}}

\item{assoc_qt_covar_data}{data to do an association analysis
for a quantitative trait with covariates
(i.e. using \link{assoc_qt_covar}),
as can be created by \link{create_assoc_qt_covar_data}}

\item{assoc_qt_covar_params}{parameters to do an association analysis
for a quantitative trait with covariates
(i.e. using \link{assoc_qt_covar}),
as can be created by \link{create_assoc_qt_covar_params}}

\item{assoc_qt_data}{data use to used by \link{assoc_qt}
(as created by \link{create_assoc_qt_data}),
which equals 'regular' data (see \link{check_data}) and a
phenotype table/file}

\item{assoc_qt_params}{parameters to do an association analysis
for a quantitative trait (i.e. using \link{assoc_qt}),
as can be created by \link{create_assoc_qt_params}}

\item{assoc_qt_result}{the result of \link{assoc_qt},
which is a \link{list} with elements:
\itemize{
\item \code{qassoc_table}: the quantitative analysis results table,
as can be checked by \link{check_qassoc_table}
\item \code{log}: the text from the log file created by \code{PLINK}/\code{PLINK2}
when doing \link{assoc_qt}
}}

\item{assoc_table}{the table that hold the result of a
case-control association, as created by \code{PLINK}/\code{PLINK2}
and stored as a \code{.assoc} file (for \code{PLINK}) or
\verb{[basename].<phenotype>.[?TODO].linear} (for \code{PLINK2}).}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{base_output_plink1_filename}{temporary folder to create the \code{PLINK}
binary files in}

\item{base_output_plink2_filename}{temporary folder to create the \code{PLINK2}
binary files in}

\item{base_phenotype_value}{the base phenotypic value for an additive trait,
i.e. the phenotypic value for homozygotes of the common allele}

\item{bed_filename}{name of a \code{PLINK} \code{.bed} file
Use \link{read_plink_bed_file} to read a \code{PLINK} \code{.bed} file.}

\item{bed_table}{a table that maps the SNPs to the individuals,
of which the column names are the names of the individuals,
the row names are the names of the SNPs,
and the values are the SNP variant.
Use \link{get_test_bed_table} to get a \code{.bed} table as used in testing.
Use \link{read_plink_bed_file} to read a \code{PLINK} \code{.bed} file.
Use \link{check_bed_table} to test if a \code{.bed} table is valid.
See also the \code{bed} file format reference at
\url{https://www.cog-genomics.org/plink2/formats#bed}}

\item{bfile}{the base filename of the binary files (i.e.
a \code{.bed}, \code{.bim} and \code{.fam} file).
This parameter is named after the \code{PLINK}
\code{--bfile} flag}

\item{bim_filename}{name of a \code{PLINK} \code{.bim} file
Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.}

\item{bim_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{chr}: the chromosome number
\item \code{id}: the SNP ID
\item \code{posg}: the position
\item \code{pos}: the position
\item \code{ref}: something
\item \code{alt}: something
}

Use \link{get_test_bim_table} to get a \code{.bim} table as used in testing.
Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.
Use \link{check_bim_table} to test if a \code{.bim} table is valid.}

\item{calc_phenotype_function}{a function that calculate the phenotypes
from genotypes. The input is the genetic data as a \link[tibble]{tibble},
in which each row is an individual and the columns are the SNVs.
The first two columns are named \code{snv_1a}, \code{snv_1b} and
hold the genetic data for the first SNV of a diploid organism.
If there are more SNVs, columns continue with names,
\code{snv_2a}, \code{snv_2b}, \code{snv_3a}, \code{snv_3b}, etc.
Nucleotides are in uppercase.
The output must be the phenotypic values, as a numeric vector,
which has the same length as the number of individuals.
Use \link{check_calc_phenotype_function}
to check a \code{calc_phenotype_function}.}

\item{chromosome_number}{the chromosome number,
as can be checked by \link{check_chromosome_number}}

\item{chromosome_selector}{a SNP selector (see \link{create_snps_selector})
that allows one to select a single chromsome,
as created by \link{create_chromosome_selector}
and checked by \link{check_chromosome_selector}}

\item{confidence_interval}{confidence interval, a value between (and
excluding both) 0.0 and 1.0. The confidence interval helps assess
the certainty of an estimation: you can be 99 percent sure
a value is within the range of the 0.99 confidence interval}

\item{cov_filename}{name of a covariates (\code{.cov}) file.
Use \link{read_plink_cov_file} to read a covariates file.}

\item{cov_table}{a table of covariates.

A \code{cov_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of covariate values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_cov_file} to read a covariates file.
Use \link{check_cov_table} to test if a covariates table is valid.}

\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
\item data to be used by \link{assoc},
as created by \link{create_assoc_data}
\item data to be used by \link{assoc_qt},
as created by \link{create_assoc_qt_data}
}}

\item{epistatic_phenotype_value}{the phenotypic value when the
epistatic phenotype is expressed}

\item{example_filename}{name of the example file}

\item{fam_filename}{name of a \code{PLINK} \code{.fam} file
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.}

\item{fam_table}{a tibble of the genetic mapping,
with as many rows as SNPs.

Each row contains:
\itemize{
\item \code{fam}: the family ID
\item \code{id}: the individual's ID
\item \code{pat}: ID of father
\item \code{mat}: ID of mother
\item \code{sex}: the gender
\item \code{pheno}: a phenotype
}

Use \link{get_test_fam_table} to get a \code{.fam} table as used in testing.
Use \link{read_plink_fam_file} to read a \code{PLINK} \code{.fam} file.
Use \link{check_fam_table} to test if a \code{.fam} table is valid.}

\item{fid}{the family ID, which is called \code{FID} in \code{PLINK}.
Use \link{check_fid} to check if a \code{fid} is valid.}

\item{fids}{one or more family IDs (which are called \code{FID}s in \code{PLINK}).
Use \link{check_fids} to check if the elements of \code{fids} are valid.}

\item{frq_filename}{name of a \code{PLINK} \code{.frq} file
Use \link{read_plink_frq_file} to read a \code{PLINK} \code{.frq} file.}

\item{frq_strat_filename}{name of a \code{PLINK} \code{.frq.strat} file
Use \link{read_plink_frq_strat_file}
to read a \code{PLINK} \code{.frq.strat} file.}

\item{iid}{a within-family ID, as can be checked by \link{check_iid}}

\item{iids}{one or more within-family IDs,
as can be checked by \link{check_iids}}

\item{ld_filename}{name of a \code{PLINK}/\code{PLINK2} linkage
disequilibrium (\code{.ld}) file
Use \link{read_plink_ld_file} to read a \code{PLINK} \code{.ld} file.}

\item{ld_window_r2}{the minimal squared correlation coefficient (also
known as \code{r^2}) between markers. This parameter is named after the
\code{--ld_window_r2} flag, as documented at
\url{https://www.cog-genomics.org/plink/1.9/ld}.}

\item{imiss_filename}{name of a \code{PLINK} \code{.imiss} file
Use \link{read_plink_imiss_file} to read a \code{PLINK} \code{.imiss} file.}

\item{lmiss_filename}{name of a \code{PLINK} \code{.lmiss} file
Use \link{read_plink_lmiss_file} to read a \code{PLINK} \code{.lmiss} file.}

\item{log_filename}{name of a \code{PLINK} \code{.log} file}

\item{maf}{minor allele frequency threshold.
Alleles that have a frequency lower than the MAF
are excluded from the \code{PLINK} analysis.
\code{maf} must be a value between zero and 0.5
(i.e. excluding zero and excluding 0.5).
By default, \code{maf} is set to the lowest
representable non-zero floating-point value,
as obtained by \link{get_lowest_maf}

The parameter name \code{maf} is named after the
\code{PLINK} \code{--maf} flag. This was chosen over
more specific names such as \code{min_allele_frequency}).}

\item{mafs}{one or more minor allele frequencies.
These allele frequencies must be ordered decreasingly,
i.e. the MAF is at the first position, where the
even rarer alleles are at the second and third positions.

Note that \code{PLINK} cannot handle triallelic nor
quadallelic SNPs: \code{PLINK} will give a warning that it
is setting the rarest alleles to missing.}

\item{map_filename}{name of a \code{PLINK} \code{.map} file
Use \link{read_plink_map_file} to read a \code{PLINK} \code{.map} file.}

\item{map_table}{a genetic mapping table.

A \code{map_table} is a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{CHR}: the chromosome code or contig name.
The chromosome number cannot exceed 95
(\url{https://www.cog-genomics.org/plink/1.9/input#chr_set})
\item \code{SNP}: Variant identifier
\item \code{position_cm}: Position in morgans or centimorgans.
This value is optional. Zeroes denote it is unused
\item \code{BP}: Base-pair coordinat
}

Use \link{read_plink_map_file} to read a \code{PLINK} \code{.map} file.
Use \link{check_map_table} to test if a genetic mapping table is valid.}

\item{n_individuals}{the number of individuals.
Use \link{check_n_individuals} to check if this is a valid value}

\item{n_phenotypes}{the number of phenotypes}

\item{n_samples}{the number of samples/individuals,
as can be checked by \link{check_n_samples}}

\item{n_snps}{the number of SNPs,
as can be checked by \link{check_n_snps}}

\item{n_snps_per_phenotype}{the number of SNPs that determine one phenotype}

\item{os}{name of the operating system,
as returned by \link[rappdirs]{app_dir}}

\item{out}{the base filename of the output files.
This parameter is named after the \code{PLINK}
\code{--out} flag}

\item{ped_filename}{name of a \code{PLINK} \code{.ped} file.
Use \link{read_plink_ped_file} to read a \code{PLINK} \code{.ped} file.}

\item{ped_table}{a 'pedigree' table.

A \code{ped_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item \code{within_family_id_father} Within-family ID of father
(\code{0} if father isn't in dataset)
\item \code{within_family_id_mother} Within-family ID of mother
(\code{0} if mother isn't in dataset)
\item \code{sex_code} Sex code
(\code{1} = male, \code{2} = female, \code{0} = unknown)
\item \code{case_control_code} Case control code
(\code{1} = control, \code{2} = case,
\code{9}/\code{0}/non-numeric = missing data if case/control)
\item \code{snv_[x][y]} Nucleotide for the \code{x}th variant
for haplotype \code{y} (\code{y} is either \code{a} or \code{b})
in the \code{.map file} (\code{0} = no call)
}

The \code{FID} and \code{IID} column names match the PLINK names, see
\url{https://www.cog-genomics.org/plink/1.9/input#pheno}.

Use \link{read_plink_ped_file} to read a \code{PLINK} \code{.ped} file.
Use \link{check_ped_table} to test if a pedigree table is valid.}

\item{pgen_filename}{name of a \code{PLINK2} \code{.pgen} file
Use \link{read_plink2_pgen_file} to read a \code{PLINK2} \code{.pgen} file}

\item{pgen_table}{an \link{array} that maps the individuals
to their SNPs, with as much rows as individuals, and as much
SNPs as columns. Optionally, the row names are the individuals' IDs,
where the column names are the SNP ID's}

\item{phenotype}{one phenotype, named after its genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive
(the nucleotides used are A and T as these are in the word 'additive'):
\itemize{
\item \code{AA} 11.0
\item \code{AT} 10.5
\item \code{TT} 10.0
}
}}

\item{phe_filename}{name of a phenotype (\code{.phe}) file,
as checked by \link{check_phe_filename}.
Use \link{read_plink_phe_file} to read a raw phenotype file.}

\item{phenotype_increase}{the phenotypic value increase per rare allele,
for an additive trait.}

\item{phe_table}{a table of phenotypes.

A \code{phe_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID (cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

These names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_phe_file} to read a phenotype file.
Use \link{check_phe_table} to test if a phenotype table is valid.}

\item{phenotypes}{one ore more phenotypes,
named after their genetic background:
\itemize{
\item \code{random} the phenotype is a random value,
i.e. there is no association between the genetics and
this phenotype
\item \code{additive} the phenotype is perfectly additive,
see \link{calc_additive_phenotype_values} for the exact calculation
\item \code{epistatic} the phenotype is epistatic,
see \link{calc_epistatic_phenotype_values} for the exact calculation
}}

\item{phenotype_data}{phenotype data in the form of either
a filename to a phenotype file (\code{phenotype_data_filename})
or a list with a phenotype table (\code{phenotype_data_table})}

\item{phenotype_data_filename}{phenotype data
(as checked by \link{check_phenotype_data}))
in the form of a \link{list} with a filename to a phenotype table}

\item{phenotype_data_table}{phenotype data
(as checked by \link{check_phenotype_data}))
in the form of a \link{list} with an in-memory phenotype table}

\item{phenotype_value_dominant}{phenotypic value
for the dominant variant, i.e. that genotype that has at least one
version of the common common allele}

\item{phenotype_value_recessive}{phenotypic value
for the recessive variant, i.e. that genotype that is homozygous for the
rare allele}

\item{plink_bin_data}{the in-memory binary data for \code{PLINK} to work on,
as created by \link{create_plink_bin_data}.}

\item{plink_bin_filenames}{the binary data files' names for \code{PLINK}
to work on,
as created by \link{create_plink_bin_filenames}.}

\item{plink_exe_path}{path to
the \code{PLINK} or \code{PLINK2} executable file.}

\item{plink_filenames}{a list of filenames that
is a collection of \code{PLINK} text files, \code{PLINK} binary files
or \code{PLINK2} binary files. For example, for \code{PLINK} text files,
this list has elements \code{map_filename} and \code{ped_filename}}

\item{plink_folder}{folder where \code{PLINK} is installed}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{plink1_options}{the \code{PLINK} version,
as created by \link{create_plink_options},
for any version of \code{PLINK}}

\item{plink2_options}{the \code{PLINK2} version,
as created by \link{create_plink_options},
for any version of \code{PLINK2}}

\item{plink_optionses}{a list of one or more options to run PLINK,
as created by \link{create_plink_optionses}.
The reduplicated plural was used to express this is a list
of \code{plink_options}, instead of one set of \code{plink_options}}

\item{plink_text_data}{the genetic and phenotypic data for \code{PLINK} to work
on, in \code{PLINK} text format, as created by \link{create_plink_text_data}.}

\item{plink_text_filenames}{the text data files' names for \code{PLINK}
to work on,
as created by \link{create_plink_text_filenames}.}

\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the \code{PLINK} version.
Use \link{get_plink_versions} to get all the supported \code{PLINK} versions.}

\item{plink_versions}{one or more versions of PLINK,
e.g. as can be obtained using \link{get_plink_versions}}

\item{plinkr_folder}{name of the folder where \link{plinkr}
stores its temporary files}

\item{plink2_bin_data}{the genetic and phenotypic data for \code{PLINK2} to work
on, in \code{PLINK2} binary format, as created by \link{create_plink2_bin_data}.}

\item{plink2_bin_filenames}{the binary data files' names for \code{PLINK2}
to work on,
as created by \link{create_plink2_bin_filenames}.}

\item{psam_filename}{name of a \code{PLINK2} \code{.psam} file
Use \link{read_plink2_psam_file} to read a \code{PLINK2} \code{.psam} file}

\item{psam_table}{a \link[tibble]{tibble} with as much rows as
individuals. It has the following columns:
\itemize{
\item \code{FID}: the family ID
\item \code{IID}: the within-family ID
\item \code{SEX}: the sex
\item \code{PHENO1}: a phenotypic value
}

The uppercase column names are those as used by PLINK2.}

\item{pvar_filename}{name of a \code{PLINK2} \code{.pvar} file
Use \link{read_plink2_pvar_file} to read a \code{PLINK2} \code{.pvar} file}

\item{pvar_table}{a \link[tibble]{tibble} with as much rows as
SNPs. It has the following columns:
\itemize{
\item \code{CHROM}: the chromosome
\item \code{POS}: the position on the chromosome
\item \code{ID}: the SNP ID
\item \code{REF}: the reference variant
\item \code{ALT}: the alternate variant
}

The uppercase column names are those as used by PLINK2.}

\item{qassoc_filename}{name of a \code{PLINK} \code{.qassoc} file
Use \link{read_plink_qassoc_file} to read a \code{PLINK} \code{.qassoc} file.}

\item{qassoc_filenames}{name of one or more
\code{PLINK} \code{.qassoc} files.
Use \link{read_plink_qassoc_files} to read one or more
\code{PLINK} \code{.qassoc} files.}

\item{qassoc_table}{the table that hold the result of an association
with a quantitative trait, as created by \code{PLINK}/\code{PLINK2}
and stored as a \code{.qassoc} file (for \code{PLINK}) or
\verb{[basename].<phenotype>.glm.linear} (for \code{PLINK2}).}

\item{random_samples_selector}{a sample selector
(see \link{create_samples_selector})
that allows one to select one or more random samples,
as created by \link{create_random_samples_selector}
and checked by \link{check_random_samples_selector}}

\item{random_snps_selector}{a SNP selector
(see \link{create_snps_selector})
that allows one to select one or more random SNPs,
as created by \link{create_random_snps_selector}
and checked by \link{check_random_snps_selector}}

\item{regular_phenotype_value}{the regular phenotypic value}

\item{sample_ids}{sample IDs, which is a \link[tibble]{tibble}
with two columns: the first column holds the family ID
(called \code{fid} or \code{FID} by PLINK/PLINK2),
the second column holds the within-family ID
(called \code{iid} or \code{IID} by PLINK/PLINK2),
as can be checked by \link{check_sample_ids}}

\item{sample_ids_filename}{name of a file to store \code{sample_ids}
(see \link{check_sample_ids}) to}

\item{sample_selector}{a sample/individual selector,
a way to select one or more samples/individuals.
See \link{create_samples_selector} for all sample selectors.}

\item{sim_filename}{name of a \code{PLINK} \code{.sim} file}

\item{simfreq_filename}{name of a \code{PLINK} \code{.simfreq} file}

\item{simulate_qt_params}{the parameters for a quantitative
traits simulation, as can be created by
\link{create_simulate_qt_params}}

\item{single_sample_selector}{a sample selector
(see \link{create_samples_selector} for all sample selectors)
that allows one to select a single sample,
as created by \link{create_single_sample_selector}
and checked by \link{check_single_sample_selector}}

\item{single_snp_selector}{a SNP selector
(see \link{create_snps_selector} for all SNP selectors)
that allows one to select a single SNP,
as created by \link{create_single_snp_selector}
and checked by \link{check_single_snp_selector}}

\item{snp}{a SNP, e.g. \code{rs12345678},
as can be checked by \link{check_snp}}

\item{snp_from}{a SNP (see \link{check_snp}) to start from}

\item{snp_range_selector}{a SNP selector (see \link{create_snps_selector}),
that allows one to select a range between two SNPs,
as created by \link{create_snp_range_selector}
and checked by \link{check_snp_range_selector}}

\item{snp_selector}{a SNP selector, a way to select one or more SNPs.
See \link{create_snps_selector} for all SNP selectors.}

\item{snp_to}{a SNP (see \link{check_snp}) to end at}

\item{snp_window_selector}{a SNP selector (see \link{create_snps_selector}),
that allows one to select a focal SNP and a range/window of SNPs around it,
as created by \link{create_snp_window_selector}
and checked by \link{check_snp_window_selector}}

\item{snvs}{a \link[tibble]{tibble} that contains the
two nucleotide calls for multiple individuals.
Each column is a haplotype, hence, for a diploid organism,
there are two columns.
Each individual is represented by a row.
Nucleotides are in uppercase.
Per SNV table, there can be only two different nucleotides,
as this is how a SNP works; a SNP from A to C is a different one (and
has a different SNP ID) than a SNP that encodes a mutation
for A to G.
Use \link{create_snvs} to create a \code{snvs}.}

\item{temp_folder}{temporary folder to store results in}

\item{temp_sim_filename}{temporary file to store simulation
parameters, which is a \code{PLINK} \code{.sim} file}

\item{trait}{one trait with a clear genetic architecture and a known
minor allele frequency, as created by \link{create_trait}.
Use \link{is_one_trait} to detect if something is one trait}

\item{traits}{one or more traits
with a clear genetic architecture and a known minor allele frequency,
as, for example, created by \link{create_demo_traits}.}

\item{url}{\code{PLINK} download URL}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}

\item{window_kb}{the window (i.e. the amount of base pairs around
a focal SNP) in kilobases, as checked by \link{check_window_kb}}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
