% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_plink_self_test.R
\name{do_plink_self_test}
\alias{do_plink_self_test}
\title{Do a minimal PLINK self test}
\usage{
do_plink_self_test(plink_folder = get_plink_folder())
}
\arguments{
\item{plink_folder}{folder where PLINK is installed}
}
\value{
nothing. Will stop with an error message if the self-test fails
}
\description{
Do a minimal PLINK self test
}
