% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_plink_self_test.R
\name{do_plink_self_test}
\alias{do_plink_self_test}
\title{Do a minimal PLINK self test}
\usage{
do_plink_self_test(plink_options = create_plink_options(), verbose = FALSE)
}
\arguments{
\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
nothing. Will stop with an error message if the self-test fails
}
\description{
Do a minimal PLINK self test
}
\author{
Richèl J.C. Bilderbeek
}
