% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_plink_self_test.R
\name{do_plink_self_test}
\alias{do_plink_self_test}
\title{Do a minimal PLINK self test}
\usage{
do_plink_self_test(
  plink_version = get_default_plink_version(),
  plink_folder = get_plink_folder(),
  add_noweb = TRUE
)
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the PLINK version.
Use \link{get_plink_versions} to get all the supported PLINK versions.}

\item{plink_folder}{folder where PLINK is installed}

\item{add_noweb}{add the \code{--noweb} option, which prevents
that PLINK will check if it is in its latest version.
By default, this value is set to true,
hence plinkr will skip this check.
Use \link{get_plink_version} to get the current PLINK version.}
}
\value{
nothing. Will stop with an error message if the self-test fails
}
\description{
Do a minimal PLINK self test
}
\author{
Richèl J.C. Bilderbeek
}
