% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_download_url.R
\name{get_plink_download_url}
\alias{get_plink_download_url}
\title{Get the URL to download a version of PLINK from}
\usage{
get_plink_download_url(
  plink_version = get_default_plink_version(),
  os = get_os()
)
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the \code{PLINK} version.
Use \link{get_plink_versions} to get all the supported \code{PLINK} versions.}

\item{os}{name of the operating system,
as returned by \link[rappdirs]{app_dir}}
}
\value{
a URL
}
\description{
\itemize{
\item PLINK v1.9: \url{https://www.cog-genomics.org/plink2/}
}
}
\author{
Richèl J.C. Bilderbeek
}
