% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_example_filename.R
\name{get_plink_example_filename}
\alias{get_plink_example_filename}
\title{Get the path to a PLINK example file}
\usage{
get_plink_example_filename(
  example_filename,
  plink_options = create_plink_options()
)
}
\arguments{
\item{example_filename}{name of the example file}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
the path to a PLINK example file, if it exists.
Will \link{stop} if the file does not exist
}
\description{
Get the path to a PLINK example file
}
\examples{
if (is_plink_installed()) {
  get_plink_example_filename("toy.map")
  get_plink_example_filename("toy.ped")
}

# v.17
plink_options <- create_plink_v1_7_options()
if (is_plink_installed(plink_options)) {
  get_plink_example_filename("test.map", plink_options)
  get_plink_example_filename("test.ped", plink_options)
}

# v1.9
plink_options <- create_plink_v1_9_options()
if (is_plink_installed(plink_options)) {
  get_plink_example_filename("toy.map", plink_options)
  get_plink_example_filename("toy.ped", plink_options)
}
}
\author{
Richèl J.C. Bilderbeek
}
