% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_example_filename.R
\name{get_plink_example_filename}
\alias{get_plink_example_filename}
\title{Get the path to a PLINK example file}
\usage{
get_plink_example_filename(
  example_filename,
  plink_version = get_default_plink_version(),
  plink_folder = get_plink_folder()
)
}
\arguments{
\item{example_filename}{name of the example file}

\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the PLINK version.
Use \link{get_plink_versions} to get all the supported PLINK versions.}

\item{plink_folder}{folder where PLINK is installed}
}
\value{
the path to a PLINK example file, if it exists.
Will \link{stop} if the file does not exist
}
\description{
Get the path to a PLINK example file
}
\examples{
if (is_plink_installed(plink_version = "1.7")) {
  get_plink_example_filename("test.map", plink_version = "1.7")
  get_plink_example_filename("test.ped", plink_version = "1.7")
}
if (is_plink_installed(plink_version = "1.9")) {
  get_plink_example_filename("toy.map", plink_version = "1.9")
  get_plink_example_filename("toy.ped", plink_version = "1.9")
}
}
