% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_example_filenames.R
\name{get_plink_example_filenames}
\alias{get_plink_example_filenames}
\title{Get the full paths to all PLINK example files}
\usage{
get_plink_example_filenames(
  plink_version = get_default_plink_version(),
  plink_folder = get_plink_folder()
)
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}}

\item{plink_folder}{folder where PLINK is installed}
}
\value{
the full paths to all PLINK example files
}
\description{
Get the full paths to all PLINK example files
}
\examples{
if (is_plink_installed()) {
  get_plink_example_filenames()
}
}
