% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_exe_path.R
\name{get_plink_exe_path}
\alias{get_plink_exe_path}
\title{Get the path to the default PLINK executable as used by plinkr}
\usage{
get_plink_exe_path(plink_options = create_plink_options())
}
\value{
path to the default PLINK executable  as used by plinkr
}
\description{
Get the path to the default PLINK executable as used by plinkr
}
\author{
Richèl J.C. Bilderbeek
}
