% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_help_text.R
\name{get_plink_help_text}
\alias{get_plink_help_text}
\title{Get the text of the PLINK help file}
\usage{
get_plink_help_text(
  plink_version = get_default_plink_version(),
  plink_folder = get_plink_folder()
)
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}}

\item{plink_folder}{folder where PLINK is installed}
}
\value{
the text of the PLINK help file
}
\description{
Get the text of the PLINK help file
}
