% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_version.R
\name{get_plink_version}
\alias{get_plink_version}
\title{Get the version of PLINK}
\usage{
get_plink_version(
  plink_version = get_default_plink_version(),
  plink_folder = get_plink_folder()
)
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the PLINK version.
Use \link{get_plink_versions} to get all the supported PLINK versions.}

\item{plink_folder}{folder where PLINK is installed}
}
\value{
the version of PLINK
}
\description{
Get the version of PLINK
}
