% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plink_versions.R
\name{get_plink_versions}
\alias{get_plink_versions}
\title{Get the supported version of PLINK}
\usage{
get_plink_versions()
}
\description{
Get the supported version of PLINK
}
\author{
Richèl J.C. Bilderbeek
}
