% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plinkr_filenames.R
\name{get_plinkr_filenames}
\alias{get_plinkr_filenames}
\title{Get the full paths to all plinkr files}
\usage{
get_plinkr_filenames()
}
\value{
the full paths to all plinkr files
}
\description{
Get the full paths to all plinkr files
}
\examples{
get_plinkr_filenames()
}
