% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_map_table.R
\name{get_test_map_table}
\alias{get_test_map_table}
\title{Get a testing mapping (\code{.map}) table}
\usage{
get_test_map_table()
}
\value{
a mapping (\code{.map}) table
}
\description{
Get a testing mapping (\code{.map}) table
}
