% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_ped_table.R
\name{get_test_ped_table}
\alias{get_test_ped_table}
\title{Get a testing pedigree (\code{.ped}) table}
\usage{
get_test_ped_table()
}
\value{
a pedigree (\code{.ped}) table
}
\description{
Get a testing pedigree (\code{.ped}) table
}
