% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_plink.R
\name{install_plink}
\alias{install_plink}
\title{Determine if PLINK is installed}
\usage{
install_plink(
  url = "http://zzz.bwh.harvard.edu/plink/dist/plink-1.07-x86_64.zip",
  plink_folder = get_plink_folder()
)
}
\arguments{
\item{url}{PLINK download URL}

\item{plink_folder}{folder where PLINK is installed}
}
\value{
Nothing
}
\description{
Determine if PLINK is installed
}
