% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_plink.R
\name{install_plink}
\alias{install_plink}
\title{Install a specific version of PLINK}
\usage{
install_plink(
  plink_version = get_default_plink_version(),
  plink_folder = get_plink_folder()
)
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the PLINK version.
Use \link{get_plink_versions} to get all the supported PLINK versions.}

\item{plink_folder}{folder where PLINK is installed}
}
\value{
Nothing
}
\description{
Install a specific version of PLINK
}
