% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_plinks.R
\name{install_plinks}
\alias{install_plinks}
\title{Install multiple versions of PLINK}
\usage{
install_plinks(
  plink_versions = get_plink_versions(),
  plink_folder = get_plink_folder()
)
}
\arguments{
\item{plink_versions}{one or more versions of PLINK,
e.g. as can be obtained using \link{get_plink_versions}}

\item{plink_folder}{folder where PLINK is installed}
}
\description{
Install multiple versions of PLINK
}
