% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_ci.R
\name{is_on_github_actions}
\alias{is_on_github_actions}
\title{Determines if the environment is GitHub Actions}
\usage{
is_on_github_actions()
}
\value{
\link{TRUE} if run on GitHub Actions, \link{FALSE} otherwise
}
\description{
Determines if the environment is GitHub Actions
}
\examples{
  if (is_on_github_actions()) {
    message("Running on GitHub Actions")
  }
}
\author{
Richèl J.C. Bilderbeek
}
