% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plink2_bin_data.R
\name{is_plink2_bin_data}
\alias{is_plink2_bin_data}
\title{Determine if the \code{plink2_bin_data} is indeed \code{PLINK2} binary data}
\usage{
is_plink2_bin_data(plink2_bin_data)
}
\arguments{
\item{plink2_bin_data}{the genetic and phenotypic data for \code{PLINK2} to work
on, in \code{PLINK2} binary format, as created by \link{create_plink2_bin_data}.}
}
\value{
\link{TRUE} if the \code{plink2_bin_data} is indeed \code{PLINK2} binary data
}
\description{
Determine if the \code{plink2_bin_data} is indeed \code{PLINK2} binary data
}
\examples{
plink2_bin_data <- create_plink2_bin_data(
  pgen_table = get_test_pgen_table(),
  psam_table = get_test_psam_table(),
  pvar_table = get_test_pvar_table()
)
is_plink2_bin_data(plink2_bin_data)
}
\seealso{
\code{plinkr} has multiple functions to test if a type of data is valid:
\itemize{
\item \code{PLINK} text data: use \link{is_plink_text_data}
\item \code{PLINK} binary data: use \link{is_plink_bin_data}
\item \code{PLINK2} binary data: use \link{is_plink2_bin_data}
\item any type of \code{PLINK}/\code{PLINK2} data: use \link{is_data}
\item associative trait analysis data: use \link{is_assoc_qt_data}
}
}
\author{
Richèl J.C. Bilderbeek
}
