% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plink_installed.R
\name{is_plink_installed}
\alias{is_plink_installed}
\title{Determine if \code{PLINK} is installed}
\usage{
is_plink_installed(plink_options = create_plink_options())
}
\arguments{
\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
\code{TRUE} if \code{PLINK} is installed, \code{FALSE} otherwise
}
\description{
Determine if \code{PLINK} is installed
}
\author{
Richèl J.C. Bilderbeek
}
