% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plink_installed.R
\name{is_plink_installed}
\alias{is_plink_installed}
\title{Determine if PLINK is installed}
\usage{
is_plink_installed(
  plink_version = get_default_plink_version(),
  plink_folder = get_plink_folder()
)
}
\arguments{
\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the PLINK version.
Use \link{get_plink_versions} to get all the supported PLINK versions.}

\item{plink_folder}{folder where PLINK is installed}
}
\value{
TRUE if PLINK is installed, FALSE otherwise
}
\description{
Determine if PLINK is installed
}
\author{
Richèl J.C. Bilderbeek
}
