% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_plink_text_data.R
\name{is_plink_text_data}
\alias{is_plink_text_data}
\title{Determine if the \code{plink_text_data} is indeed \code{PLINK} text data}
\usage{
is_plink_text_data(plink_text_data)
}
\description{
Determine if the \code{plink_text_data} is indeed \code{PLINK} text data
}
\examples{
plink_text_data <- create_plink_text_data(
  ped_table = get_test_ped_table(),
  map_table = get_test_map_table()
)
is_plink_text_data(plink_text_data)
}
\author{
Richèl J.C. Bilderbeek
}
