% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_random_snp_selector.R
\name{is_random_snp_selector}
\alias{is_random_snp_selector}
\title{Determine if this is a random SNP selector}
\usage{
is_random_snp_selector(random_snp_selector, verbose = FALSE)
}
\arguments{
\item{random_snp_selector}{a SNP selector (see \link{create_snp_selector})
that allows one to select one or more random SNPs,
as created by \link{create_random_snp_selector}
and checked by \link{check_random_snp_selector}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if this is a random SNP selector
}
\description{
Determine if this is a random SNP selector
}
\examples{
is_random_snp_selector(create_test_random_snp_selector())

is_random_snp_selector("nonsense")
}
\author{
Richèl J.C. Bilderbeek
}
