% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\title{Filter \code{PLINK} and \code{PLINK2} data on their family ID}
\usage{
keep(
  base_input_filename,
  sample_ids_filename,
  plink_options = create_plink_v1_9_options(),
  temp_folder = plinkr::get_plinkr_tempfilename(),
  verbose = FALSE
)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{sample_ids_filename}{name of a file to store \code{sample_ids}
(see \link{check_sample_ids}) to}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{temp_folder}{temporary folder to store results in}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
the names of the files created
}
\description{
Filter \code{PLINK} and \code{PLINK2} data on their family ID
}
\note{
this function is named after the \code{PLINK}/\code{PLINK2} \code{--keep}
parameter flag
}
\author{
Richèl J.C. Bilderbeek
}
