% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\title{Filter \code{PLINK} and \code{PLINK2} data on their family ID}
\usage{
keep(
  base_input_filename,
  sample_ids,
  plink_options = create_plink_v1_9_options(),
  temp_folder = plinkr::get_plinkr_tempfilename(),
  sample_ids_filename = file.path(temp_folder, "sample_ids.txt"),
  verbose = FALSE
)
}
\arguments{
\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{sample_ids}{sample IDs, which is a \link[tibble]{tibble}
with two columns: the first column holds the family ID
(called \code{fid} or \code{FID} by PLINK/PLINK2),
the second column holds the within-family ID
(called \code{iid} or \code{IID} by PLINK/PLINK2),
as can be checked by \link{check_sample_ids}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{temp_folder}{temporary folder to store results in}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
the names of the files created
}
\description{
Filter \code{PLINK} and \code{PLINK2} data on their family ID
}
\note{
this function is named after the \code{PLINK}/\code{PLINK2} \code{--Keep}
parameter flag
}
\author{
Richèl J.C. Bilderbeek
}
