% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkr_report.R
\name{plinkr_report}
\alias{plinkr_report}
\title{Create a \link{plinkr} report, to be used when reporting bugs}
\usage{
plinkr_report(plink_optionses = plinkr::create_plink_optionses())
}
\description{
Create a \link{plinkr} report, to be used when reporting bugs
}
\examples{
plinkr_report()
}
\author{
Richèl J.C. Bilderbeek
}
