% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink2_pgen_file.R
\name{read_plink2_pgen_file}
\alias{read_plink2_pgen_file}
\title{Read a \code{PLINK2} \code{.pgen} file}
\usage{
read_plink2_pgen_file(pgen_filename, verbose = FALSE)
}
\arguments{
\item{pgen_filename}{name of a \code{PLINK2} \code{.pgen} file
Use \link{read_plink2_pgen_file} to read a \code{PLINK2} \code{.pgen} file}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Read a \code{PLINK2} \code{.pgen} file
}
\note{
this function is inspired by
\url{https://github.com/simingz/ctwas/blob/main/R/ctwas_read_data.R#L112}
}
\examples{
pgen_filename <- get_plinkr_filename(
  "toy_v1_9_after_make-bed_after_make-pgen.pgen"
)
read_plink2_pgen_file(pgen_filename)
}
\author{
Richèl J.C. Bilderbeek
}
