% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_frq_file.R
\name{read_plink_frq_file}
\alias{read_plink_frq_file}
\title{Read a PLINK \code{.frq} file}
\usage{
read_plink_frq_file(frq_filename)
}
\arguments{
\item{frq_filename}{name of a \code{PLINK} \code{.frq} file
Use \link{read_plink_frq_file} to read a \code{PLINK} \code{.frq} file.}
}
\value{
a \link[tibble]{tibble} with column names:
\itemize{
\item \code{CHR} ?
\item \code{SNP} ?
\item \code{A1} ?
\item \code{A2} ?
\item \code{MAF} ?
\item \code{NCHROBS} ?
}

These column names are the names used by \code{PLINK}.
}
\description{
Read a PLINK \code{.frq} file
}
\examples{
frq_filename <- get_plinkr_filename("miss_stat.frq")
read_plink_frq_file(frq_filename)
}
\author{
Richèl J.C. Bilderbeek
}
