% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_log_file.R
\name{read_plink_log_file}
\alias{read_plink_log_file}
\title{Read a PLINK log file}
\usage{
read_plink_log_file(log_filename)
}
\arguments{
\item{log_filename}{name of a PLINK \code{.log} file}
}
\description{
Read a PLINK log file
}
