% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_map_file.R
\name{read_plink_map_file}
\alias{read_plink_map_file}
\title{Read a PLINK \code{.map} file}
\usage{
read_plink_map_file(map_filename)
}
\arguments{
\item{map_filename}{name of a PLINK \code{.map} file
Use \link{read_plink_map_file} to read a \code{PLINK} \code{.map} file.}
}
\value{
a [tibble]{tibble} with the following columns:
\itemize{
\item \code{CHR}: the chromosome code or contig name
\item \code{SNP}: Variant identifier
\item \code{position_cm}: Position in morgans or centimorgans.
This value is optional. Zeroes denote it is unused
\item \code{BP}: Base-pair coordinat
}
}
\description{
Read a PLINK \code{.map} file
}
\note{
The uppercase variable names are chosen to match those used by PLINK,
e.g. as can be observed from \link{assoc_qt}.

\code{position_cm} is still lowercase as I am unsure how PLINK calls this
variable. If you know, please let me know.
}
\author{
Richèl J.C. Bilderbeek
}
