% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_ped_file.R
\name{read_plink_ped_file}
\alias{read_plink_ped_file}
\title{Read a PLINK \code{.ped} file}
\usage{
read_plink_ped_file(ped_filename)
}
\arguments{
\item{ped_filename}{name of a \code{PLINK} \code{.ped} file.
Use \link{read_plink_ped_file} to read a \code{PLINK} \code{.ped} file.}
}
\value{
a \link[tibble]{tibble} with column names:
\itemize{
\item \code{family_id} The family ID (\code{FID})
\item \code{within_family_id} Within-family ID
(\code{IID}, cannot be zero)
\item \code{within_family_id_father} Within-family ID of father
(\code{0} if father isn't in dataset)
\item \code{within_family_id_mother} Within-family ID of mother
(\code{0} if mother isn't in dataset)
\item \code{sex_code} Sex code
(\code{1} = male, \code{2} = female, \code{0} = unknown)
\item \code{case_control_code} Case control code
(\code{1} = control, \code{2} = case,
\code{9}/\code{0}/non-numeric = missing data if case/control)
\item \code{snv_[x][y]} Nucleotide for the \code{x}th variant
for haplotype \code{y} (\code{y} is either \code{a} or \code{b})
in the \code{.map file} (\code{0} = no call)
}
}
\description{
Read a PLINK \code{.ped} file
}
\examples{
 if (is_plink_installed(create_plink_v1_7_options())) {
   read_plink_ped_file(
     get_plink_example_filename(
       example_filename = "test.ped",
       create_plink_v1_7_options()
     )
   )
 }
 if (is_plink_installed(create_plink_v1_9_options())) {
   read_plink_ped_file(
     get_plink_example_filename(
       example_filename = "toy.ped",
       create_plink_v1_9_options()
     )
   )
 }
}
\author{
Richèl J.C. Bilderbeek
}
