% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_phenotype_file.R
\name{read_plink_phenotype_file}
\alias{read_plink_phenotype_file}
\title{Read a PLINK raw phenotype file}
\usage{
read_plink_phenotype_file(phenotype_filename)
}
\arguments{
\item{phenotype_filename}{name of a raw phenotype file.
Use \link{read_plink_phenotype_file} to read a raw phenotype file.}
}
\value{
a tibble
}
\description{
Read a PLINK raw phenotype file
}
\examples{
read_plink_phenotype_file(
  phenotype_filename = get_plinkr_filename("pheno.raw")
)
}
