% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_phenotype_file.R
\name{read_plink_phenotype_file}
\alias{read_plink_phenotype_file}
\title{Read a PLINK raw phenotype file}
\usage{
read_plink_phenotype_file(phenotype_filename)
}
\arguments{
\item{phenotype_filename}{name of a raw phenotype file.
Use \link{read_plink_phenotype_file} to read a raw phenotype file.}
}
\value{
a \link[tibble]{tibble} with the following columns:
\itemize{
\item \code{family_id} The family ID (\code{FID})
\item \code{within_family_id} Within-family ID
(\code{IID}, cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}
}
\description{
Read a PLINK raw phenotype file
}
\examples{
read_plink_phenotype_file(
  phenotype_filename = get_plinkr_filename("pheno.raw")
)
}
\author{
Richèl J.C. Bilderbeek
}
