% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plink.R
\name{run_plink}
\alias{run_plink}
\title{Run PLINK from arguments}
\usage{
run_plink(
  args,
  plink_version = get_default_plink_version(),
  plink_folder = get_plink_folder(),
  add_noweb = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{args}{arguments for the PLINK executable}

\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the PLINK version.
Use \link{get_plink_versions} to get all the supported PLINK versions.}

\item{plink_folder}{folder where PLINK is installed}

\item{add_noweb}{add the \code{--noweb} option, which prevents
that PLINK will check if it is in its latest version.
By default, this value is set to true,
hence plinkr will skip this check.
Use \link{get_plink_version} to get the current PLINK version.}

\item{verbose}{set to TRUE for more output}
}
\value{
the text that PLINK returns
}
\description{
Run PLINK from arguments
}
\examples{
if (is_plink_installed()) {

  # Use the PLINK v1.9 example files
  ped_filename <- get_plink_example_filename("toy.ped")
  map_filename <- get_plink_example_filename("toy.map")

  # Do a case-control association
  run_plink(
    args = c(
     "--ped", ped_filename,
     "--map", map_filename
    )
  )
}
}
