% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_plink.R
\name{run_plink}
\alias{run_plink}
\title{Run PLINK from arguments}
\usage{
run_plink(args, plink_options = create_plink_options(), verbose = FALSE)
}
\arguments{
\item{args}{arguments for the PLINK executable}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{set to TRUE for more output}
}
\value{
the text that PLINK returns
}
\description{
Run PLINK from arguments
}
\examples{
if (is_plink_installed()) {

  # Use the PLINK v1.9 example files
  ped_filename <- get_plink_example_filename("toy.ped")
  map_filename <- get_plink_example_filename("toy.map")

  # Do a case-control association
  run_plink(
    args = c(
     "--ped", ped_filename,
     "--map", map_filename
    )
  )

  # Delete the created files
  file.remove("plink.bed")
  file.remove("plink.fam")
  file.remove("plink.log")
}
}
\author{
Richèl J.C. Bilderbeek
}
