% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_str_split.R
\name{safe_str_split}
\alias{safe_str_split}
\title{Call \link[stringr]{str_split} safely
stringi::stri_split_regex \strong{sometimes} gave an
'embedded nul in string' error.}
\usage{
safe_str_split(string)
}
\arguments{
\item{string}{one or more strings}
}
\description{
Call \link[stringr]{str_split} safely
stringi::stri_split_regex \strong{sometimes} gave an
'embedded nul in string' error.
}
