% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_map_table_to_file.R
\name{save_map_table_to_file}
\alias{save_map_table_to_file}
\title{Save a \code{.map} table to file}
\usage{
save_map_table_to_file(map_table, map_filename)
}
\arguments{
\item{map_table}{a table obtained by reading a PLINK \code{.map} file
Use \link{read_plink_map_file} to read a PLINK \code{.map} file.}

\item{map_filename}{name of a PLINK \code{.map} file
Use \link{read_plink_map_file} to read a PLINK \code{.map} file.}
}
\description{
Save a \code{.map} table to file
}
