% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ped_table_to_file.R
\name{save_ped_table_to_file}
\alias{save_ped_table_to_file}
\title{Save a \code{.ped} table to file}
\usage{
save_ped_table_to_file(ped_table, ped_filename)
}
\arguments{
\item{ped_table}{a table obtained by reading a PLINK \code{.ped} file
Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.}

\item{ped_filename}{name of a PLINK \code{.ped} file.
Use \link{read_plink_ped_file} to read a PLINK \code{.ped} file.}
}
\description{
Save a \code{.ped} table to file
}
