% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_phenotype_table_to_file.R
\name{save_phenotype_table_to_file}
\alias{save_phenotype_table_to_file}
\title{Save a \code{.phenotype} table to file}
\usage{
save_phenotype_table_to_file(phenotype_table, phenotype_filename)
}
\arguments{
\item{phenotype_table}{a table of phenotypes.

A \code{phenotype_table} is a \link[tibble]{tibble} with these columns:
\itemize{
\item \code{FID} The family ID
\item \code{IID} Within-family ID
(cannot be zero)
\item One or more columns of phenotype values,
columns can have any name
}

Thes names match the PLINK column
names (\url{https://www.cog-genomics.org/plink/1.9/input#pheno}).
Use \link{read_plink_phenotype_file} to read a phenotype file.
Use \link{check_phenotype_table} to test if a phenotype table is valid.}

\item{phenotype_filename}{name of a raw phenotype file.
Use \link{read_plink_phenotype_file} to read a raw phenotype file.}
}
\description{
Save a \code{.phenotype} table to file
}
\author{
Richèl J.C. Bilderbeek
}
