% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_phenotype_table_to_file.R
\name{save_phenotype_table_to_file}
\alias{save_phenotype_table_to_file}
\title{Save a \code{.phenotype} table to file}
\usage{
save_phenotype_table_to_file(phenotype_table, phenotype_filename)
}
\arguments{
\item{phenotype_table}{a table of phenotypes.
Use \link{read_plink_phenotype_file} to read a phenotype file.}

\item{phenotype_filename}{name of a raw phenotype file.
Use \link{read_plink_phenotype_file} to read a raw phenotype file.}
}
\description{
Save a \code{.phenotype} table to file
}
