% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plink_text_data.R
\name{save_plink_text_data}
\alias{save_plink_text_data}
\title{Save \code{PLINK} text data to files.}
\usage{
save_plink_text_data(
  plink_text_data,
  base_input_filename = get_plinkr_tempfilename(),
  verbose = FALSE
)
}
\arguments{
\item{plink_text_data}{the genetic and phenotypic data for \code{PLINK} to work
on, in \code{PLINK} text format, as created by \link{create_plink_text_data}.}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{PLINK}/\code{PLINK2}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
the full paths to the \code{.bed}, \code{.bim} and \code{.fam} files created
}
\description{
Save \code{PLINK} text data to files.
}
\author{
Richèl J.C. Bilderbeek
}
