% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_snps.R
\name{select_snps}
\alias{select_snps}
\title{Select SNPs}
\usage{
select_snps(data, snps)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
}}
}
\description{
Select SNPs
}
\author{
Richèl J.C. Bilderbeek
}
