% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_snps.R
\name{select_snps}
\alias{select_snps}
\title{Select SNPs}
\usage{
select_snps(
  data,
  snp_selector,
  base_output_filename = plinkr::get_plinkr_tempfilename(),
  plink_options = plinkr::create_plink_options(),
  verbose = FALSE
)
}
\arguments{
\item{data}{the data source, which can be:
\itemize{
\item PLINK text data, as created by \link{create_plink_text_data}.
\item PLINK binary data, as created by \link{create_plink_bin_data}.
\item PLINK2 binary data, as created by \link{create_plink2_bin_data}.
\item PLINK text filenames, as created by \link{create_plink_text_filenames}.
\item PLINK binary filenames, as created by \link{create_plink_bin_filenames}.
\item PLINK2 binary filenames, as created by \link{create_plink2_bin_filenames}.
}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{PLINK}/\code{PLINK2}}

\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\description{
Select SNPs
}
\examples{
# Use PLINK files
plink_bin_filenames <- create_plink_bin_filenames(
  bed_filename = get_plinkr_filename("select_snps.bed"),
  bim_filename = get_plinkr_filename("select_snps.bim"),
  fam_filename = get_plinkr_filename("select_snps.fam")
)
# Selects 3 SNPs, with 'snp_5' in the middle
snp_window_selector <- create_snp_window_selector(
  snp = "snp_5",
  window_kb = 0.003
)
plink_bin_data <- select_snps(
  data = plink_bin_filenames,
  snp_selector = snp_window_selector
)
}
\author{
Richèl J.C. Bilderbeek
}
