% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_qt.R
\name{simulate_qt}
\alias{simulate_qt}
\title{Simulate a quantitative trait}
\usage{
simulate_qt(
  simulate_qt_params,
  n_individuals,
  temp_sim_filename,
  plink_version = get_default_plink_version(),
  add_noweb = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{simulate_qt_params}{the parameters for a quantitative
traits simulation, as can be created by
\link{create_simulate_qt_params}}

\item{n_individuals}{the number of individuals}

\item{temp_sim_filename}{temporary file to store simulation
parameters, which is a PLINK \code{.sim} file}

\item{plink_version}{version of PLINK, e.g. \code{"1.7"}
Use \link{get_plink_version} to get the PLINK version.
Use \link{get_plink_versions} to get all the supported PLINK versions.}

\item{add_noweb}{add the \code{--noweb} option, which prevents
that PLINK will check if it is in its latest version.
By default, this value is set to true,
hence plinkr will skip this check.
Use \link{get_plink_version} to get the current PLINK version.}

\item{verbose}{set to TRUE for more output}
}
\description{
Simulate a quantitative trait
}
\note{
This function is named after the \code{--simulate-qt} PLINK flag.
}
