% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_plink.R
\name{uninstall_plink}
\alias{uninstall_plink}
\title{Uninstall PLINK}
\usage{
uninstall_plink(plink_options = create_plink_options())
}
\arguments{
\item{plink_options}{options to run PLINK,
as created by \link{create_plink_options}}
}
\value{
Nothing
}
\description{
Uninstall PLINK
}
\author{
Richèl J.C. Bilderbeek
}
