% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_plinks.R
\name{uninstall_plinks}
\alias{uninstall_plinks}
\title{Uninstall multiple versions of PLINK}
\usage{
uninstall_plinks(
  plink_versions = get_plink_versions(),
  plink_folder = get_plink_folder()
)
}
\arguments{
\item{plink_versions}{one or more versions of PLINK,
e.g. as can be obtained using \link{get_plink_versions}}

\item{plink_folder}{folder where PLINK is installed}
}
\description{
Uninstall multiple versions of PLINK
}
