% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_plinks.R
\name{uninstall_plinks}
\alias{uninstall_plinks}
\title{Uninstall multiple versions of PLINK}
\usage{
uninstall_plinks(plink_optionses = create_plink_optionses())
}
\arguments{
\item{plink_optionses}{a list of one or more options to run PLINK,
as created by \link{create_plink_optionses}.
The reduplicated plural was used to express this is a list
of \code{plink_options}, instead of one set of \code{plink_options}}
}
\description{
Uninstall multiple versions of PLINK.
This deletes only the PLINK folders that are expected to be
present.
Use \link{clear_plink_folder} to ruthlessly delete all files in the PLINK
folder.
}
\author{
Richèl J.C. Bilderbeek
}
